### 2023-07-10 #5646 @editorialbot
[PRE REVIEW]: matscipy: materials science at the atomic scale with Python
**Submitting author:** @pastewka (<a href="http://orcid.org/0000-0001-8351-7336">Lars Pastewka</a>)
**Repository:** https://github.com/libAtoms/matscipy
**Branch with paper.md** (empty if default branch):
**Version:** 0.8.0
