### 2023-07-10 #5646 @Kevin-Mattheus-Moerman
@pastewka thanks for this submission. I am the AEiC for this track and here to help process initial steps. I will be looking for a handling editor but for the moment, can you start on the points below:

- [ ] The paper compilation error says `failure: Author (Jacek Golebiowski) is missing affiliation`, can you try to address this? You can call `@editorialbot generate pdf` to try to compile it again.
- [ ] Can you check the above potentially missing/invalid DOIs :point_up:, you can call `@editorialbot check references` to check the DOIs again. Also make sure your bib file does not contain any entries you are not citing in the paper (I have not checked this but just asking you to make sure given that the list is long).
