### 2023-07-10 f903e67
DOC: Fixed DOIs and Jacek's affiliation
