+++ paper/paper.bib
@@ -242,8 +242,7 @@ abstract = {Calculations of dislocation-defect interactions are essential to mod
 	author = {Tersoff, J},
 	year = {1986},
 	pages = {632--635},
-	file = {Tersoff 1986 - New Empirical Model for the Structural Properties of Silicon:C\:\\Users\\Lars Pastewka\\Zotero\\storage\\QCIUZDUB\\Tersoff 1986 - New Empirical Model for the Structural Properties of Silicon.pdf:application/pdf},
-	optdoi = {10.1103/PhysRevLett.56.632},
+	doi = {10.1103/PhysRevLett.56.632}
 }
 
 @article{Tersoff1989,
@@ -373,7 +372,8 @@ abstract = {Calculations of dislocation-defect interactions are essential to mod
   title = {Analysis and simulation of semiconductor devices},
   author = {Selberherr, Siegfried},
   year = {1984},
-  publisher = {Springer Science \& Business Media}
+  publisher = {Springer Science \& Business Media},
+  doi = {10.1007/978-3-7091-8752-4}
 }
 
 @article{Seidl2021,
@@ -455,7 +455,7 @@ abstract = {Calculations of dislocation-defect interactions are essential to mod
   volume    =  124,
   pages     = "105501",
   year      =  2020,
-  doi       = "10.1103/PhysRevLett.124.10550"
+  doi       = "10.1103/PhysRevLett.124.105501"
 }
 
 @ARTICLE{Reichenbach2020,
