+++ paper/paper.bib
@@ -548,11 +548,12 @@ abstract = {Calculations of dislocation-defect interactions are essential to mod
                    results, we outline the effect of multiaxial stress states
                    as opposed to volumetric deformation on the limits of
                    stability of their crystalline lattices.",
-  month         =  feb,
   year          =  2023,
-  archivePrefix = "arXiv",
-  primaryClass  = "cond-mat.mtrl-sci",
-  eprint        = "2302.08754"
+  journal       = "Phys. Rev. Mater.",
+  volume        = 7,
+  number        = 7,
+  pages         = "073603",
+  doi           = "10.1103/PhysRevMaterials.7.073603"
 }
 
 @ARTICLE{Miller2009,
