+++ paper/paper.md
@@ -146,7 +146,7 @@ Besides generating and analysing atomic-scale configurations, `matscipy` impleme
 - **Interatomic potentials.** The module `matscipy.calculators` has implementations of classical pair-potentials, Coulomb interactions, the embedded-atom method (EAM) [@Daw1984] and other many-body potentials [e.g. @StillingerWeber1985;@Tersoff1989].
 
 - **Second-order derivatives.** The thermodynamic and elastic properties of solid materials are closely connected to the Hessian of the overall system, which contains the second derivatives of the total energy with respect to position and macroscopic strains. `matscipy` implements analytic second-order potential derivatives for pair-potentials [@Lennard1931], EAM potentials [@Daw1984], bond-order potentials [@Kumagai2007;@Tersoff1989;@Brenner1990], cluster potentials [@StillingerWeber1985] and electrostatic interaction [@BKS1990].
-This is achieved through a generic mathematical formulation of the manybody total energy [@Griesser2023b] in `matscipy.calculators.manybody`.
+This is achieved through a generic mathematical formulation of the manybody total energy [@Muser2023;@Griesser2023crystal] in `matscipy.calculators.manybody`.
 The module `matscipy.numerical` additionally provides routines for the numerical (finite-differences) evaluation of these properties. These analytic second-order derivatives allow a fast and accurate computation of the aforementioned properties in crystals, polymers and amorphous solids, even for unstable configurations where numerical methods are not applicable.
 
 - **Quantum mechanics/molecular mechanics.** The module `matscipy.calculators.mcfm` implements a generalised force-mixing potential [@Bernstein2009] with support for multiple concurrent QM clusters, named MultiClusterForceMixing (MCFM). It has been applied to model failure of graphene-nanotube composites [@Golebiowski2018;@Golebiowski2020].
