### 2023-07-10 #5646 @pastewka
@Kevin-Mattheus-Moerman I resolved the issues and the proof looks good. Note that the DOI listed as invalid is for a paper that was literally posted online an hour ago. It should because valid over the next couple of days.
