### 2023-07-10 fd429c7
DOC: Fixed middle authors to alphabetical
