### 2023-07-11 278705a
DOC: minor changes to plasticity docs
