### 2023-07-13 e986529
WIP: Seperated F and E evaluation into two seperate functions
Also added functionality to supply E functions with multiple coordinate
systems
