+++ matscipy/cauchy_born.py
@@ -323,34 +323,33 @@ class CubicCauchyBorn:
         except FileNotFoundError:
             print('No files found containing model parameters - make sure to call save_taylor() to save the model.')
 
-    def F_3D_from_atoms(
+    def tensor_field_3D_from_atoms(
             self,
             atoms,
-            F_func,
+            func,
             coordinates='cart3D',
             cell=None,
             *args,
             **kwargs):
-        """Generate the 3D deformation gradient tensor field (F) for an atoms
-        object in a given coordinate system.
+        """Generate a 3D tensor field (F or E) for an atoms
+        object in a given coordinate system, using a provided
+        function
 
         Parameters
         ----------
         atoms : ASE atoms object
             Atoms object where atoms sit at the coordinates needed
-            to generate the deformation gradient tensor field
-            from F_func.
-        F_func : Function
-            Function to calculate the deformation gradient tensor field
+            to generate the tensor field from func.
+        func : Function
+            Function to calculate the tensor field
             from a set of coordinates. The system of coordinates that
             the function accepts is given by 'coordinates' and the
             function also accepts anything extra specified in *args
-            and **kwargs. Function must return the deformation
-            gradient field for the atoms in form of a numpy array
-            shape [natoms,ndims,ndims], where F is specified in
-            cartesian coordinates.
+            and **kwargs. Function must return the field for the atoms 
+            in form of a numpy array shape [natoms,ndims,ndims], 
+            where output is specified in cartesian coordinates.
         coordinates: string
-            The coordinates system that F_func accepts. Must
+            The coordinates system that func accepts. Must
             be 'cylind2D', 'cylind3D', 'spherical', 'cart2D', 'cart3D'
         cell : array_like, optional
             An optional argument allowing the user to specify
@@ -358,8 +357,8 @@ class CubicCauchyBorn:
 
         Returns
         -------
-        F3D : array_like
-            The 3D deformation gradient tensor field for all
+        field3D : array_like
+            The 3D tensor field for all
             the atoms in the system, expressed in the lab frame.
         """
         natoms = len(atoms)
@@ -373,11 +372,11 @@ class CubicCauchyBorn:
             cx, cy = sx / 2, sy / 2
             r = np.sqrt((x - cx)**2 + (y - cy)**2)
             theta = np.arctan2((y - cy), (x - cx))
-            F_2D = F_func(r, theta, *args, **kwargs)
+            TF_2D = func(r, theta, *args, **kwargs)
             # pad F into a 3x3 matrix
-            F_3D = (np.zeros([natoms, 3, 3]))
-            F_3D[:, 0:2, 0:2] = F_2D[:, :, :]
-            F_3D[:, 2, 2] = 1
+            TF_3D = (np.zeros([natoms, 3, 3]))
+            TF_3D[:, 0:2, 0:2] = TF_2D[:, :, :]
+            TF_3D[:, 2, 2] = 1
 
         elif coordinates == 'cylind3D':
             x, y, z = atoms.positions[:,
@@ -385,29 +384,29 @@ class CubicCauchyBorn:
             cx, cy, cz = sx / 2, sy / 2, sz / 2
             r = np.sqrt((x - cx)**2 + (y - cy)**2)
             theta = np.arctan2((y - cy), (x - cx))
-            F_3D = F_func(r, theta, z, *args, **kwargs)
+            TF_3D = func(r, theta, z, *args, **kwargs)
 
         elif coordinates == 'spherical':
             raise NotImplementedError
 
         elif coordinates == 'cart2D':
             x, y = atoms.positions[:, 0], atoms.positions[:, 1]
[…]
