### 2023-07-13 e7c57f6
WIP: removed superfluous function
