+++ matscipy/fracture_mechanics/crack.py
@@ -570,7 +570,7 @@ class CubicCrystalCrack:
 
     def __init__(self, crack_surface, crack_front, C11=None, C12=None,
                  C44=None, stress_state=PLANE_STRAIN, C=None, Crot=None,
-                 cb=None):
+                 cauchy_born=None):
         """
         Initialize a crack in a cubic crystal with elastic constants C11, C12
         and C44 (or optionally a full 6x6 elastic constant matrix C).
@@ -620,7 +620,7 @@ class CubicCrystalCrack:
         self.crack_surface = crack_surface
         self.crack_front = crack_front
         self.RotationMatrix = A
-        self.cb = cb
+        self.cauchy_born = cauchy_born
 
 
     def k1g(self, surface_energy):
@@ -1063,97 +1063,6 @@ class SinclairCrack:
         return u
 
 
-    #-----------------delete this part from main branch-------------------
-
-    def relax_crack_surface(self):
-        """Relaxes the free surface of the crack in regions II-IV. This is crucial
-        as the unrelaxed surface leads to huge f_{alpha} components which leads to a 
-        strong finite size effect. """
-        self.atoms.set_pbc([False, False, True])
-        self.atoms.calc = self.calc
-
-        self.atoms.info['k'] = self.k
-        self.atoms.info['alpha'] = self.alpha
-        
-        # x = x_cryst + K * u_cle + u
-        self.atoms.positions[:, :] = self.cryst.positions
-        #self.u_cle()[:,1].tofile('uCLEsolution.csv',sep=',')
-        u_cle_solution = self.k * self.u_cle()
-        self.atoms.positions[:, :] += u_cle_solution
-        self.atoms.positions[self.regionI, :] += self.u
-
-        # add vacuum
-        self.atoms.cell = self.cryst.cell
-        self.atoms.cell[0, 0] += self.vacuum
-        self.atoms.cell[1, 1] += self.vacuum
-
-        #clone atoms for relaxing
-        atoms = self.atoms.copy()
-        i = neighbour_list('i', atoms, 12.0) #large cutoff to get the first 2 layers
-        coord = np.bincount(i)
-        np.savetxt('coords.txt',coord)
-        #build a mask for relaxing based on the following criteria:
-        # - the coordination of the atoms must be lower than 0.8*the max coordination
-        # - the atoms x coordinate should lie between rI - cutoff and 
-        #  rI + 2*cutoff (if non extended) and rIII + cutoff/2 if extended.
-        # - the atoms y coordinate should lie between the positive and negative
-        # maximum y Ucle displacement + cutoff
-        coordination_criteria = (coord<(0.8*np.max(coord)))
-        #get the coords of the centre of the cell.
-        mid_cell_x = self.cryst.cell.diagonal()[0] / 2.0
-        mid_cell_y = self.cryst.cell.diagonal()[1] / 2.0
-        if self.extended_far_field:
-            x_criteria = ((atoms.positions[:,0]-mid_cell_x)<-((self.rI)-(2*self.cutoff)))&\
-                (((atoms.positions[:,0]-mid_cell_x)>-((self.rIII)+(self.cutoff/2))))
-        else:
-            x_criteria = ((atoms.positions[:,0]-mid_cell_x)<-((self.rI)-(2*self.cutoff)))&\
-                (((atoms.positions[:,0]-mid_cell_x)>-((self.rI)+((2*self.cutoff)))))
-        y_criteria = ((atoms.positions[:,1]-mid_cell_y)<(np.max(u_cle_solution[:,1])+(2*self.cutoff)))&\
[…]
