### 2023-07-13 #5646 @pastewka
@diehlpk We have the following suggestions for reviewers:

Pierre Hirel (Lille, maintains atoms package)
Noam Bernstein (Naval Research Lab)
Lucas Hale (NIST, maintains atomman package)
Markus Stricker (ICAMS Bochum)
Gabor Csanyi (Cambridge)

Let me know if you need contact info.
