+++ tests/test_cauchy_born_corrector.py
@@ -1,0 +1,186 @@
+import unittest
+import numpy as np
+from matscipy.cauchy_born import CubicCauchyBorn
+from ase.build import bulk
+from ase.lattice.cubic import Diamond
+from ase.optimize.precon import PreconLBFGS
+from ase.constraints import ExpCellFilter
+from matscipy.cauchy_born import CubicCauchyBorn
+from matscipy.calculators.manybody.explicit_forms.stillinger_weber import StillingerWeber, Holland_Marder_PRL_80_746_Si
+from matscipy.calculators.manybody import Manybody
+from scipy.linalg import sqrtm
+import ase.io
+import matscipytest
+
+
+class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
+    """
+    Tests of Cauchy-Born shift prediction in the case of multilattices.
+
+    We test with a cubic Silicon diamond multi-lattice
+    """
+    def setUp(self):
+        el = 'Si' # chemical element
+        si = Diamond(el, latticeconstant=5.43, size=[1, 1, 1], 
+                       directions=[[1,0,0],[0,1,0],[0,0,1]])
+        calc = Manybody(**StillingerWeber(Holland_Marder_PRL_80_746_Si))
+        si.calc = calc
+        ecf = ExpCellFilter(si)
+        opt = PreconLBFGS(ecf)
+        opt.run(fmax=1e-6, smax=1e-6)
+        #print(si.get_positions())
+        a0 = si.cell[0, 0]
+        #print(a0)
+        self.cb = CubicCauchyBorn(el,a0,calc,lattice=Diamond)
+        self.unit_cell = si
+
+
+    def test_set_sublattice(self):
+        A = np.eye(3)
+        self.cb.set_sublattices(self.unit_cell,A)
+        assert np.array_equal(self.cb.lattice1mask,np.array([True,False,True,False,True,False,True,False],dtype=bool))
+        assert np.array_equal(self.cb.lattice2mask,np.array([False,True,False,True,False,True,False,True],dtype=bool))
+
+    def test_fit_taylor_model(self):
+        self.cb.fit_taylor()
+        #print(self.cb.grad_f)
+        #print(self.cb.hess_f)
+        grad_f = np.array([0.00000000e+00,0.00000000e+00,0.00000000e+00,-1.09995082e+00,-2.22044605e-12,-3.33066907e-12])
+        hess_f = np.array([[0.00000000e+00,0.00000000e+00,0.00000000e+00,2.03958794e-01,0.00000000e+00,1.11022302e-08],
+                        [0.00000000e+00,0.00000000e+00,0.00000000e+00,9.12311482e-01,-5.55111512e-09,0.00000000e+00],
+                        [0.00000000e+00,0.00000000e+00,0.00000000e+00,9.12311465e-01,0.00000000e+00,0.00000000e+00],
+                        [2.03958794e-01,9.12311493e-01,9.12311465e-01,0.00000000e+00,0.00000000e+00,0.00000000e+00],
+                        [0.00000000e+00,-5.55111512e-09,0.00000000e+00,0.00000000e+00,0.00000000e+00,0.00000000e+00],
+                        [1.11022302e-08,0.00000000e+00,0.00000000e+00,0.00000000e+00,0.00000000e+00,0.00000000e+00]])
+        assert np.allclose(self.cb.grad_f,grad_f,atol=1e-8)
+        assert np.allclose(self.cb.hess_f,hess_f,atol=1e-8)
+
+    def E_cart3D(self,x,y,z,eps=None):
+        E = np.zeros([len(x),3,3])
[…]
