### 2023-07-17 9539bd3
DOC: Made Cauchy-Born test PEP8 compliant
