+++ tests/test_cauchy_born_corrector.py
@@ -90,7 +90,9 @@ class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
         E[:, 1, 2], E[:, 2, 1] = eps[3], eps[3]
         E[:, 0, 2], E[:, 2, 0] = eps[4], eps[4]
         E[:, 0, 1], E[:, 1, 0] = eps[5], eps[5]
-        U = sqrtm((2*E[i, :, :])+np.eye(3))
+        U = np.zeros_like(E)
+        for i in range(np.shape(E)[0]):
+            U[i,:,:] = sqrtm((2*E[i, :, :])+np.eye(3))
         return U  # with no rigid rotation, U is F
 
     def F_cylind3D(self, r, theta, z, eps=None):
@@ -101,7 +103,9 @@ class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
         E[:, 1, 2], E[:, 2, 1] = eps[3], eps[3]
         E[:, 0, 2], E[:, 2, 0] = eps[4], eps[4]
         E[:, 0, 1], E[:, 1, 0] = eps[5], eps[5]
-        U = sqrtm((2*E[i, :, :])+np.eye(3))
+        U = np.zeros_like(E)
+        for i in range(np.shape(E)[0]):
+            U[i,:,:] = sqrtm((2*E[i, :, :])+np.eye(3))
         return U  # with no rigid rotation, U is F
 
     def model_prediction(self, dirs, eps, method, E_func=None, F_func=None, coordinates='cart3D', atol=1.5e-5, returnvals=False):
@@ -136,7 +140,7 @@ class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
         if returnvals:
             return atoms_copy, shifts, np.linalg.norm(shift_diff_predicted-shift_diff_true), A
 
-    def test_taylor_model(self):
+    def test_taylor_model_E(self):
         self.cb.fit_taylor()
         eps_vals = [np.array([0, 0, 0, 0.0001, 0, 0]),
                     np.array([0.0001, 0, 0.0001, 0.0001, 0.0001, 0]),
@@ -154,10 +158,28 @@ class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
             self.model_prediction(
                 dir_vals[i], eps_vals[i], method='taylor', E_func=funcs[i], coordinates=coords[i], atol=1e-7)
 
+    def test_taylor_model_F(self):
+        self.cb.fit_taylor()
+        eps_vals = [np.array([0, 0, 0, 0.0001, 0, 0]),
+                    np.array([0.0001, 0, 0.0001, 0.0001, 0.0001, 0]),
+                    np.array([0, 0, 0, 0.0001, 0, 0]),
+                    np.array([0, 0, 0, 0.0001, 0, 0])]
+
+        dir_vals = [[[1, 0, 0], [0, 1, 0], [0, 0, 1]],
+                    [[1, 1, 0], [-1, 1, 0], np.cross([1, 1, 0], [-1, 1, 0])],
+                    [[1, 1, 1], [-2, 1, 1], np.cross([1, 1, 1], [-2, 1, 1])],
+                    [[1, 1, 0], [-1, 1, 0], np.cross([1, 1, 0], [-1, 1, 0])]]
+
+        funcs = [self.F_cart3D, self.F_cart3D, self.F_cart3D, self.F_cylind3D]
+        coords = ['cart3D', 'cart3D', 'cart3D', 'cylind3D']
+        for i in range(len(funcs)):
+            self.model_prediction(
+                dir_vals[i], eps_vals[i], method='taylor', F_func=funcs[i], coordinates=coords[i], atol=1e-7)
+    
     def test_fit_regression_model(self):
[…]
