### 2023-07-17 31410e1
DOC: Made cauchy-born test PEP8 compliant
