+++ matscipy/surface_reconstruction.py
@@ -7,8 +7,7 @@ from ase.constraints import FixAtoms
 
 from matscipy.cauchy_born import CubicCauchyBorn
 from ase.optimize.precon import PreconLBFGS
-import ase.io
-
+import ase
 
 class SurfaceReconstruction:
     """Object for mapping and applying a surface reconstruction in simple and multi-lattices.
@@ -19,7 +18,7 @@ class SurfaceReconstruction:
     can be mapped to the internal crack surfaces of an atoms object.
     """
 
-    def __init__(self, el, a0, calc, directions, surf_dir, lattice='diamond'):
+    def __init__(self, el, a0, calc, directions, surf_dir, lattice='diamond',multilattice=False):
         """Parameters
         ----------
         el : string
@@ -34,8 +33,8 @@ class SurfaceReconstruction:
         surf_dir : int
             Direction which is normal to the free surface. Can be 0, 1 or 2 for
             lab x, lab y or lab z (as defined by directions).
-        lattice : string
-            Crystal lattice type.
+        lattice : ASE ase.lattice.cubic function
+            Callable function which can be used to generate the crystal.
         """
         # directions is a list of 3 vectors specifiying the surface of interest
         # in the orientation specified by the problem. surf_dir says which axis is free surface
@@ -45,7 +44,7 @@ class SurfaceReconstruction:
         self.calc = calc
         self.directions = directions
         self.surf_dir = surf_dir
-
+        self.inter_planar_dist = a0/np.linalg.norm(directions[surf_dir])
         # build the rotation matrix from the directions
         A = np.zeros([3, 3])
         for i, direction in enumerate(directions):
@@ -53,12 +52,13 @@ class SurfaceReconstruction:
             A[:, i] = direction / np.linalg.norm(direction)
         self.A = A  # rotation matrix
         self.eval_cb = None
-        if lattice == 'diamond':
-            self.lattice = Diamond
+        self.lattice = lattice
+        self.cb = None
+        if multilattice:
             self.cb = CubicCauchyBorn(
-                self.el, self.a0, self.calc, lattice=Diamond)
+                self.el, self.a0, self.calc, lattice=self.lattice)
 
-    def map_surface(self, fmax=0.0001, layers=6):
+    def map_surface(self, fmax=0.0001, layers=6, cutoff=10, shift=0):
         """Map the relaxation of the crystal surface a certain number of layers deep.
         Parameters
         ----------
@@ -66,9 +66,40 @@ class SurfaceReconstruction:
             Force tolerance for relaxation
         layers : int
             Number of layers deep to map the free surface
+        shift : float
+            Amount to shift the bulk structure before the surface is mapped
+            such that the top layer of surface is physically meaningful. One
+            use case of this would be to adjust the top layer of atoms being mapped
+            in diamond such that it matches the top layer of atoms seen on the cleavage
+            plane.
         """
-        # set the number of atomic layers to model
-        height = 30
+        #build a single cell to determine how many atomic surface
+        #layers there are in a single unit cell made by ASE
+        single_cell = self.lattice(
+            directions=self.directions,
+            size=[1,1,1],
+            symbol=self.el,
+            latticeconstant=self.a0,
+            pbc=(1, 1, 1))
+        #ase.io.write(f'{self.directions[self.surf_dir]}_unitcell.xyz',single_cell)
+
+        if self.cb is not None:
+            # for a multilattice, get number of layers by just looking
+            # at one of the sublattices
+            self.cb.set_sublattices(single_cell, self.A)
[…]
