+++ matscipy/surface_reconstruction.py
@@ -9,6 +9,7 @@ from matscipy.cauchy_born import CubicCauchyBorn
 from ase.optimize.precon import PreconLBFGS
 import ase
 
+
 class SurfaceReconstruction:
     """Object for mapping and applying a surface reconstruction in simple and multi-lattices.
 
@@ -18,7 +19,7 @@ class SurfaceReconstruction:
     can be mapped to the internal crack surfaces of an atoms object.
     """
 
-    def __init__(self, el, a0, calc, directions, surf_dir, lattice='diamond',multilattice=False):
+    def __init__(self, el, a0, calc, directions, surf_dir, lattice='diamond', multilattice=False):
         """Parameters
         ----------
         el : string
@@ -73,33 +74,35 @@ class SurfaceReconstruction:
             in diamond such that it matches the top layer of atoms seen on the cleavage
             plane.
         """
-        #build a single cell to determine how many atomic surface
-        #layers there are in a single unit cell made by ASE
+        # build a single cell to determine how many atomic surface
+        # layers there are in a single unit cell made by ASE
         single_cell = self.lattice(
             directions=self.directions,
-            size=[1,1,1],
+            size=[1, 1, 1],
             symbol=self.el,
             latticeconstant=self.a0,
             pbc=(1, 1, 1))
-        #ase.io.write(f'{self.directions[self.surf_dir]}_unitcell.xyz',single_cell)
+        # ase.io.write(f'{self.directions[self.surf_dir]}_unitcell.xyz',single_cell)
 
         if self.cb is not None:
             # for a multilattice, get number of layers by just looking
             # at one of the sublattices
             self.cb.set_sublattices(single_cell, self.A)
-            n_layer_per_cell = len(np.unique(single_cell.get_positions()[self.cb.lattice1mask][:,self.surf_dir].round(decimals=6)))
+            n_layer_per_cell = len(np.unique(single_cell.get_positions()[
+                                   self.cb.lattice1mask][:, self.surf_dir].round(decimals=6)))
         else:
-            #otherwise, just look at all the layers directly
-            n_layer_per_cell = len(np.unique(single_cell.get_positions()[:,self.surf_dir].round(decimals=6)))
+            # otherwise, just look at all the layers directly
+            n_layer_per_cell = len(np.unique(single_cell.get_positions()[
+                                   :, self.surf_dir].round(decimals=6)))
 
+        # print('nlayer per cell', n_layer_per_cell)
+        # ase.io.write('single_surface_cell.xyz',single_cell)
 
-        #print('nlayer per cell', n_layer_per_cell)
-        #ase.io.write('single_surface_cell.xyz',single_cell)
- 
         # set the number of atomic unit cells to map
         # as twice the number of layers to map + a number of layers
         # equal to the potential cutoff for fixing
-        height = 2*int(layers/n_layer_per_cell) + int(np.round(cutoff/self.inter_planar_dist))
+        height = 2*int(layers/n_layer_per_cell) + \
+            int(np.round(cutoff/self.inter_planar_dist))
         size = [1, 1, 1]
         size[self.surf_dir] *= height
         self.layers = layers
@@ -112,9 +115,9 @@ class SurfaceReconstruction:
             symbol=self.el,
             latticeconstant=self.a0,
             pbc=(1, 1, 1))
-        
-        #shift the cell slightly and wrap to stop dodgy surfaces when vacuum added
-        #as well as to add on any user-defined shift
+
+        # shift the cell slightly and wrap to stop dodgy surfaces when vacuum added
+        # as well as to add on any user-defined shift
         shift_vec = np.zeros([3])
         shift_vec[self.surf_dir] += 0.01 + shift
         bulk.positions += shift_vec
[…]
