### 2023-07-18 c449148
TST: Added tests for surface reconstruction
