+++ tests/test_surface_reconstruction.py
@@ -1,0 +1,168 @@
+import unittest
+import numpy as np
+from matscipy.cauchy_born import CubicCauchyBorn
+from ase.build import bulk
+from ase.lattice.cubic import Diamond, FaceCenteredCubic
+from ase.optimize.precon import PreconLBFGS
+from ase.constraints import ExpCellFilter
+from matscipy.cauchy_born import CubicCauchyBorn
+from matscipy.calculators.manybody.explicit_forms.stillinger_weber import StillingerWeber, Holland_Marder_PRL_80_746_Si
+from matscipy.calculators.manybody.explicit_forms.tersoff_brenner import TersoffBrenner, Erhart_PRB_71_035211_Si
+from matscipy.calculators.manybody import Manybody
+from matscipy.surface_reconstruction import SurfaceReconstruction
+import ase
+from matscipy.calculators.eam import EAM
+import matscipytest
+
+
+class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
+    """
+    Tests of Cauchy-Born shift prediction in the case of multilattices.
+
+    Test uses EAM for gold (Au-Grochola-JCP05.eam.alloy)
+    and 
+    """
+
+    def setupmultilattice(self):
+        el = 'Si'  # chemical element
+        unitcell = Diamond(el, latticeconstant=5.43, size=[1, 1, 1],
+                        directions=[[1, 0, 0], [0, 1, 0], [0, 0, 1]])
+        #model = "c_v3.yace"
+        #d2_table = "d2.table"
+        #cmds = ['pair_style hybrid/overlay pace table linear 10000',
+        #'pair_coeff * * pace c_v3.yace C', 'pair_coeff * * table d2.table D2 9.0']
+        #calc = LAMMPSlib(amendments=["mass 1 12.101"],lmpcmds = cmds,log_file='test.log',keep_alive=True)
+        calc = Manybody(**TersoffBrenner(Erhart_PRB_71_035211_Si))
+        unitcell.calc = calc
+        ecf = ExpCellFilter(unitcell)
+        opt = PreconLBFGS(ecf)
+        opt.run(fmax=0.0001, smax=0.0001)
+        a0 = unitcell.cell[0, 0]
+        self.cb = CubicCauchyBorn(el, a0, calc, lattice=Diamond)
+        self.el = el
+        self.a0 = a0
+        self.calc = calc
+
+    def test_surface_mapping_multilattice(self):
+        self.setupmultilattice()
+        dir_vals = [[[1, 1, 0], [-1, 1, 0],[0,0,1]],
+                    [[1, 1, 1], [-2, 1, 1],[0,-1,1]]]
+        surf_dirs = [0,1,2]
+        for dirs in dir_vals:
+            for surf_dir in surf_dirs:
+                #print(f'testing relaxation of {dirs[surf_dir]} surface')
+                sr = SurfaceReconstruction(self.el,self.a0,self.calc,dirs,surf_dir,lattice=Diamond,multilattice=True)
+                sr.map_surface(fmax=0.0005,shift=0)
+                self.check_surface_mapping_multilattice(sr,dirs,Diamond,surf_dir)
+    
+    def check_surface_mapping_multilattice(self,sr,directions,lattice,surf_dir,shift=0):
+        cb = CubicCauchyBorn(self.el,self.a0,self.calc,lattice=lattice)
+
+        size = [1, 1, 1]
+        size[surf_dir] *= 20
+
+        shift_vec = np.zeros([3])
+        shift_vec[surf_dir] += 0.01 + shift
+
+        bulk = lattice(directions=directions, size=size, symbol=self.el, latticeconstant=self.a0, pbc=(1,1,1))
+        calc = self.calc
+        bulk.positions += 0.01+shift
+        bulk.wrap()
+
[…]
