### 2023-07-18 b9ae1a2
DOC: Made surface reconstruction tests PEP8 compliant
