+++ tests/test_surface_reconstruction.py
@@ -26,12 +26,12 @@ class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
     def setupmultilattice(self):
         el = 'Si'  # chemical element
         unitcell = Diamond(el, latticeconstant=5.43, size=[1, 1, 1],
-                        directions=[[1, 0, 0], [0, 1, 0], [0, 0, 1]])
-        #model = "c_v3.yace"
-        #d2_table = "d2.table"
-        #cmds = ['pair_style hybrid/overlay pace table linear 10000',
-        #'pair_coeff * * pace c_v3.yace C', 'pair_coeff * * table d2.table D2 9.0']
-        #calc = LAMMPSlib(amendments=["mass 1 12.101"],lmpcmds = cmds,log_file='test.log',keep_alive=True)
+                           directions=[[1, 0, 0], [0, 1, 0], [0, 0, 1]])
+        # model = "c_v3.yace"
+        # d2_table = "d2.table"
+        # cmds = ['pair_style hybrid/overlay pace table linear 10000',
+        # 'pair_coeff * * pace c_v3.yace C', 'pair_coeff * * table d2.table D2 9.0']
+        # calc = LAMMPSlib(amendments=["mass 1 12.101"],lmpcmds = cmds,log_file='test.log',keep_alive=True)
         calc = Manybody(**TersoffBrenner(Erhart_PRB_71_035211_Si))
         unitcell.calc = calc
         ecf = ExpCellFilter(unitcell)
@@ -45,18 +45,20 @@ class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
 
     def test_surface_mapping_multilattice(self):
         self.setupmultilattice()
-        dir_vals = [[[1, 1, 0], [-1, 1, 0],[0,0,1]],
-                    [[1, 1, 1], [-2, 1, 1],[0,-1,1]]]
-        surf_dirs = [0,1,2]
+        dir_vals = [[[1, 1, 0], [-1, 1, 0], [0, 0, 1]],
+                    [[1, 1, 1], [-2, 1, 1], [0, -1, 1]]]
+        surf_dirs = [0, 1, 2]
         for dirs in dir_vals:
             for surf_dir in surf_dirs:
-                #print(f'testing relaxation of {dirs[surf_dir]} surface')
-                sr = SurfaceReconstruction(self.el,self.a0,self.calc,dirs,surf_dir,lattice=Diamond,multilattice=True)
-                sr.map_surface(fmax=0.0005,shift=0)
-                self.check_surface_mapping_multilattice(sr,dirs,Diamond,surf_dir)
-    
-    def check_surface_mapping_multilattice(self,sr,directions,lattice,surf_dir,shift=0):
-        cb = CubicCauchyBorn(self.el,self.a0,self.calc,lattice=lattice)
+                # print(f'testing relaxation of {dirs[surf_dir]} surface')
+                sr = SurfaceReconstruction(
+                    self.el, self.a0, self.calc, dirs, surf_dir, lattice=Diamond, multilattice=True)
+                sr.map_surface(fmax=0.0005, shift=0)
+                self.check_surface_mapping_multilattice(
+                    sr, dirs, Diamond, surf_dir)
+
+    def check_surface_mapping_multilattice(self, sr, directions, lattice, surf_dir, shift=0):
+        cb = CubicCauchyBorn(self.el, self.a0, self.calc, lattice=lattice)
 
         size = [1, 1, 1]
         size[surf_dir] *= 20
@@ -64,43 +66,47 @@ class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
         shift_vec = np.zeros([3])
[…]
