### 2023-07-19 #5668 @editorialbot
[REVIEW]: matscipy: materials science at the atomic scale with Python
**Submitting author:** @pastewka (<a href="http://orcid.org/0000-0001-8351-7336">Lars Pastewka</a>)
**Repository:** https://github.com/libAtoms/matscipy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
