### 2023-08-01 #5646 @pastewka
Yes, I do:

Pierre Hirel (Lille, maintains atoms package) - "pierrehirel"
Noam Bernstein (Naval Research Lab) - "bernstei"
Lucas Hale (NIST, maintains atomman package) - "lmhale99"
Markus Stricker (ICAMS Bochum) - "mastricker"
Gabor Csanyi (Cambridge) - "gabor1"
Alexander Stukowski (Ovito GmbH) - "stukowski"
