### 2023-08-15 #5646 @bernstei
> Hi @bernstei do you have time to review this paper?

Is there a full author list and/or abstract I can check before deciding?
