### 2023-08-22 8a8ff3d
MAINT: switch optimizer to preconditioned LBFGS for improved convergence
