### 2023-08-31 #5646 @diehlpk
> So, FWIW, I'd be willing to review, but I'm a pretty close collaborator of at least one of the authors (Kermode), and somewhat of a couple of others (Pastewka, Wengert). I feel like I could give an unbiased review, but you should let me know how you feel about that level of previous collaboration.
>
> Also, can you point me at the reviewer guidelines? It's not obvious to me to what extent I'm reviewing the text vs. the software itself, and how much responsibility I'm taking to actually run it as opposed to just make conclusions based on the available documentation.

Thanks for letting me know, but I think that is some COI.
