+++ matscipy/io/opls.py
@@ -44,29 +44,30 @@ except:
 
 def read_extended_xyz(fileobj):
     """
-    Read extended xyz file with labeled atoms. The number of atoms should
-    be given in the first line, the second line contains the cell
+    Read an extended xyz file with labeled atoms. The number of atoms
+    should be given in the first line, the second line contains the cell
     dimensions and the definition of the columns. The file should contain
     the following columns: element (1 or 2 characters), x(float), y(float),
     z (float), molecule id (int), name (1 or 2 characters). A full
     description of the extended xyz format can be found for example in the
     ASE documentation. An example for a file defining an H2 molecule is
     given below.
+    ::
+      2
+      Lattice="10.0 0.0 0.0 0.0 10.0 0.0 0.0 0.0 10.0" Properties=species:S:1:pos:R:3:molid:I:1:type:S:1
+      H 4.5 5.0 5.0 1 H1
+      H 5.5 5.0 5.0 1 H1
 
-    ====== Example ======
-    2
-    Lattice="10.0 0.0 0.0 0.0 10.0 0.0 0.0 0.0 10.0" Properties=species:S:1:pos:R:3:molid:I:1:type:S:1
-    H 4.5 5.0 5.0 1 H1
-    H 5.5 5.0 5.0 1 H1
-    ====== End of example ======
 
     Parameters
     ----------
     filename : str
+        Name of the file to read from.
 
     Returns
     -------
-    opls_struct : matscipy.opls.OPLSStructure
+    matscipy.opls.OPLSStructure
+        Atomic structure as defined in the input file.
     """
     atoms = ase.io.read(fileobj)
     opls_struct = matscipy.opls.OPLSStructure(atoms)
@@ -86,24 +87,34 @@ def read_extended_xyz(fileobj):
 def read_block(filename, name):
     """
     Read a named data block from a parameter file for a non-reactive
-    potential. Blocks begin with '# name' and are terminated by empty
-    lines. More information and an example can be found in the
-    documentation of the 'read_parameter_file' function.
+    potential. Blocks begin with ``# name`` and are terminated by empty
+    lines. Using the function on a file named ``parameterfile`` which
+    contains a block named ``Bonds`` and the following entries
+    ::
+      # Bonds
+      C1-C1 10.0 1.0
+      H1-H1 20.0 2.0
+
+    will return the following dictionary:
+    ::
+      {'C1-C1': [10., 1.], 'H1-H1': [20., 2.]}
 
     Parameters
     ----------
     filename : str
+        Name of the file to read from.
     name : str
+        Name of the data block to search for.
 
     Returns
     -------
-    data : dict
+    dict
         Name-Value pairs. Each value is a list of arbitrary length.
 
     Raises
     ------
     RuntimeError
-        If data block 'name' is not found in the file.
+        If data block ``name`` is not found in the file.
 
     """
     data = {}
@@ -146,23 +157,23 @@ def read_block(filename, name):
 def read_cutoffs(filename):
     """
     Read the cutoffs for construction of a non-reactive system from a
-    file. Comments in the file begin with '#', the file should be
+    file. Comments in the file begin with ``#``, the file should be
     structured like this:
-
-    ====== Example ======
-    # Cutoffs
-    C1-C1 1.23  # name, cutoff (A)
-    H1-H1 4.56  # name, cutoff (A)
-    C1-H1 7.89  # name, cutoff (A)
-    ====== End of example ======
+    ::
+      # Cutoffs
+      C1-C1 1.23  # name, cutoff (A)
+      H1-H1 4.56  # name, cutoff (A)
+      C1-H1 7.89  # name, cutoff (A)
 
     Parameters
[…]
