### 2023-08-31 9f7fe9e
DOC: Moved OPLS module examples to Jupyter notebook.
