### 2023-09-01 26d46bf
ENH: GammaSurface class
Class for generating gamma surface/generalised stacking fault images & plots
