+++ matscipy/gamma_surface.py
@@ -1,0 +1,314 @@
+import numpy as np
+from ase.build import rotate, cut
+from matscipy.dislocation import FixedLineAtoms
+from ase.optimize import BFGSLineSearch
+import warnings
+import matplotlib.pyplot as plt
+from ase.constraints import UnitCellFilter
+
+
+class GammaSurface():
+    '''
+    A class for generating gamma surface/generalised stacking fault images & plots
+    '''
+    def __init__(self, ats, surface_direction, y_dir=None, compress=True):
+        '''
+        Initialise by cutting and rotating the input structure.
+
+        ats: ase.Atoms
+            Starting structure to generate gamma surface from
+        surface_direction: np.array
+            Vector direction of gamma surface, in miller index notation
+            EG: np.array([0, 0, 1]), np.array([-1, 1, 0]), np.array([1, 1, 1])
+        y_dir: np.array | None
+            Basis vector (in miller indeces) to form "y" axis, which should be orthogonal to surface_direction
+            If None, a suitable y_dir will be found automatically
+        compress: bool
+            Search for "compressed" representation of surface where number of atoms
+            is minimised
+
+        Useful attributes:
+        self.cut_at
+            Cut Atoms object used as base for gamma surface image generation
+        self.surf_directions
+            Dict giving miller indeces for "x", "y" and "z" directions of gamma surface plot
+            -   self.surf_directions["z"] = surface_direction
+            -   self.surf_directions["y"] = y_dir, if y_dir was specified
+        self.cell_directions
+            Dict giving miller indeces for "x", "y" and "z" axes of self.cut_at, relative to the initial structure
+            Will differ from self.surf_directions if compress=True, and a smaller representation was found
+        self.nx, self.ny
+            Dimensions of the (nx, ny) gamma surface grid
+        self.images
+            Generated gamma surface images (populated after self.generate_images is called)
+        '''
+
+        self.images = []
+        self.nx = 0
+        self.ny = 0
+        self.x_disp = 0
+        self.y_disp = 0
+        self.surface_area = 0
+
+        def _cut_and_rotate(ats, surface_direction, y_dir):
+            # Cut such that surface_direction lies in z
+            at = cut(ats.copy(), a=surface_direction, b=y_dir)
+            rotate(at, [1, 0, 0], [0, 0, 1], [0, 1, 0], [0, 1, 0])
+            rotate(at, at.cell[0, :], [1, 0, 0], at.cell[1, :], [0, 1, 0])
+            return at
+
+        self._base_ats = ats.copy()
+
+        if y_dir is None:
+            _y_dir = self._y_dir_search(surface_direction)
+        else:
+            _y_dir = y_dir
+
+            if np.abs(np.dot(surface_direction, _y_dir)) >= 1E-3:
+                # Vector basis is not orthogonal
+                msg = f"y_dir vector {_y_dir} is not orthogonal to surface_direction vector {surface_direction}; dot(surface_direction, y_dir) = {float(np.dot(surface_direction, _y_dir))}\n" + \
+                    "Gamma Surface plot may not show the correct directions"
+                warnings.warn(msg, RuntimeWarning)
+
+        _x_dir = np.cross(_y_dir, surface_direction)
+
+        if not compress:
+            _cell_y_dir = _y_dir
+            _cell_x_dir = _x_dir
+
+            self.mapping = {
+                "x" : np.array([1, 0, 0]),
+                "y" : np.array([0, 1, 0])
+            }
+            self.cut_at = _cut_and_rotate(self._base_ats, surface_direction, _y_dir)
+        
+        else:
+            # Find linear combinations of _x_dir and _y_dir
[…]
