### 2023-09-01 11ddefe
ENH: StackingFault class
Shell class for convenient stacking fault images/measurements
