### 2023-09-05 f4854af
BUG: Gamma Surface Bug Fixes
