### 2023-09-07 c14e9ed
DOC: Fix author list
