### 2023-09-23 8a2fd25
TST: Added tests for Cauchy-Born shift gradient
