+++ tests/test_cauchy_born_corrector.py
@@ -232,6 +232,104 @@ class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
             dirs, eps, method='regression', E_func=func, coordinates=coordinates, atol=1e-4, returnvals=True)
         assert shift_err_after < shift_err_before
         # print(shift_err_before,shift_err_after)
+    
+    def E_cart3D_with_de(self, x, y, z, eps=None, de=0):
+        E = np.zeros([len(x), 3, 3])
+        E[:, 0, 0] = eps[0]
+        E[:, 1, 1] = eps[1]+de
+        E[:, 2, 2] = eps[2]
+        E[:, 1, 2], E[:, 2, 1] = eps[3], eps[3]
+        E[:, 0, 2], E[:, 2, 0] = eps[4], eps[4]
+        E[:, 0, 1], E[:, 1, 0] = eps[5], eps[5]
+        return E
+
+    def test_regression_model_gradient_E(self):
+        self.cb.initial_regression_fit()
+        eps = np.array([0.01, 0, 0.01, 0.01, 0, -0.01])
+        dirs = [[1, 1, 1], [-2, 1, 1], np.cross([1, 1, 1], [-2, 1, 1])]
+        func = self.E_cart3D_with_de
+        coordinates = 'cart3D'
+        #print('MAKING RAW PREDICION WITH NO ADDED EPS')
+        atoms, shifts, shift_err_before, A = self.model_prediction(
+            dirs, eps, method='regression', E_func=func, coordinates=coordinates, atol=1e-4, returnvals=True)
+        atoms_copy_init = atoms.copy()
+        self.cb.apply_shifts(atoms_copy_init,shifts)
+        #find the gradient using the strain function finite differences
+        de = 1e-5
+        nu_grad = self.cb.get_shift_gradients(A,atoms,\
+            E_func=func, coordinates='cart3D',eps=eps,de=de)
+
+        #find the gradients manually by finite differences
+        eps_down = eps.copy()
+        eps_down[1] -= de
+
+        atoms, shifts_down, shift_err_before, A = self.model_prediction(
+            dirs, eps_down, method='regression', E_func=func, coordinates=coordinates, atol=1e-4, returnvals=True)
+        atoms_copy_down = atoms.copy()
+        self.cb.apply_shifts(atoms_copy_down, shifts_down)
+        
+        eps_up = eps.copy()
+        eps_up[1] += de
+        atoms, shifts_up, shift_err_before, A = self.model_prediction(
+            dirs, eps_up, method='regression', E_func=func, coordinates=coordinates, atol=1e-4, returnvals=True)
+        
+        atoms_copy_up = atoms.copy()
+        self.cb.apply_shifts(atoms_copy_up, shifts_up)
+        #print(shifts_up-shifts_down)
+        nu_grad_fd = (atoms_copy_up.get_positions()-atoms_copy_down.get_positions())/(2*de)
+        #print(nu_grad_fd-nu_grad)
+        assert np.allclose(nu_grad_fd,nu_grad,1e-8)
+
+    def F_cart3D_with_de(self, x, y, z, eps=None, de=0):
+        E = np.zeros([len(x), 3, 3])
+        E[:, 0, 0] = eps[0]
+        E[:, 1, 1] = eps[1]+de
+        E[:, 2, 2] = eps[2]
+        E[:, 1, 2], E[:, 2, 1] = eps[3], eps[3]
+        E[:, 0, 2], E[:, 2, 0] = eps[4], eps[4]
+        E[:, 0, 1], E[:, 1, 0] = eps[5], eps[5]
+        U = np.zeros_like(E)
+        for i in range(np.shape(E)[0]):
+            U[i,:,:] = sqrtm((2*E[i, :, :])+np.eye(3))
+        return U
+
[…]
