+++ matscipy/surface_reconstruction.py
@@ -1,4 +1,5 @@
 import numpy as np
+from numpy import sqrt
 from ase.lattice.cubic import Diamond, FaceCenteredCubic, SimpleCubic, BodyCenteredCubic
 
 from ase.optimize import LBFGS
@@ -55,11 +56,13 @@ class SurfaceReconstruction:
         self.eval_cb = None
         self.lattice = lattice
         self.cb = None
+        self.multilattice = False
         if multilattice:
             self.cb = CubicCauchyBorn(
                 self.el, self.a0, self.calc, lattice=self.lattice)
+            self.multilattice = True
 
-    def map_surface(self, fmax=0.0001, layers=6, cutoff=10, shift=0):
+    def map_surface(self, fmax=0.0001, layers=6, cutoff=10, shift=0, switch=False):
         """Map the relaxation of the crystal surface a certain number of layers deep.
         Parameters
         ----------
@@ -67,12 +70,18 @@ class SurfaceReconstruction:
             Force tolerance for relaxation
         layers : int
             Number of layers deep to map the free surface
+        cutoff : float
+            Cutoff of the potential being used - used to decide the number of
+            atoms beneath the mapped layers that need to be modelled 
         shift : float
             Amount to shift the bulk structure before the surface is mapped
             such that the top layer of surface is physically meaningful. One
             use case of this would be to adjust the top layer of atoms being mapped
             in diamond such that it matches the top layer of atoms seen on the cleavage
             plane.
+        switch : bool
+            Whether or not to switch the sublattices around in the mapped structure.
+            (Essentially, the correct surface can always be found from some combination of shift and switch)
         """
         # build a single cell to determine how many atomic surface
         # layers there are in a single unit cell made by ASE
@@ -82,14 +91,21 @@ class SurfaceReconstruction:
             symbol=self.el,
             latticeconstant=self.a0,
             pbc=(1, 1, 1))
-        # ase.io.write(f'{self.directions[self.surf_dir]}_unitcell.xyz',single_cell)
+        ase.io.write(f'{self.directions[self.surf_dir]}_unitcell.xyz',single_cell)
 
-        if self.cb is not None:
+        if self.multilattice:
             # for a multilattice, get number of layers by just looking
             # at one of the sublattices
             self.cb.set_sublattices(single_cell, self.A)
             n_layer_per_cell = len(np.unique(single_cell.get_positions()[
                                    self.cb.lattice1mask][:, self.surf_dir].round(decimals=6)))
+            #check if multilattice atoms are coplanar
+            diff = single_cell.get_positions()[self.cb.lattice1mask][0, self.surf_dir] - single_cell.get_positions()[self.cb.lattice2mask][0, self.surf_dir]
+            print(diff)
+            if np.abs(diff)<0.001:
+                self.split_cb_pair = False
+            else:
+                self.split_cb_pair = True
         else:
             # otherwise, just look at all the layers directly
             n_layer_per_cell = len(np.unique(single_cell.get_positions()[
@@ -136,27 +152,29 @@ class SurfaceReconstruction:
         slab.set_constraint(FixAtoms(mask=mask))
         slab.calc = self.calc
         # print('fbefore',slab.get_forces())
-        # ase.io.write('0_slab.xyz',slab)
+        ase.io.write('0_slab.xyz',slab)
 
         # run an optimisation to relax the surface
         opt_slab = PreconLBFGS(slab)
         opt_slab.run(fmax=fmax)
-        # ase.io.write('1_slab.xyz',slab)
+        ase.io.write('1_slab.xyz',slab)
[…]
