### 2023-09-23 8c1dbea
WIP: Improved preconditioning, added shift gradients, added parallel functionality
