+++ matscipy/fracture_mechanics/crack.py
@@ -31,13 +31,14 @@ try:
     from scipy.optimize import brentq, leastsq, minimize, root
     from scipy.sparse import csc_matrix, spdiags
     from scipy.sparse.linalg import spsolve, spilu, LinearOperator
+    from scipy.optimize.nonlin import NoConvergence
 except ImportError:
     warnings.warn('Warning: no scipy')
 
 import ase.units as units
 from ase.optimize.precon import Exp
 #from ase.optimize.ode import ode12r
-
+from ase.optimize.sciopt import OptimizerConvergenceError
 from matscipy.atomic_strain import atomic_strain
 from matscipy.elasticity import (rotate_elastic_constants,
                                  rotate_cubic_elastic_constants,
@@ -53,6 +54,7 @@ from ase.constraints import FixAtoms
 from ase import Atom
 import ase.io
 
+import os
 ###
 
 # Constants
@@ -736,7 +738,8 @@ class SinclairCrack:
     def __init__(self, crk, cryst, calc, k, alpha=0.0, vacuum=6.0,
                  variable_alpha=True, variable_k=False,
                  alpha_scale=None, k_scale=None,
-                 extended_far_field=False,rI=0.0,rIII=0.0,cutoff=0.0):
+                 extended_far_field=False,rI=0.0,rIII=0.0,cutoff=0.0,
+                 incl_rI_f_alpha=False):
         """
 
         Parameters
@@ -751,6 +754,10 @@ class SinclairCrack:
         variable_k - if True, stress intensity factor can vary
                      (needed for arc-length continuation)
         extended_far_field - if True, crack tip force includes region III contrib
+        rI - Size of region I
+        rIII - Size of region III
+        cutoff - Potential cutoff
+        incl_rI_f_alpha - Whether to include region I when calculating the f_alpha forces
         """
         self.crk = crk # instance of CubicCrystalCrack
         self.cryst = cryst # Atoms object representing crystal
@@ -780,6 +787,11 @@ class SinclairCrack:
         self.rIII = rIII
         self.cutoff = cutoff
         self.shiftmask = None
+        self.f_alpha_vals = []
+        self.force_call_vals = []
+        self.norm_F_vals = []
+        self.alpha_vals = []
+        self.incl_rI_f_alpha = incl_rI_f_alpha
         # check atoms are sorted by distance from centre so we can use N1,N2,N3
         tip_x = cryst.cell.diagonal()[0] / 2.0
         tip_y = cryst.cell.diagonal()[1] / 2.0
@@ -803,6 +815,8 @@ class SinclairCrack:
         self.precon_count = 0
         self.f_alpha_correction = 0
         self.alpha0 = alpha
+        self.force_calls = 0
+
 
     def pack(self, u, alpha, k):
         dofs = list(u.reshape(-1))
@@ -907,8 +921,14 @@ class SinclairCrack:
                 raise RuntimeError('CLE fit failed')
         return res
 
-    def get_deformation_gradient(self,r,theta,k):
-        return self.crk.crack.deformation_gradient(r, theta, k)
+    def get_deformation_gradient(self,x,y,k=None,de=0):
+        alpha = self.alpha + de #take finite differences if necessary
+        if k is None:
+            k = self.k
+        tip_x = self.cryst.cell.diagonal()[0] / 2.0 + alpha
+        tip_y = self.cryst.cell.diagonal()[1] / 2.0
+        dg = self.crk.deformation_gradient(x, y ,tip_x, tip_y, k)
+        return dg
 
     def set_shiftmask(self,radial_dist):
         self.shiftmask = self.r>radial_dist
@@ -935,7 +955,7 @@ class SinclairCrack:
             # very important to pass cryst rather than atoms here as the displacement gradient field
             # is found from the positions of the original atoms, not the deformed atoms
[…]
