+++ scripts/fracture_mechanics/parallel_NCFlex.py
@@ -1,0 +1,734 @@
+import matscipy; print(matscipy.__file__)
+
+import os
+import numpy as np
+
+import h5py
+import ase.io
+from ase.units import GPa
+from ase.constraints import FixAtoms
+#from ase.optimize import LBFGSLineSearch
+from ase.optimize.precon import PreconLBFGS
+from ase.optimize import LBFGS
+
+from matscipy import parameter
+from matscipy.elasticity import  fit_elastic_constants
+from matscipy.fracture_mechanics.crack import CubicCrystalCrack, SinclairCrack
+
+from scipy.optimize import fsolve, fminbound, minimize
+from scipy.optimize.nonlin import NoConvergence
+
+import sys
+sys.path.insert(0, '.')
+
+import multiprocessing
+import params
+import time
+
+from queue import Empty
+from copy import deepcopy
+
+
+class RegressionModel:
+    def fit(self,alphas,Ks):
+        phi = self.basis_function_evaluation(alphas)
+        self.model = np.linalg.lstsq(phi,Ks,rcond=None)[0]
+        #print(phi,self.model)
+        
+    def predict(self,alphas):
+        phi = self.basis_function_evaluation(alphas)
+        #print(phi,self.model)
+        return phi @ self.model
+
+    def basis_function_evaluation(self,alphas):
+        if isinstance(alphas,np.ndarray):
+            xdim = len(alphas)
+        else:
+            xdim = 1
+        phi = np.zeros([xdim,3])
+        phi[:,0] = 1.0
+        phi[:,1] = np.sin(((2*np.pi)/alpha_period)*alphas)
+        phi[:,2] = np.cos(((2*np.pi)/alpha_period)*alphas)
+        return phi
+
+
+
+def init_worker():
+    global data_queue
+    print('worker initialised')
+    data_queue.put(os.getpid())
+
+def assign_sc(sc):
+    return [os.getpid(),sc]
+
+def assign_calc(sc):
+    #originally was passing around sc objects, but some
+    #calculators aren't pickleable
+    calc = params.new_calculator_instance()
+    sc.calc = calc
+    sc.atoms.calc = calc
+    sc.cryst.calc = calc
+    if sc.crk.cauchy_born is not None:
+        sc.crk.cauchy_born.calc = calc
+
+def delete_calc(sc):
+    del sc.calc
+    del sc.atoms.calc
+    del sc.cryst.calc
+    if sc.crk.cauchy_born is not None:
+        del sc.crk.cauchy_born
+
+def search(K0,alpha0,sc_dict):
+    global status_queue, search_queue
+    #set status as 'searching'
+
+    status_queue.put([os.getpid(),'searching'],block=False)
+    sc = sc_dict[os.getpid()]
+    assign_calc(sc)
+
+    print(f'pid {os.getpid()} sc is {sc} sc calc is {sc.calc}')
+    sc.variable_k = False
+    k1g = sc.k1g
+    sc.k = K0*k1g
+    sc.alpha = alpha0
+    sc.update_atoms()
+
+    converged = False
+    try:
+        sc.optimize(ftol=0.0005, steps=1000,method='ode12r',precon=True)
+    except:
+        print('did not converge fully in ode12r section, proceeding with krylov')
+
+    try:
+        sc.optimize(ftol=fmax, steps=100,method='krylov')
+        converged = True
+    except:
+        print('No convergence')
+
+    if converged:
+        k_x0 = sc.k/k1g
+        x0 = np.r_[sc.get_dofs(), k_x0 * k1g]
+        alpha_x0 = sc.alpha
+
[…]
