### 2023-09-25 7076a6b
BUG: Gamma Surface Bugfixes
