### 2023-09-26 10aa552
BUG: Changed default of cb to not be None due to errors with parameter
