### 2023-09-26 a4eae34
BUG: Fixed bug to ensure correct behaviour of alpha period finder  when surface reconstruction moves atoms outward
