+++ scripts/fracture_mechanics/parallel_NCFlex.py
@@ -683,7 +683,7 @@ if __name__ == '__main__':
     ypos = pos[:,1] - sy/2
 
     #find the closest y atoms by finding which atoms lie within 1e-2 of the min
-    closest_y_mask = np.abs(ypos)<(np.min(np.abs(ypos))+(1e-2))
+    closest_y_mask = np.abs(ypos)<(np.min(np.abs(ypos[xmask]))+(1e-2))
 
     #filter out all atoms with x less than 0
     xmask = xpos>0
@@ -692,11 +692,8 @@ if __name__ == '__main__':
     closest_x = xpos[closest_y_mask&xmask]
 
     #sort these atoms and find the largest x gap
-    print('closest_x', closest_x)
     sorted_x = np.sort(closest_x)
-    print('sorted_x', sorted_x)
     diffs = np.diff(sorted_x)
-    print('diffs',diffs)
     alpha_period = np.sum(np.unique(np.round(np.diff(sorted_x),decimals=4)))
     print('alpha_period',alpha_period)
     # setup the crack
