### 2023-09-26 a54fb20
BUG: Remove GammaSurface compression, as not functional
