+++ matscipy/gamma_surface.py
@@ -11,7 +11,7 @@ class GammaSurface():
     '''
     A class for generating gamma surface/generalised stacking fault images & plots
     '''
-    def __init__(self, ats, surface_direction, y_dir=None, compress=True):
+    def __init__(self, ats, surface_direction, y_dir=None):
         '''
         Initialise by cutting and rotating the input structure.
 
@@ -23,9 +23,6 @@ class GammaSurface():
         y_dir: np.array | None
             Basis vector (in miller indeces) to form "y" axis, which should be orthogonal to surface_direction
             If None, a suitable y_dir will be found automatically
-        compress: bool
-            Search for "compressed" representation of surface where number of atoms
-            is minimised
 
         Useful attributes:
         self.cut_at
@@ -71,52 +68,16 @@ class GammaSurface():
 
         _x_dir = np.cross(_y_dir, surface_direction)
 
-        if not compress:
-            _cell_y_dir = _y_dir
-            _cell_x_dir = _x_dir
+    
+        _cell_y_dir = _y_dir
+        _cell_x_dir = _x_dir
 
-            self.mapping = {
-                "x" : np.array([1, 0, 0]),
-                "y" : np.array([0, 1, 0])
-            }
-            self.cut_at = _cut_and_rotate(self._base_ats, surface_direction, _y_dir)
-        
-        else:
-            # Find linear combinations of _x_dir and _y_dir
-            d4_vecs = self._find_new_axes(_y_dir, _x_dir)
-
-            # Test each comb to find smallest representation
-            at_len = np.inf
-            for vec in d4_vecs:
-                test_cut = _cut_and_rotate(self._base_ats, surface_direction, vec)
-                if len(test_cut) < at_len:
-                    at_len = len(test_cut)
-                    cand_ats = test_cut.copy()
-                    _cell_y_dir = vec
-
-            _cell_x_dir = np.cross(_cell_y_dir, surface_direction)
-
-            if y_dir is None:  # No user specified y_dir, use most compressed direction instead
-                _y_dir = _cell_y_dir
-                _x_dir = _cell_x_dir
-
-            # Define mapping between cell coord system andngamma surface coord system
-            self.mapping = {
-                "x": np.linalg.solve(np.array([_cell_x_dir, _cell_y_dir, surface_direction]).T, _x_dir),
-                "y": np.linalg.solve(np.array([_cell_x_dir, _cell_y_dir, surface_direction]).T, _y_dir)
-            }
-
-            x_sgn = np.sign(self.mapping["x"])[0]
-            y_sgn = np.sign(self.mapping["y"])[0]
-
-            if x_sgn < 0:
-                self.mapping["x"] *= x_sgn
-
-            if y_sgn < 0:
-                self.mapping["y"] *= y_sgn
-                
-            self.cut_at = cand_ats
-        
+        self.mapping = {
+            "x" : np.array([1, 0, 0]),
+            "y" : np.array([0, 1, 0])
+        }
+        self.cut_at = _cut_and_rotate(self._base_ats, surface_direction, _y_dir)
+    
         self.surf_directions = {
             "x": _x_dir,
             "y": _y_dir,
@@ -156,32 +117,6 @@ class GammaSurface():
         
         # No nice integer vector found!
         raise RuntimeError(f"Could not automatically find an integer basis from basis vector {d1}")
-
-    def _find_new_axes(self, d2, d3, depth=3):
-        '''
-        Search for new orthogonal vectors d4, d5 via:
-
-        d4 = a_1 * d2 + a_2 * d3
-        d5 = cross(d4, d1)
-
-        depth controls the maximum allowed magnitudes of a_1, a_2
-
[…]
