### 2023-09-26 1d036e9
BUG: Fixed crash when searching for new K, alpha when no value yet found
