### 2023-09-27 eb581b7
BUG: Fixed bug with xmask not defined
