### 2023-09-27 822e613
BUG: Fixed divide by 0 error if no K values available
