### 2023-10-03 471cd78
WIP: Added functions for building 3D sincliar kink-crack geometries
