+++ matscipy/fracture_mechanics/clusters.py
@@ -24,8 +24,163 @@ import numpy as np
 
 from ase.lattice.cubic import Diamond, FaceCenteredCubic, SimpleCubic, BodyCenteredCubic
 from matscipy.neighbours import neighbour_list
+import ase.io
+####
 
-###
+def get_alpha_period(cryst):
+    pos = cryst.get_positions()
+    sx, sy, sz = cryst.cell.diagonal()
+    xpos = pos[:,0] - sx/2
+    ypos = pos[:,1] - sy/2
+    #find the closest y atoms by finding which atoms lie within 1e-2 of the min
+    #filter out all atoms with x less than 0
+    xmask = xpos>0
+    closest_y_mask = np.abs(ypos)<(np.min(np.abs(ypos[xmask]))+(1e-2))
+    #find the x positions of these atoms
+    closest_x = xpos[closest_y_mask&xmask]
+    #sort these atoms and find the largest x gap
+    sorted_x = np.sort(closest_x)
+    diffs = np.diff(sorted_x)
+    alpha_period = np.sum(np.unique(np.round(np.diff(sorted_x),decimals=4)))
+    return alpha_period
+
+def generate_3D_structure(cryst_2D,nzlayer,el,a0,lattice,crack_surface,crack_front,shift=np.array([0.0,0.0,0.0]),cb=None,switch_sublattices=False):
+    alpha_period = get_alpha_period(cryst_2D)
+    #make a single layer of cell 
+    single_cell = lattice(el,a0,[1,1,1],crack_surface,crack_front,cb=cb,shift=shift,switch_sublattices=switch_sublattices)
+    ase.io.write('single_cell.xyz',single_cell)
+    cell_x_period = single_cell.get_cell()[0,0]
+    print('NUM KINK PER CELL,', cell_x_period/alpha_period)
+    #original x,y dimensions
+    big_cryst = cryst_2D*[1,1,nzlayer]
+    og_cell = big_cryst.get_cell()
+    h = og_cell[2,2]
+    theta = np.arctan(cell_x_period/h)
+    pos = big_cryst.get_positions()
+    xpos = pos[:,0]
+    zpos = pos[:,2]
+    cell_xlength = og_cell[0,0]
+    ztantheta = zpos*np.tan(theta)
+    mask = xpos>(cell_xlength - ztantheta)
+    pos[mask,0] = pos[mask,0] - cell_xlength
+    big_cryst.set_positions(pos)
+    new_cell = og_cell.copy()
+    new_cell[2,0] = -cell_x_period
+    ase.io.write('big_cryst_pre_rotation.xyz',big_cryst)
+    big_cryst.set_cell(new_cell)
+    ase.io.write('big_cryst.xyz',big_cryst)
+
+    # Define the rotation matrix
+    R = np.array([[np.cos(theta), 0, np.sin(theta)],
+            [0, 1, 0],
+            [-np.sin(theta), 0, np.cos(theta)]])
+
+    # Get the cell of big_cryst
+    cell = big_cryst.get_cell()
+
+    # Rotate the cell using the rotation matrix
+    rotated_cell = np.dot(cell, R.T)
+
+    # Set the new cell of big_cryst
+    big_cryst.set_cell(rotated_cell,scale_atoms=True)
+    return big_cryst
+
+def generate_3D_cubic_111(cryst_2D, nzlayer, el, a0, lattice, crack_surface, crack_front, shift=np.array([0.0,0.0,0.0]), cb=None, switch_sublattices=False):
+    """
+    Generate a kink-periodic cell, using the high symmetry of the 111 cubic surface
+    to reduce the number of kinks in the cell
+
+    Parameters:
+    -----------
+    cryst_2D : ASE atoms object
+        The 2D cryst structure for the crack to use as a template for the 3D structure.
+    nzlayer : int
[…]
