### 2023-10-05 1349682
ENH: Gamma Surface Vacuum layers
