+++ matscipy/gamma_surface.py
@@ -58,7 +58,7 @@ class GammaSurface():
         if y_dir is None:
             _y_dir = self._y_dir_search(surface_direction)
         else:
-            _y_dir = y_dir
+            _y_dir = np.array(y_dir)
 
             if np.abs(np.dot(surface_direction, _y_dir)) >= 1E-3:
                 # Vector basis is not orthogonal
@@ -79,14 +79,14 @@ class GammaSurface():
         self.cut_at = _cut_and_rotate(self._base_ats, surface_direction, _y_dir)
     
         self.surf_directions = {
-            "x": _x_dir,
-            "y": _y_dir,
-            "z": surface_direction
+            "x": np.array(_x_dir),
+            "y": np.array(_y_dir),
+            "z": np.array(surface_direction)
             }
         self.cell_directions = {
-            "x": _cell_x_dir,
-            "y": _cell_y_dir,
-            "z": surface_direction
+            "x": np.array(_cell_x_dir),
+            "y": np.array(_cell_y_dir),
+            "z": np.array(surface_direction)
         }
 
     def _y_dir_search(self, d1):
@@ -118,7 +118,7 @@ class GammaSurface():
         # No nice integer vector found!
         raise RuntimeError(f"Could not automatically find an integer basis from basis vector {d1}")
 
-    def generate_images(self, nx, ny, z_replications=1, vert_strain=0.0):
+    def generate_images(self, nx, ny, z_replications=1, atom_offset=0.0, cell_strain=0.0, vacuum=0.0, vacuum_offset=0.0, path_limsx=[0.0, 1.0], path_limsy=None):
         '''
         Generate gamma surface images on an (nx, ny) grid
 
@@ -128,36 +128,81 @@ class GammaSurface():
             Number of points in the y direction
         z_replications: int
             Number of supercell copies in z (increases separation between periodic surfaces)
-        vert_strain: float
-            Percentage strain to apply in z direction (0.1 = +10% strain)
+        atom_offset: float
+            Offset in the z direction (in A) to apply to all atoms
+        cell_strain: float
+            Fractional strain to apply in z direction to cell only (0.1 = +10% strain)
+        vacuum: float
+            Additional vacuum layer (in A) to add between periodic images of the gamma surface
+        vacuum_offset: float
+            Offset (in A) applied to the position of the vacuum layer in the cell
+            The position of the vacuum layer is given by:
+            vac_pos = self.cut_at.cell[2, 2] * (1 + cell_strain) / 2 + vacuum_offset
+        path_limsx: list|array of floats
+            Limits (in fractional coordinates) of the stacking fault path in the x direction
+        path_limsy: list|array of floats
+            Limits (in fractional coordinates) of the stacking fault path in the x direction
+            If not supplied, will be set to path_limsx
         '''
+
+        if path_limsy is None:
+            y_lims = path_limsx
+        else:
+            y_lims = path_limsy
+        
+        x_lims = path_limsx
+
         self.nx = nx
         self.ny = ny
 
         base_struct = self.cut_at * (1, 1, z_replications)
 
-        cell = base_struct.cell[:, :]
+        # Atom offset
+        pos = base_struct.get_positions()
+        pos[:, 2] += atom_offset
+        base_struct.set_positions(pos)
+        base_struct.wrap()
+        
+        # Apply cell strain
+        new_cell = base_struct.cell[:, :].copy()
+        new_cell[2, 2] += cell_strain
+        base_struct.set_cell(new_cell, scale_atoms=False)
 
+        # Add vacuum
[…]
