+++ tests/test_gamma_surface.py
@@ -9,127 +9,34 @@ from quippy.potential import Potential
 
 class GammaSurfaceTest(matscipytest.MatSciPyTestCase):
     def setUp(self) -> None:
-            self.fmax = 1e-4
-            self.at0 = Diamond('Si', latticeconstant=5.43)
+        self.fmax = 1e-3
+        self.at0 = Diamond('Si', latticeconstant=5.43)
 
-            # Si testing framework SW model string
-            # https://github.com/libAtoms/silicon-testing-framework/blob/master/models/SW/model.py
-            self.model = Potential('IP SW', param_str="""
-                    <SW_params n_types="1">
-                    <per_type_data type="1" atomic_num="14" />
-                    <per_pair_data atnum_i="14" atnum_j="14" AA="7.049556277" BB="0.6022245584"
-                        p="4" q="0" a="1.80" sigma="2.0951" eps="2.1675" />
-                    <per_triplet_data atnum_c="14" atnum_j="14" atnum_k="14"
-                        lambda="21.0" gamma="1.20" eps="2.1675" />
-                    </SW_params>
-                    """)
+        # Si testing framework SW model string
+        # https://github.com/libAtoms/silicon-testing-framework/blob/master/models/SW/model.py
+        self.model = Potential('IP SW', param_str="""
+            <SW_params n_types="1">
+            <per_type_data type="1" atomic_num="14" />
+            <per_pair_data atnum_i="14" atnum_j="14" AA="7.049556277" BB="0.6022245584"
+                p="4" q="0" a="1.80" sigma="2.0951" eps="2.1675" />
+            <per_triplet_data atnum_c="14" atnum_j="14" atnum_k="14"
+                lambda="21.0" gamma="1.20" eps="2.1675" />
+            </SW_params>
+            """)
     
     def test_gamma_surface(self):
-          # Test functionality of gamma surface code + using model to relax
+        surface = GammaSurface(self.at0, np.array([0, 0, 1], dtype=int), np.array([1, 0, 0], dtype=int))
+        surface.generate_images(3, 3)
+        surface.relax_images(self.model, self.fmax)
+        Es = surface.get_surface_energies(self.model)
 
-            surface = GammaSurface(self.at0, np.array([0, 0, 1], dtype=int), np.array([1, 0, 0], dtype=int))
-            surface.generate_images(7, 7)
-            surface.relax_images(self.model, self.fmax)
-            Es = surface.get_surface_energies(self.model)
-
-            assert np.min(Es) >= 0.0
-            print(np.max(Es))
-    
+        assert np.min(Es) >= 0.0
+        assert np.allclose(np.max(Es), 0.1946074084327025)
 
     def test_stacking_fault(self):
-          import matplotlib.pyplot as plt
-          from ase.io import write
-          # Test that stacking fault reproduces Si results from Si testing framework for SW potential
-          surface = StackingFault(self.at0, np.array([0, 0, 1]), np.array([1, 1, 0]), compress=True)
-          surface.generate_images(21, z_replications=1)
-          surface.relax_images(self.model, self.fmax)
-          Es = surface.get_surface_energies(self.model)
-          
-          surface.plot_gamma_surface()
-          plt.savefig("Stack_Compressed.png")
+        surface = StackingFault(self.at0, np.array([1, 1, 0]), np.array([-1, 1, 0]))
+        surface.generate_images(9, z_replications=1, path_ylims=[0, 0.5])
+        surface.relax_images(self.model, self.fmax)
+        Es = surface.get_surface_energies(self.model)[0, :]
 
-          Es1 = Es.copy()[0, :]
[…]
