### 2023-10-06 5ecb44c
WIP: added explore direction choice
