### 2023-10-06 cbc3bcb
Add depreciation warning on gamma_line
