### 2023-10-09 fe37ddc
Add EU funders to JOSS acknowlegements
Add grant references for NOMAD (@jameskermode and others) and ENTENTE (@lakshenoy and @jameskermode ) to JOSS acknowlegements sections.
