+++ matscipy/surface_reconstruction.py
@@ -491,14 +491,14 @@ class SurfaceReconstruction:
                     directions=self.pandey_dirs
                     )
 
-
         self.cb.set_sublattices(a,self.U)
         if switch:
             self.cb.switch_sublattices(a)
         sorted_z_vals = np.sort(a.get_positions()[self.cb.lattice1mask][:,2])
         self.inter_surf_dist = sorted_z_vals[2] - sorted_z_vals[0]
-        a.translate([0,0,-shift*self.inter_surf_dist])
-
+        #a.translate([0,+0.001,-shift*self.inter_surf_dist]) #REMOVE SHIFT FROM HERE
+        a.translate([0.01,0.01,0]) 
+        a.wrap()
         print('INTER SURF DIST',self.inter_surf_dist)
         #seed the 2x1 reconstruction on the top plane
         sx, sy, sz = a.get_cell().diagonal()
@@ -517,14 +517,16 @@ class SurfaceReconstruction:
                                                     np.logical_not(mask))]
         a.set_distance(top1, top2, bondlen)
         a.set_distance(topA, topB, bondlen)
-
+        #ase.io.write('intermediate_atoms.xyz',a)
         x, y, z = a.positions.T
         y1 = (y[top1]+y[top2])/2
         yA = (y[topA]+y[topB])/2
         y[top1] += yA-y1+a.cell[1,1]/2
         y[top2] += yA-y1+a.cell[1,1]/2
+        #x[top1] += a.cell[0,0]/2
+        #x[topB] -= a.cell[0,0]/2
         a.set_positions(np.transpose([x, y, z]))
-        a.wrap()
+        #a.wrap()
 
         ase.io.write('0_bulk.xyz',bulk)
         ase.io.write('1_bulk.xyz',a)
@@ -547,6 +549,20 @@ class SurfaceReconstruction:
         opt_a.run(fmax=fmax)
         ase.io.write('2_bulk.xyz',a)
         pos_after_unrotated = a.get_positions()
+        print(pos_after_unrotated[:,1]<0)
+        print(len(pos_after_unrotated[:,1]<0))
+        print(pos_after_unrotated[:,1]>a.cell[1,1])
+        print(len(pos_after_unrotated[:,1]>a.cell[1,1]))
+        atoms_outside_mask = (pos_after_unrotated[:,1]<0) | (pos_after_unrotated[:,1]>a.cell[1,1])
+        pushed_out_atoms = a[atoms_outside_mask]
+        wrapped=False
+        if len(pushed_out_atoms)>0:
+            print('detected atoms to be wrapped')
+            wrapped=True
+            a.wrap()
+            pos_after_unrotated = a.get_positions()
+            ase.io.write('3_bulk.xyz',a)
+        
         pos_diff_unrotated = pos_after_unrotated-pos_initial_unrotated
         #print('pos diff unrotated',pos_diff_unrotated)
         #map out the layers in the original structure
@@ -556,6 +572,69 @@ class SurfaceReconstruction:
         layer_mask_set_lattice1_atom1, layer_mask_set_lattice1_atom2,layer_mask_set_lattice2_atom1, layer_mask_set_lattice2_atom2 = \
             self.identify_pandey_layers(bulk,surface_coords,read_from_atoms=True,frame_dirs=self.pandey_dirs)
         self.surf_dir = tmp
+
+        """
+        #The following is code written in order to minimise the amount that atoms
+        move during the relaxation- this is not necessary when there is no chance 
+        that the position of xlim always lies at the edge of a pandey unit cell
+        #also, i think actually that the atoms do end up in positions that minimise the travel
+        distance anyway
+
+        a.new_array('atomno',np.zeros(len(a),dtype=int))
+        a_atomno = a.arrays['atomno']
[…]
