### 2023-10-09 110a436
WIP: Added functionality to create kink-periodic fracture cells for NCflex
