### 2023-10-10 8680f8f
WIP: Added parallel script that allows for measurement of simultaneous NCFlex curves
