+++ scripts/fracture_mechanics/parallel_NCFlex_seperate_curves.py
@@ -1,0 +1,435 @@
+import matscipy; print(matscipy.__file__)
+
+import os
+import numpy as np
+
+import h5py
+import ase.io
+from ase.units import GPa
+from ase.constraints import FixAtoms
+#from ase.optimize import LBFGSLineSearch
+from ase.optimize.precon import PreconLBFGS
+from ase.optimize import LBFGS
+
+from matscipy import parameter
+from matscipy.elasticity import  fit_elastic_constants
+from matscipy.fracture_mechanics.crack import CubicCrystalCrack, SinclairCrack
+
+from scipy.optimize import fsolve, fminbound, minimize
+from scipy.optimize.nonlin import NoConvergence
+
+import sys
+sys.path.insert(0, '.')
+
+import multiprocessing
+import params
+import time
+
+from queue import Empty
+from copy import deepcopy
+
+
+
+def init_worker():
+    global data_queue
+    print('worker initialised')
+    data_queue.put(os.getpid())
+
+def assign_sc(sc):
+    return [os.getpid(),sc]
+
+def assign_calc(sc):
+    #originally was passing around sc objects, but some
+    #calculators aren't pickleable
+    calc = params.new_calculator_instance()
+    sc.calc = calc
+    sc.atoms.calc = calc
+    sc.cryst.calc = calc
+    if sc.crk.cauchy_born is not None:
+        sc.crk.cauchy_born.calc = calc
+
+def delete_calc(sc):
+    del sc.calc
+    del sc.atoms.calc
+    del sc.cryst.calc
+    if sc.crk.cauchy_born is not None:
+        del sc.crk.cauchy_born
+
+def search(K0,alpha0,sc_dict):
+    global status_queue, search_queue
+    #set status as 'searching'
+
+    status_queue.put([os.getpid(),'searching'],block=False)
+    sc = sc_dict[os.getpid()]
+    assign_calc(sc)
+
+    print(f'pid {os.getpid()} sc is {sc} sc calc is {sc.calc}')
+    sc.variable_k = False
+    k1g = sc.k1g
+    sc.k = K0*k1g
+    sc.alpha = alpha0
+    sc.update_atoms()
+
+    converged = False
+    try:
+        sc.optimize(ftol=0.0005, steps=1000,method='ode12r',precon=True)
+    except:
+        print('did not converge fully in ode12r section, proceeding with krylov')
+
+    try:
+        sc.optimize(ftol=fmax, steps=100,method='krylov')
+        converged = True
+    except:
+        print('No convergence')
+
+    if converged:
+        k_x0 = sc.k/k1g
+        x0 = np.r_[sc.get_dofs(), k_x0 * k1g]
+        alpha_x0 = sc.alpha
+
+        print(f'Rescaling K_I from {sc.k} to {sc.k + dk * k1g}')
+        k_x1 = k_x0 + dk
+        print(k0)
+        sc.k = k_x1*k1g
+        sc.update_atoms()
+        sc.optimize(ftol=0.0001, steps=100,method='krylov')
+        x1 = np.r_[sc.get_dofs(), k_x1 * k1g]
+        alpha_x1 = sc.alpha
+        print(f'k0={k_x0}, k1={k_x1} --> alpha0={alpha_x0}, alpha1={alpha_x1}')
+        
+    else:
+        #return sentinal values
+        #FIXME really this should do something more drastic
+        x0 = [-10000,-10000]
+        x1 = [-10000,-10000]
+
+    #place result in search queue (blocking)
+    print('Braodcasting search result')
+    search_queue.put([os.getpid(),[x0,x1]])
+
+    delete_calc(sc)
+    #at end,
[…]
