### 2023-10-10 7072308
BUG: Fixed bug where occasionally manager would recieve message from dead workers
