+++ matscipy/dislocation.py
@@ -40,6 +40,8 @@ from ase.io import read
 
 from matscipy.neighbours import neighbour_list, mic
 from matscipy.elasticity import fit_elastic_constants
+from matscipy.elasticity import Voigt_6x6_to_full_3x3x3x3
+from matscipy.elasticity import cubic_to_Voigt_6x6
 
 
 def make_screw_cyl(alat, C11, C12, C44,
@@ -2147,6 +2149,195 @@ def check_duplicates(atoms, distance=0.1):
     # print(f"found {duplicates.sum()} duplicates")
     return duplicates.astype(np.bool)
 
+class CubicCrystalDislocationGrad:
+    """
+    Displacement and displacement gradient field of straight 
+    dislocation in anisotropic elastic media.
+    Ref: pp. 467 in J.P. Hirth and J. Lothe, Theory of Dislocations, 2nd ed.
+    The function is similar as class `CubicCrystalDislocation`.
+    """
+
+    def __init__(self, a0, crystal_surface, crack_front, n, xi, Burgers, C11=None, C12=None,
+                 C44=None):
+        """
+        Insert a dislocation in a cubic crystal containing a crack. 
+        Elastic constants are C11, C12 and C44 (or optionally a full 6x6 elastic constant matrix C).
+        The dislocation is defined by m (dislocation glide direction), n (the slip plane normal), 
+        and xi (dislocation line direction).
+        """
+
+        # x (third_dir) - direction in which the crack is running
+        # y (crack_surface) - free surface that forms due to the crack
+        # z (crack_front) - direction of the crack front
+        # normalize axis of crack coordinates 
+        third_dir = np.cross(crack_surface, crack_front)
+        third_dir = np.array(third_dir) / np.sqrt(np.dot(third_dir,
+                                                         third_dir))
+        crack_surface = np.array(crack_surface) / \
+            np.sqrt(np.dot(crack_surface, crack_surface))
+        crack_front = np.array(crack_front) / \
+            np.sqrt(np.dot(crack_front, crack_front))
+        
+        # define rotation matrix from dislocation system to crack system
+        A = np.array([third_dir, crack_surface, crack_front])
+        if np.linalg.det(A) < 0:
+            third_dir = -third_dir
+        A = np.array([third_dir, crack_surface, crack_front])
+        m = np.cross(n, xi)
+
+        # normalize m, n and xi
+        m = np.array(m) / np.sqrt(np.dot(m, m))
+        n = np.array(n) / np.sqrt(np.dot(n, n))
+        xi = np.array(xi) / np.sqrt(np.dot(xi, xi))
+        disloc_sys = np.array([m, n, xi])
+        if np.linalg.det(disloc_sys) < 0:
+            m = -m        
+        
+        # Rotate vectors from dislocation system to crack coordinates
+        self.m = np.einsum('ij,j', A, m)
+        self.n = np.einsum('ij,j', A, n)
+        self.xi = np.einsum('ij,j', A, xi)
+        self.b = np.einsum('ij,j', A, Burgers)
+        
+        # define elastic constant in Voigt notation
+        C_v = cubic_to_Voigt_6x6(C11, C12, C44)
+        # convert Voigt to tensor
+        C = Voigt_6x6_to_full_3x3x3x3(C_v)
+        # rotate elastic matrix
+        cijkl = np.einsum('ig,jh,ghmn,km,ln', \
+            A, A, C, A, A)
+    
[…]
