### 2023-10-16 libAtoms/matscipy#183 @mbarzegary
[JOSS Review] Paper and statement of need
Dear authors,

The provided paper is comprehensive and well-written. However, these comments may help improve it:
* As mentioned in the paper, there exists other packages for multi-scale coupling such as LibMultiScale and MultiBench, but the necessity of the proposed package is not well elaborated. Since it’s the most important paragraph of the statement of need, I appreciate it if the authors elaborate on this.
* I see duplicate references. For example, there are two identical Seidel et al. (2021a and 2021b). Can you please check the paper for this?
* The authors have referred to 2 papers using the coupling of the numerical models of the Poisson–Nernst–Planck equation in electrochemistry (Hörmann et al., 2023; Seidl et al. 2021b), but it doesn't seem that those papers use this coupling. Can you please elaborate on the mentioned usage?

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5668
