### 2023-10-16 libAtoms/matscipy#184 @mbarzegary
 [JOSS Review] Reproducibility and functionality
Dear authors,

Generally speaking, the software runs more or less as described in the documentation. However, there are some comments in this regard:
* It's not clear where the user should start using the package after installing it. The way the provided examples are placed is a bit difficult to follow. Except for the electrochemistry and the fracture_mechanics directories, the rest of the examples directories are not clear regarding which functionality they belong to. I suggest that you add a new example directory for getting started with the package, putting some basic examples grouped per feature.
* It's not clear how the distributed computation example is supposed to work.
* Generally speaking, the example directory contains a large number of non-trivial files, most of which  are not documented unfortunately. I think this makes it difficult to get started with the package.
* Running the `samples_pb_c2d.py` and `samples_pnp_c2d.py` examples results in the following error: `ModuleNotFoundError: No module named 'matscipy.electrochemistry.utility'`. The error for `samples_pnp_by_fenics_fem.py` is related to lack of `matscipy.electrochemistry.poisson_nernst_planck_solver_fenics` module.
* Some examples rely on IPython, which doesn't get installed via the installation procedure of the package.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5668
