### 2023-10-16 libAtoms/matscipy#185 @mbarzegary
[JOSS Review] Tests
Dear authors,

The software has standard and comprehensive automated tests implemented using pytest. However, these minor points can be taken into account in this regard:
* I cannot find the FEniCS functionality being tested via the pytest tests. Is it because of the complicated installation procedure? I care about this functionality not only due to its usefulness but also because it's mentioned in the paper.
* All tests pass on my local machine, but some of them produce unclear characters. This is the output of the pytest run:
```
test_angle_distribution.py ....                                                                                             [  0%]
test_atomic_strain.py .                                                                                                     [  0%]
test_bop.py ............................................................................................................... [ 17%]
......                                                                                                                      [ 18%]
test_bulk_properties.py .............                                                                                       [ 20%]
test_c2d.py ......                                                                                                          [ 21%]
test_cauchy_born_corrector.py ........                                                                                      [ 22%]
test_committee.py ............                                                                                              [ 24%]
test_cubic_crystal_crack.py ......                                                                                          [ 25%]
test_dislocation.py ssssssssssssss.sssssssss.ss.ssssssss.s                                                                  [ 30%]
test_eam_average_atom.py ..                                                                                                 [ 30%]
test_eam_calculator.py .......                                                                                              [ 32%]
test_eam_calculator_forces_and_hessian.py ......                                                                            [ 32%]
test_eam_io.py ....                                                                                                         [ 33%]
test_elastic_moduli.py ..                                                                                                   [ 33%]
test_energy_release.py s                                                                                                    [ 33%]
test_ewald.py .................................                                                                             [ 38%]
test_ffi.py .                                                                                                               [ 39%]
test_full_to_Voigt.py ....                                                                                                  [ 39%]
test_hessian_finite_differences.py ...                                                                                      [ 40%]
test_hessian_precon.py ...                                                                                                  [ 40%]
test_hydrogenate.py .                                                                                                       [ 40%]
test_idealbrittlesolid.py ...                                                                                               [ 41%]
test_invariants.py .                                                                                                        [ 41%]
test_io.py .....                                                                                                            [ 42%]
test_mcfm.py ...                                                                                                            [ 42%]
test_neighbours.py ...................                                                                                      [ 45%]
test_numpy_tricks.py .                                                                                                      [ 45%]
test_opls.py .....                                                                                                          [ 46%]
test_opls_io.py .......                                                                                                     [ 47%]
test_pair_potential_calculator.py ......................                                                                    [ 50%]
test_poisson_nernst_planck_solver.py .                                                                                      [ 50%]
test_polydisperse_calculator.py ............                                                                                [ 52%]
test_pressurecoupling.py ....                                                                                               [ 53%]
test_ring_statistics.py .....                                                                                               [ 53%]
test_rotation_of_elastic_constants.py .                                                                                     [ 54%]
test_spatial_correlation_function.py ..                                                                                     [ 54%]
test_supercell_calculator.py .                                                                                              [ 54%]
test_surface_reconstruction.py ..                                                                                           [ 54%]
manybody/test_manybody_molecules.py .........xxxxxxxxx                                                                      [ 57%]
manybody/test_manybody_potentials.py x.x..x..x..........                                                                    [ 60%]
manybody/test_newmb.py XxXx.x............XXXX.X............XxXx.x............XXXX.X............XxXx.x............XXXX.X.... [ 75%]
........xxxx.x............xxxx.x............xxxx.x............xxxx.x............XxXx.x............XXXX.X............xxxx.x. [ 93%]
```

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5668
