### 2023-10-16 libAtoms/matscipy#186 @mbarzegary
[JOSS Review] Documentation 
Dear authors,

The documentation of the software is nicely done, but you may consider improving it with these suggestions:
* Some of the mentioned functionalities of the package are not documented. For example, there is no documentation for the electrochemistry. Also, some of the atomic analysis tools and all the interatomic potential features are only documented via autogenerated API docs. I think this makes it hard for beginner users to get started with the package. The application domains (except the electrochemistry, as mentioned) are well documented and described. It can be beneficial to talk a little bit about the package features in the documentation.
* It's mentioned that the package has an interface to the FEniCS code. This functionality is supported by one example that does not run. FEniCS dependency is not documented and elaborated.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5668
