### 2023-10-17 a195858
fix: protecting sympy imports in manybody potential tests
