+++ matscipy/dislocation.py
@@ -2971,6 +2971,104 @@ class FCCEdge110Dislocation(CubicCrystalDissociatedDislocation):
         super().__init__(left_shockley, right_shockley, burgers)
 
 
+def _validate_disloc_inputs(a, symbol="w", axes=None, crystalstructure="fcc", pbc=[True, True, True]):
+    ''''
+    Validate & transform args of dislocation class __init__ methods
+
+    a: float or ase.atoms
+        EITHER lattice constant (in A), or the cubic bulk structure
+    symbol: str
+        Elemental symbol, passed to relevant crystalstructure generator when a is float
+    axes: np.array
+        Axes transform to apply to bulk cell
+    crystalstructure: str
+        Base Structure of bulk system, currently supported: fcc, bcc, diamond
+    pbc: list of bool
+        Periodic Boundary Conditions in x, y, z
+    '''
+
+    from ase.lattice.cubic import FaceCenteredCubic, BodyCenteredCubic, Diamond
+    from ase.atoms import Atoms
+    from ase.build import cut, rotate
+    from warnings import warn
+
+    constructors = {
+        "fcc": FaceCenteredCubic,
+        "bcc": BodyCenteredCubic,
+        "diamond": Diamond
+    }
+
+    # Choose correct ase.lattice.cubic constructor given crystalstructure
+    try:
+        cell_builder = constructors[crystalstructure.lower()]
+    except (KeyError, AttributeError) as e:
+        if type(e) == KeyError:
+            # KeyError when crystalstructure not a key in constructors
+            raise ValueError(f"Crystal Structure {crystalstructure} not in accepted list: {list(constructors.keys())}")
+        else:
+            # AttributeError when type(crystalstructure) doesn't support .lower() (not a valid input type)
+            raise TypeError(f"crystalstructure argument {crystalstructure} is not a string")
+
+
+    if np.issubdtype(type(a), np.floating) or np.issubdtype(type(a), np.integer):
+        # Reproduce legacy behaviour with a==alat
+        alat = a
+            
+        unit_cell = cell_builder(symbol, directions=axes.tolist(),
+                                      pbc=pbc,
+                                      latticeconstant=alat)
+    elif type(a) == Atoms:
+        # New behaviour for arbitrary cubic unit cells (Zincblende, L12, ...)
+        alat = a.cell[0, 0]
+
+        ats = a.copy()
+
+        # Try to validate that "a" matches expected structure given by crystalstructure
+
+        tol = 1E-3
+        ref_ats = cell_builder("C", directions=np.eye(3).astype(int).tolist(),
+                                      latticeconstant=alat)
+
+        # Check that number of atoms in ats is a factor of ref_ats
+        # (ats has to be the cubic primitive, or a supercell of the cubic primitive)
+        # Also check that the structure is cubic
+        rel_size = len(ats) / len(ref_ats)
+        skip_fractional_check = False
+        try:
+            diag = np.diag(ats.cell[:, :])
+            assert abs(rel_size - np.floor(rel_size)) < tol and np.all(np.isclose(diag, diag[0], atol=tol))
+        except AssertionError:
+            skip_fractional_check = True
+            
+        # Check fractional coords match expectation
+        frac_match = True
+        if not skip_fractional_check:
+            sup_size = int(rel_size**(1/3))
+            ref_supercell = ref_ats * (sup_size, sup_size, sup_size)
+
+            try:
+
+                apos = np.sort(ats.get_scaled_positions(), axis=0)
+                rpos = np.sort(ref_supercell.get_scaled_positions(), axis=0)
+                assert np.allclose(apos, rpos, atol=tol)
[…]
