+++ tests/test_dislocation.py
@@ -731,6 +731,77 @@ class TestDislocation(matscipytest.MatSciPyTestCase):
         np.testing.assert_almost_equal(E, target_E, decimal=3)
         np.testing.assert_almost_equal(shift, target_shift, decimal=3)
 
+    def test__validate_disloc_inputs(self):
+        from ase.build import bulk as build_bulk
+        import warnings
+
+        warnings.simplefilter("always")
+        
+        bulk_fcc = build_bulk("Cu", crystalstructure="fcc", cubic=True)
+        bulk_bcc = build_bulk("W", crystalstructure="bcc", cubic=True)
+        bulk_dia = build_bulk("C", crystalstructure="diamond", cubic=True)
+
+        bulks = {
+            "fcc": bulk_fcc,
+            "bcc": bulk_bcc,
+            "diamond": bulk_dia
+        }
+
+        axes = np.eye(3).astype(int)
+
+        # Check legacy behaviour works
+        for structure in bulks.keys():
+            bulk = bulks[structure]
+            alat = bulk.cell[0, 0]
+
+            alat_valid, struct = sd._validate_disloc_inputs(alat, symbol="Cu", axes=axes, crystalstructure=structure)
+
+            assert alat_valid == alat
+            bpos = np.sort(bulk.get_scaled_positions(), axis=0)
+            spos = np.sort(struct.get_scaled_positions(), axis=0)
+            assert np.allclose(struct.cell[:, :], bulk.cell[:, :])
+            assert np.allclose(bpos, spos)
+
+        # Check new functionality accepted
+        for structure in bulks.keys():
+            bulk = bulks[structure]
+            alat = bulk.cell[0, 0]
+            alat_valid, struct = sd._validate_disloc_inputs(bulk, symbol="Cu", axes=axes, crystalstructure=structure)
+
+            assert alat_valid == alat
+            bpos = np.sort(bulk.get_scaled_positions(), axis=0)
+            spos = np.sort(struct.get_scaled_positions(), axis=0)
+            assert np.allclose(struct.cell[:, :], bulk.cell[:, :])
+            assert np.allclose(bpos, spos)
+
+        # Check Fail/Warn states
+
+        # "a" validation
+        self.assertRaises(TypeError, sd._validate_disloc_inputs, "Not a bulk or lattice constant", symbol="Cu", axes=axes, crystalstructure="fcc")
+        
+        # Crystalstructure validation
+        self.assertRaises(ValueError, sd._validate_disloc_inputs, bulk_fcc, symbol="Cu", axes=axes, crystalstructure="made up crystalstructure")
+        self.assertRaises(TypeError, sd._validate_disloc_inputs, bulk_fcc, symbol="Cu", axes=axes, crystalstructure=["not the ", "right datatype"])
+
+        # Bulk structure validation
+        self.assertWarns(UserWarning, sd._validate_disloc_inputs, bulk_fcc, symbol="Cu", axes=axes, crystalstructure="bcc")
+        self.assertWarns(UserWarning, sd._validate_disloc_inputs, bulk_fcc, symbol="Cu", axes=axes, crystalstructure="diamond")
+        self.assertWarns(UserWarning, sd._validate_disloc_inputs, bulk_bcc, symbol="Cu", axes=axes, crystalstructure="fcc")
+
+        ats = bulk_fcc.copy()
+        del ats[0]
+        # FCC with vacancy is not bulk FCC!
+        self.assertWarns(UserWarning, sd._validate_disloc_inputs, ats, symbol="Cu", axes=axes, crystalstructure="fcc")
+
+        
+        # Test w/ supercells of bulk
+        print()
[…]
