### 2023-10-18 #5668 @pgrigorev
Dear @mastricker, thank you very much for your review and comments. Here I will try to answer your questions about `dislocation` module and the related tests.

To generate dislocations  the module relies on anisotropic elasticity solution within Stroh formalism as implemented in [atomman](https://www.ctcms.nist.gov/potentials/atomman/) package. Unfortunately this package has an insane number of dependencies and thus we prefer not to include it as a default dependency of `matscipy`. This is one of the reasons so many tests `test_dislocation.py` are skipped (marked as `s`). We have been working on our own implementation of Stroh solution in order to have `matscipy` more independent. This is still in progress and requires extensive testing before the change.

 Another reason why most of test are skipped is related to the use of DXA algorithm for testing. We simply tests if the [DXA algorithm](https://www.ovito.org/manual/reference/pipelines/modifiers/dislocation_analysis.html) from ovito package detects the same dislocation that we want to create. This requires `ovito` python interface and thus is skipped in the automated testing if `ovito` python module not found. Again, we want to avoid including `ovito` as a standard dependency since we only use it for testing.

Finally, some of the tests rely on calling `LAMMPS` via `LAMMPSlib` `ASE` interface since it is one of the most common calculators for MD simulations. The test use an eam potential for tungsten to run a couple of basic checks. This tests are skipped if `lammps` python is not found. I think in your case you have `lammps.py` installed, but it might not have [manybody](https://docs.lammps.org/Packages_details.html#pkg-manybody) package installed. This package is required to use eam potentials and is not a part of standard installation. This is the reason you have `test_elastic_constants_lammpslib` and `test_gamma_line` tests failing.

Normally I use these tests to verify a fresh installation of `matscipy` on a new machine in order to run calculations with `LAMMPS` via `ASE`. In this case I would have `matscipy` + `atomman` + `lammps` installed. I use the tests requiring `ovito` DXA when adding new dislocation configurations during development. It is true that this is not explained at all in the documentation and I will add a `installation` section to [plasticity documentation](http://libatoms.github.io/matscipy/applications/plasticity.html) with the detailed explanation.

In the plasticity section of the documentation I have added few examples how to build dislocations in [Building cylindrical configurations with dislocations](http://libatoms.github.io/matscipy/applications/cylinder_configurations.html) section. At the moment it is only for the case of BCC metals, I will add FCC and Diamond structure this week. I was thinking to add a section to show how one can add a new types of structures and dislocations to the set. If you have any particularly interesting system in mind, I would be happy to try to use it for this example.
