### 2023-10-20 libAtoms/matscipy#189 @rashatwi
JOSS Review: Examples and Tutorials
I see there is an examples folder in the root directory with descriptive names for the python scripts, but it is still not very clear to me what these examples do. I think it will be helpful to have a tutorials section on the website or Jupyter notebooks added to the main GitHub repository that guide the user step-by-step to how they can start using the functionalities described in the paper and also contain detailed explanations.

On a side note, a lot of the examples provided in that directory have not been updated in a long time, e.g., in the example under fracture_mechanics/energy_barrier/params.py, I’m getting this import error:

_ImportError: cannot import name 'diamond_111_110' from 'matscipy.fracture_mechanics.clusters’_

I believe this should be replaced by: _from matscipy.fracture_mechanics.clusters import diamond_.

I also encountered similar errors in other examples. Overall, I think these examples should be updated to match the most recent version of the code and to be more explanatory.

Disclaimer: I'm reviewing the JOSS paper here [openjournals/joss-reviews/#5668](https://github.com/openjournals/joss-reviews/issues/5668)
