### 2023-10-20 #5668 @rashatwi
@diehlpk and @pastewka:

I  finished my review of the code and paper. It is clear to me that matscipy is well suited for JOSS and will be very useful to the computational materials science community.

In terms of the paper, I think the authors did a good job in terms of describing clearly what the functionalities of matscipy are and in terms of citing the relevant and commonly used packages in the field.

In terms of the checklist items, I have checked off most of them, but I think there are a few things that need to be addressed before acceptance of this work in JOSS. Most of my comments are minor, except for the fact that there is very limited/not enough examples or tutorials on how to use the different functionalities explained in the paper. I think the availability of detailed and clear tutorials will make it more accessible and much easier for new users to adopt the methods implemented in matscipy.

# Paper-related comments:
1. Typo on line 155 (ot?)
2. Wondering if there is a way to show some of the equations used in the analysis codes like the radial and angular correlation functions? Multiple variations exist for these calculations (e.g., different normalization methods) and it would be nice for the user to know how these are exactly calculated without having to go over the code (either in the paper or on the package website).
3. In the _All-purpose atomic analysis tools_ section of the paper, only the _topology building for non-reactive MD simulations_ subsection includes the specific module where the functionality is implemented, the other 4 subsections do not refer to the module and I had to go over the package modules to know where these are implemented.

# Package-related comments:
1. I suggest adding the package sphinx documentation link in the “About” section at the top right of the GitHub repository. It makes it much more accessible than just referring to it in the README file.
2. I see there is support for command line interface, but there is no details about how to use it. I’m not sure if this is a preliminary implementation of the cli. If not, then maybe adding simple examples in the README file will be helpful.
3. In the matscipy.elasticity module, lines 1303 and 1304, spilu and LinearOperator are not properly imported from scipy.
4. I created a separate issue for the examples folder provided in the package root directory.
