### 2023-10-20 a8c9482
DOC: remodeled electrochemistry abstract to properly address https://github.com/libAtoms/matscipy/issues/183
