+++ paper/paper.md
@@ -117,7 +117,7 @@ Within materials science, the package has different application domains:
 
 - **Tribology.** Tribology is the study of two interfaces sliding relative to one another, as encountered in frictional sliding or adhesion. Molecular dynamics simulations of representative volume elements of tribological interfaces are routinely used to gain insights into the atomistic mechanisms underlying friction and wear. The module `matscipy.pressurecoupling` provides tools to perform such simulations under a constant normal load and sliding velocity. It includes an implementation of the pressure coupling algorithm described by @Pastewka2010. By dynamically adjusting the distance between the two sliding surfaces according to the local pressure, the algorithm ensures mechanical boundary conditions that account for the inertia of the bulk material which is not explicitly included in the simulation. This algorithm was used to study friction [@Seidl2021-vn] and wear [@Pastewka2011-rd;@Moras2011-my;@Peguiron2016-wf;@Moras2018-lm;@Reichenbach2021-pi].
 
-- **Electrochemistry.** Electrochemistry describes the motion and spatial distribution of charged atoms and molecules (ions) within an external electric field. Classical treatment of charged systems leads to continuous field that describe mean charge distributions, while true atomic systems consist of discrete particles with fixed charges. The `matscipy.electrochemistry` module provides tools that statistically sample discrete coordinate sets from continuum fields and apply steric corrections [@Martinez2009] to avoid overlap of finite size species. To generate continuum charge distributions, the package also contains a control-volume solver [@Selberherr1984] for the one-dimensional Poisson--Nernst--Planck equations [@Bazant2006], as well as an interface to the finite-element solver `FEniCS` [@LoggMardalEtAl2012]. This was used to study friction in electrochemical environments [@Seidl2021;@Hormann2023-ml].
+- **Electrochemistry.** Electrochemistry describes the motion and spatial distribution of charged atoms and molecules (ions) within an external electric field. Classical treatment of charged systems leads to continuous fields that describe mean concentration distributions, while true atomic systems consist of discrete particles with fixed charges. Neither do continuum models account for structured layering of ions in a polar solvent like water [@Seidl2021] nor do they describe finite size effects at high concentrations such as densely packed monolayers [@Hormann2023-ml]. Sampling discrete particle positions from smooth distributions may, however, yield good initial configurations that accelerate equilibration in atomistic calculations. The `matscipy.electrochemistry` module provides tools that statistically sample discrete coordinate sets from continuum fields and apply steric corrections [@Martinez2009] to avoid overlap of finite size species. To generate continuum concentration distributions, the package also contains a control-volume solver [@Selberherr1984] for the one-dimensional Poisson--Nernst--Planck equations [@Bazant2006], as well as an interface to the finite-element solver `FEniCS` [@LoggMardalEtAl2012].
 
 # All-purpose atomic analysis tools
 
