### 2023-10-20 ea5713f
BUG: Fix error in gamma surface si unit conversions
