### 2023-10-20 f0d24ee
ENH: Add test of dislocs from bulk configs
