### 2023-10-20 920e73f
WIP: include fenics interface in published package, add tests on electrochemistry samples including fenics, update fenics interface docs
