+++ examples/electrochemistry/samples_pnp_by_fenics_fem.py
@@ -29,8 +29,11 @@
 
 # %%
 # stretching notebook width across whole window
-from IPython.core.display import display, HTML
-display(HTML("<style>.container { width:100% !important; }</style>"))
+try:
+    from IPython.core.display import display, HTML
+    display(HTML("<style>.container { width:100% !important; }</style>"))
+except:
+    pass
 
 # %%
 # for dynamic module reload during testing, code modifications take immediate effect
@@ -41,20 +44,11 @@ display(HTML("<style>.container { width:100% !important; }</style>"))
 # %matplotlib inline
 
 # %%
-from IPython.core.display import display, HTML
-
-# %%
 # basics
 import logging
 import numpy as np
 import scipy.constants as sc
 import matplotlib.pyplot as plt
-
-# sampling
-# from scipy import interpolate
-# from matscipy.electrochemistry import continuous2discrete
-# from matscipy.electrochemistry import get_histogram
-# from matscipy.electrochemistry.utility import plot_dist
 
 # electrochemistry basics
 from matscipy.electrochemistry import debye, ionic_strength
