+++ examples/electrochemistry/c2d_batch/cell_1d/concentration_sweep/run.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-./run_std.sh   | tee run_std.log
+bash ./run_std.sh | tee run_std.log
