+++ examples/electrochemistry/pnp_batch/cell_1d/concentration_sweep/run.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
-./run_std.sh   | tee run_std.log
-./eval.sh      | tee eval.log
+bash ./run_std.sh   | tee run_std.log
+bash ./eval.sh      | tee eval.log
