### 2023-10-20 25cfecc
TST: simplify githu workflow
