### 2023-10-20 389dcab
TST: simplify test workflow
