### 2023-10-20 2c48ee5
CI: install jupyter
