+++ .github/workflows/tests_fenics.yml
@@ -23,6 +23,7 @@ jobs:
           sudo apt-get install -y python3-dev python3-pip python3-venv libxml2-dev libxslt-dev zlib1g-dev
           # Upgrade to latest meson and ninja
           sudo pip install --upgrade meson ninja
+          sudo pip install --upgrade jupyter jupytext
 
       - name: install_fenics
         run: |
