### 2023-10-20 4d6ad0f
MAINT: create ipynb
