+++ examples/electrochemistry/to_html.sh
@@ -16,7 +16,7 @@ echo "Convert ${infiles}..."
 for file in ${infiles}; do
     BASENAME=$(basename "${file}" .py)
     echo "${BASENAME}.py to ${BASENAME}.ipynb"
-    jupytext --sync "$file"
+    jupytext --to notebook "$file"
     echo "${BASENAME}.ipynb to ${BASENAME}.html"
     python to_html.py "${BASENAME}.ipynb"
 done
