### 2023-10-20 6aaa90e
CI: actually run notebooks
