### 2023-10-20 7d4b308
CI: install pandas
