### 2023-10-20 d0d0430
MAINT: try with fixed tolerance
