### 2023-10-23 862beb7
ENH: tolerance in samples
