### 2023-10-23 0150204
ENH: Convert all dislocs to simpler attribute-based solution
