+++ matscipy/dislocation.py
@@ -2149,10 +2149,23 @@ def check_duplicates(atoms, distance=0.1):
 
 
 class CubicCrystalDislocation:
-    def __init__(self, unit_cell, alat, C11, C12, C44, axes, burgers,
-                 unit_cell_core_position=None,
-                 parity=None, glide_distance=None, n_planes=None,
-                 self_consistent=None):
+
+    # Mandatory Attributes of CubicCrystalDislocation with no defaults
+    # These should be set by the child dislocation class, or by CubicCrystalDislocation.__init__
+    # (see https://stackoverflow.com/questions/472000/usage-of-slots for more details on __slots__)
+    __slots__ = ("__dict__", "burgers_unscaled", "unit_cell_core_position_unscaled",
+                 "glide_distance_unscaled", "crystalstructure", "axes",
+                 "C11", "C12", "C44", "alat", "unit_cell")
+    
+    # Attributes with defaults
+    # These may be overridden by child classes
+    parity = np.zeros(2)
+    n_planes = 3
+    self_consistent = True
+    pbc = [True, True, True]
+    stroh = None
+
+    def __init__(self, a, C11, C12, C44, symbol="W"):
         """
         This class represents a dislocation in a cubic crystal
 
@@ -2162,63 +2175,80 @@ class CubicCrystalDislocation:
 
         Parameters
         ----------
-        unit_cell : unit cell to build the dislocation configuration
-        alat : lattice constant
+        a : lattice constant OR cubic bulk atoms object to build the dislocation configuration
         C11 : elastic constants
         C12
         C44
-        axes : cell axes (b is normally along z direction)
-        burgers : burgers vector of the dislocation
-        unit_cell_core_position : dislocation core position in the unit cell
-                                  used to shift atomic positions to
-                                  make the dislocation core the center
-                                  of the cell
-        parity
+        symbol : Chemical symbol used to construct
+
+
+        Useful Dislocation Properties
+        ------------------------------
+        alat : lattice parameter (in A)
+        unit_cell : bulk cell used to generate dislocations
+        axes : np.array of float
+            unit cell axes (b is normally along z direction)
+        burgers_unscaled : np.array of float
+            burgers vector of the dislocation class, in an abstract alat=1 system
+        burgers : np.array of float
+            burgers vector of the atomistic dislocation
+        unit_cell_core_position : np.array of float
+            dislocation core position in the unit cell used to shift atomic positions to
+            make the dislocation core the center of the cell
+        parity : np.array
         glide_distance : distance to the next equivalent
                          core position in the glide direction
         n_planes : int
             number of non equivalent planes in z direction
         self_consistent : float
             default value for the displacement calculation
+        crystalstructure : str
+            Name of basic structure defining atomic geometry ("fcc", "bcc", "diamond")
         """
 
-        self.unit_cell = unit_cell.copy()
-        self.alat = alat
         self.C11 = C11
         self.C12 = C12
         self.C44 = C44
+        
+        self.alat, self.unit_cell = _validate_disloc_inputs(a, symbol, self.axes, self.crystalstructure, self.pbc)
 
-        self.axes = axes
-        self.burgers = burgers
-        if unit_cell_core_position is None:
-            unit_cell_core_position = np.zeroes(3)
[…]
