### 2023-10-23 ddb2075
MAINT: Formatting & Structure of dislocation.py
