+++ matscipy/dislocation.py
@@ -2153,8 +2153,8 @@ class CubicCrystalDislocation:
     # Mandatory Attributes of CubicCrystalDislocation with no defaults
     # These should be set by the child dislocation class, or by CubicCrystalDislocation.__init__
     # (see https://stackoverflow.com/questions/472000/usage-of-slots for more details on __slots__)
-    __slots__ = ("__dict__", "burgers_unscaled", "unit_cell_core_position_unscaled",
-                 "glide_distance_unscaled", "crystalstructure", "axes",
+    __slots__ = ("__dict__", "burgers_dimensionless", "unit_cell_core_position_dimensionless",
+                 "glide_distance_dimensionless", "crystalstructure", "axes",
                  "C11", "C12", "C44", "alat", "unit_cell")
     
     # Attributes with defaults
@@ -2184,12 +2184,14 @@ class CubicCrystalDislocation:
 
         Useful Dislocation Properties
         ------------------------------
-        alat : lattice parameter (in A)
-        unit_cell : bulk cell used to generate dislocations
+        alat : float
+            lattice parameter (in A)
+        unit_cell : ase.atoms.Atoms object
+            bulk cell used to generate dislocations
         axes : np.array of float
             unit cell axes (b is normally along z direction)
-        burgers_unscaled : np.array of float
-            burgers vector of the dislocation class, in an abstract alat=1 system
+        burgers_dimensionless : np.array of float
+            burgers vector of the dislocation class, in a dimensionless alat=1 system
         burgers : np.array of float
             burgers vector of the atomistic dislocation
         unit_cell_core_position : np.array of float
@@ -2212,7 +2214,7 @@ class CubicCrystalDislocation:
         
         self.alat, self.unit_cell = _validate_disloc_inputs(a, symbol, self.axes, self.crystalstructure, self.pbc)
 
-        self.glide_distance = self.alat * self.glide_distance_unscaled
+        self.glide_distance = self.alat * self.glide_distance_dimensionless
 
     def init_stroh(self):
         from atomman import ElasticConstants
@@ -2220,14 +2222,14 @@ class CubicCrystalDislocation:
         c = ElasticConstants(C11=self.C11, C12=self.C12, C44=self.C44)
         self.stroh = Stroh(c, burgers=self.burgers, axes=self.axes)
 
-    # @property & @var.setter decorators used to ensure var and var_unscaled don't get out of sync
+    # @property & @var.setter decorators used to ensure var and var_dimensionless don't get out of sync
     @property
     def burgers(self):
-        return self.burgers_unscaled  * self.alat
+        return self.burgers_dimensionless  * self.alat
     
     @burgers.setter
     def burgers(self, burgers):
-        self.burgers_unscaled = burgers / self.alat
+        self.burgers_dimensionless = burgers / self.alat
         if self.stroh is None:
             self.init_stroh()
 
@@ -2236,19 +2238,19 @@ class CubicCrystalDislocation:
 
     @property
     def unit_cell_core_position(self):
-        return self.unit_cell_core_position_unscaled * self.alat
+        return self.unit_cell_core_position_dimensionless * self.alat
     
     @unit_cell_core_position.setter
     def unit_cell_core_position(self, position):
[…]
