### 2023-10-25 87c7d53
BUG: Fix disloc classes sharing vector objects
