### 2023-10-25 b7ec4bf
BUG: Remove erroneous float.copy()
