### 2023-10-25 #5668 @pgrigorev
@mastricker thank you very much for your feedback. I am working on an update of the documentation of dislocation module to take into account your comments. If you have any problems with the package in the future please do not hesitate to open an issue and/or contact me directly.
