+++ matscipy/surface_reconstruction.py
@@ -62,7 +62,7 @@ class SurfaceReconstruction:
                 self.el, self.a0, self.calc, lattice=self.lattice)
             self.multilattice = True
 
-    def map_surface(self, fmax=0.0001, layers=6, cutoff=10, shift=0, switch=False):
+    def map_surface(self, fmax=0.0001, layers=6, cutoff=10, shift=0, switch=False,invert_dirs=[False,False,False]):
         """Map the relaxation of the crystal surface a certain number of layers deep.
         Parameters
         ----------
@@ -82,6 +82,11 @@ class SurfaceReconstruction:
         switch : bool
             Whether or not to switch the sublattices around in the mapped structure.
             (Essentially, the correct surface can always be found from some combination of shift and switch)
+        invert_dirs : bool array
+            Whether to invert the components of the mapped shifts. This is useful in the case where a surface which originally has mirror
+            symmetry in a certain plane can break this symmetry during reconstruction. This is the case for the 110 silicon reconstructed surface,
+            where the surface breaks mirror symmetry in the {001} type plane during reconstruction. Two possible surfaces can form, related through
+            diad symmetry. To map the other surface, one has to switch sublattices and invert the mapped displacements along the <001> type direction.
         """
         # build a single cell to determine how many atomic surface
         # layers there are in a single unit cell made by ASE
@@ -102,10 +107,10 @@ class SurfaceReconstruction:
             #check if multilattice atoms are coplanar
             diff = single_cell.get_positions()[self.cb.lattice1mask][0, self.surf_dir] - single_cell.get_positions()[self.cb.lattice2mask][0, self.surf_dir]
             print(diff)
-            if np.abs(diff)<0.001:
-                self.split_cb_pair = False
-            else:
-                self.split_cb_pair = True
+            #if np.abs(diff)<0.001:
+            #    self.split_cb_pair = False
+            #else:
+            #    self.split_cb_pair = True
         else:
             # otherwise, just look at all the layers directly
             n_layer_per_cell = len(np.unique(single_cell.get_positions()[
@@ -165,6 +170,11 @@ class SurfaceReconstruction:
         print('POS SHIFT',pos_shift)
         # print('fafter',slab.get_forces())
 
+        #invert measured shifts if required
+        for direc in range(len(invert_dirs)):
+            if invert_dirs[direc]:
+                pos_shift[:,direc] *= -1
+        
         # for a multi-lattice:
         if self.multilattice:
             # if this is a multi-lattice
@@ -472,7 +482,7 @@ class SurfaceReconstruction:
         # map out the pandey reconstruction on a diamond structure 111 surface
         # can then apply it later (this requires a seperate function as the unit cell of 2x1 reconstruction
         # breaks the symmetry of the surface)
-        self.split_cb_pair = True
+        #self.split_cb_pair = True
         #111 type surface always splits a cb pair
         ny = 1
         nx = 1
@@ -792,8 +802,11 @@ class SurfaceReconstruction:
             print('lattice2')
             print(np.where(layer_mask_set_lattice2[:,i]))
             print(np.where(layer_mask_set_lattice2[:,i]^(mask&layer_mask_set_lattice2[:,i])))
+
+            #TODO uncomment these lines? why did I even comment them? this really needs testing
[…]
