### 2023-10-25 17bf399
ENH: Added feature to allow alpha backtracking
