### 2023-10-26 1c00d45
WIP: Added expressions and test for mode II crack fields
