+++ matscipy/fracture_mechanics/crack.py
@@ -140,9 +140,9 @@ class RectilinearAnisotropicCrack:
         self.mu2_q1 = self.mu2 * self.q1
 
 
-    def displacements(self, r, theta, k):
+    def displacements(self, r, theta, kI, kII=0.0):
         """
-        Displacement field in mode I fracture. Positions are passed in cylinder
+        Displacement field in mode I/II fracture. Positions are passed in cylinder
         coordinates.
 
         Parameters
@@ -162,20 +162,26 @@ class RectilinearAnisotropicCrack:
             Displacements normal to the plane of the crack.
         """
 
-        h1 = k * np.sqrt(2.0*r/math.pi)
+        h1 = kI * np.sqrt(2.0*r/math.pi)
 
         h2 = np.sqrt( np.cos(theta) + self.mu2*np.sin(theta) )
         h3 = np.sqrt( np.cos(theta) + self.mu1*np.sin(theta) )
 
-        u = h1*( self.inv_mu1_mu2*( self.mu1_p2*h2 - self.mu2_p1*h3 ) ).real
-        v = h1*( self.inv_mu1_mu2*( self.mu1_q2*h2 - self.mu2_q1*h3 ) ).real
+        u_I = h1*( self.inv_mu1_mu2*( self.mu1_p2*h2 - self.mu2_p1*h3 ) ).real
+        v_I = h1*( self.inv_mu1_mu2*( self.mu1_q2*h2 - self.mu2_q1*h3 ) ).real
 
+        h4 = kII*np.sqrt(2.0*r/math.pi)
+        u_II = h4*(self.inv_mu1_mu2*(self.p2*h2 - self.p1*h3)).real
+        v_II = h4*(self.inv_mu1_mu2*(self.q2*h2 - self.q1*h3)).real
+
+        u = u_I + u_II
+        v = v_I + v_II
         return u, v
 
 
-    def deformation_gradient(self, r, theta, k):
+    def deformation_gradient(self, r, theta, kI, kII=0.0):
         """
-        Deformation gradient tensor in mode I fracture. Positions are passed in
+        Deformation gradient tensor in mode I/II fracture. Positions are passed in
         cyclinder coordinates.
 
         Parameters
@@ -202,17 +208,31 @@ class RectilinearAnisotropicCrack:
             perpendicular to the plane.
         """
 
-        f = k / np.sqrt(2*math.pi*r)
+        f_I = kI / np.sqrt(2*math.pi*r)
 
         h1 = (self.mu1*self.mu2)*self.inv_mu1_mu2
         h2 = np.sqrt( np.cos(theta) + self.mu2*np.sin(theta) )
         h3 = np.sqrt( np.cos(theta) + self.mu1*np.sin(theta) )
 
-        du_dx = f*( self.inv_mu1_mu2*( self.mu1_p2/h2 - self.mu2_p1/h3 ) ).real
-        du_dy = f*( h1*( self.p2/h2 - self.p1/h3 ) ).real
+        du_dx_I = f_I*( self.inv_mu1_mu2*( self.mu1_p2/h2 - self.mu2_p1/h3 ) ).real
+        du_dy_I = f_I*( h1*( self.p2/h2 - self.p1/h3 ) ).real
 
-        dv_dx = f*( self.inv_mu1_mu2*( self.mu1_q2/h2 - self.mu2_q1/h3 ) ).real
-        dv_dy = f*( h1*( self.q2/h2 - self.q1/h3 ) ).real
[…]
