### 2023-10-26 450e0f3
TST: Added more mode II unit tests
