+++ tests/test_cubic_crystal_crack.py
@@ -201,6 +201,7 @@ class TestCubicCrystalCrack(matscipytest.MatSciPyTestCase):
                 self.assertTrue(np.max(residual[mask]) < 0.2)
 
     def test_consistency_of_deformation_gradient_and_stress(self):
+        #test pure mode 1 first
         for C11, C12, C44, surface_energy, k1 in self.materials:
             crack = CubicCrystalCrack([1,0,0], [0,1,0], C11, C12, C44)
             k = crack.k1g(surface_energy)*k1
@@ -224,8 +225,60 @@ class TestCubicCrystalCrack(matscipytest.MatSciPyTestCase):
                 self.assertAlmostEqual(eps[0, 0], eps_xx)
                 self.assertAlmostEqual(eps[1, 1], eps_yy)
                 self.assertAlmostEqual(eps[0, 1], eps_xy)
+        
+        #now test mode 2
+        for C11, C12, C44, surface_energy, k1 in self.materials:
+            crack = CubicCrystalCrack([1,0,0], [0,1,0], C11, C12, C44)
+            k = crack.k1g(surface_energy)*k1
+            for i in range(10):
+                x = np.random.uniform(-10, 10)
+                y = np.random.uniform(-10, 10)
+                F = crack.deformation_gradient(x, y, 0, 0, 0, kII=k)
+                eps = (F+F.T)/2-np.eye(2)
+                sig_x, sig_y, sig_xy = crack.stresses(x, y, 0, 0, 0,kII=k)
+                eps_xx = crack.crack.a11*sig_x + \
+                         crack.crack.a12*sig_y + \
+                         crack.crack.a16*sig_xy
+                eps_yy = crack.crack.a12*sig_x + \
+                         crack.crack.a22*sig_y + \
+                         crack.crack.a26*sig_xy
+                # Factor 1/2 because elastic constants and matrix product are
+                # expressed in Voigt notation.
+                eps_xy = (crack.crack.a16*sig_x + \
+                          crack.crack.a26*sig_y + \
+                          crack.crack.a66*sig_xy)/2
+                self.assertAlmostEqual(eps[0, 0], eps_xx)
+                self.assertAlmostEqual(eps[1, 1], eps_yy)
+                self.assertAlmostEqual(eps[0, 1], eps_xy)
+        
+        #now test mixed mode
+        for C11, C12, C44, surface_energy, k1 in self.materials:
+            crack = CubicCrystalCrack([1,0,0], [0,1,0], C11, C12, C44)
+            k = crack.k1g(surface_energy)*k1
+            for i in range(10):
+                x = np.random.uniform(-10, 10)
+                y = np.random.uniform(-10, 10)
+                F = crack.deformation_gradient(x, y, 0, 0, k/2, k/2)
+                eps = (F+F.T)/2-np.eye(2)
+                sig_x, sig_y, sig_xy = crack.stresses(x, y, 0, 0, k/2, k/2)
+                eps_xx = crack.crack.a11*sig_x + \
+                         crack.crack.a12*sig_y + \
+                         crack.crack.a16*sig_xy
+                eps_yy = crack.crack.a12*sig_x + \
+                         crack.crack.a22*sig_y + \
+                         crack.crack.a26*sig_xy
+                # Factor 1/2 because elastic constants and matrix product are
+                # expressed in Voigt notation.
+                eps_xy = (crack.crack.a16*sig_x + \
+                          crack.crack.a26*sig_y + \
[…]
