### 2023-10-31 7ae42c4
WIP: Added option to follow a constant G contour NCFlex
