### 2023-11-02 14662d3
ENH: Create utils.py, shift cubic bulk validation there
