### 2023-11-02 7a969e6
BUG: BCCEdge100110Dislocation unit_cell_core_position -> unit_cell_core_position_dimensionless
