### 2023-11-02 b39e596
MAINT: Minor PEP8 changes to dislocation.py
