### 2023-11-02 ba6e97d
ENH: Added a stable sort based on r, theta and z for fracture clusters
