### 2023-11-03 d61ff89
MAINT: Remove random print, minor formatting
