### 2023-11-03 8592e01
DOC: moved visualisation code to an external file
