+++ docs/applications/visualisation.py
@@ -1,0 +1,130 @@
+import numpy as np
+
+from ovito.io.ase import ase_to_ovito
+from ovito.modifiers import CommonNeighborAnalysisModifier, IdentifyDiamondModifier
+from ovito.pipeline import StaticSource, Pipeline
+
+# Get the results of Ovito Common Neighbor Analysis 
+# https://www.ovito.org/docs/current/reference/pipelines/modifiers/common_neighbor_analysis.html
+# and Identify Diamond modifier
+# https://www.ovito.org/docs/current/reference/pipelines/modifiers/identify_diamond.html
+# for better visualisation of the dislocation core
+# it will be identified as "other" structure type
+def get_structure_types(structure, diamond_structure=False):
+    """Get the results of Common Neighbor Analysis and 
+        Identify Diamond modifiers from Ovito
+    Args:
+        structure (ase.atoms): input structure
+    Returns:
+        atom_labels (array of ints): per atom labels of the structure types
+        structure_names (list of strings): names of the structure types
+        colors (list of strings): colors of the structure types in hex format
+    """
+    ovito_structure = structure.copy()
+    if "fix_mask" in ovito_structure.arrays:
+        del ovito_structure.arrays["fix_mask"]
+    
+    if diamond_structure:
+        modifier = IdentifyDiamondModifier()
+    else:
+        modifier = CommonNeighborAnalysisModifier() 
+    
+    data = ase_to_ovito(ovito_structure)
+    pipeline = Pipeline(source=StaticSource(data=data))
+    pipeline.modifiers.append(modifier)
+    data = pipeline.compute()
+
+    atom_labels = data.particles['Structure Type'].array
+
+    structure_names = [structure.name for structure in modifier.structures]
+    colors = [structure.color for structure in modifier.structures]
+    hex_colors = ['#{:02x}{:02x}{:02x}'.format(int(r*255), int(g*255), int(b*255)) for r, g, b in colors] 
+
+    return atom_labels, structure_names, hex_colors
+
+# interactive visualisation  inside the notebook with nglview
+from nglview import show_ase, ASEStructure
+
+def add_dislocation(view, system, name, color=[0, 1, 0], x_shift=0.0):
+    '''Add dislocation line to the view as a cylinder and two cones.
+    The cylinder is hollow by default so the second cylinder is needed to close it.
+    In case partial distance is provided, two dislocation lines are added and are shifter accordingly.
+    '''
+    
+    center = system.positions.mean(axis=0)
+    view.shape.add_cylinder((center[0] + x_shift, center[1], -2.0), 
+                            (center[0] + x_shift, center[1], system.cell[2][2] - 0.5),
+                            color,
+                            [0.3],
+                            name)
+    
+    view.shape.add_cone((center[0] + x_shift, center[1], -2.0), 
+                        (center[0] + x_shift, center[1], 0.5),
+                        color,
+                        [0.3],
+                        name)
+       
+    view.shape.add_cone((center[0] + x_shift, center[1], system.cell[2][2] - 0.5), 
+                        (center[0] + x_shift, center[1], system.cell[2][2] + 1.0),
+                        color,
+                        [0.55],
+                        name)
+    
+
+def show_dislocation(system, scale=0.5, diamond_structure=False, add_bonds=False,
+                     d_name='', d_color=[0, 1, 0], partial_distance=None):
+
+    atom_labels, structure_names, colors = get_structure_types(system, 
+                                                               diamond_structure=diamond_structure)
+
+    view = show_ase(system)
+    view.hide([0])
+    
+    if add_bonds: # add bonds between all atoms to have bonds between structures
+        component = view.add_component(ASEStructure(system), default_representation=False, name='between structures')
[…]
