### 2023-11-03 40508fa
ENH: Made modeII file reading back-compatible with old scripts
