+++ scripts/fracture_mechanics/parallel_rI_convergence_test.py
@@ -1,0 +1,190 @@
+import matscipy; print(matscipy.__file__)
+from matscipy import parameter
+
+from mpi4py import MPI
+from matscipy.fracture_mechanics.crack import CubicCrystalCrack, SinclairCrack
+
+from ase.units import GPa
+import numpy as np
+
+import h5py
+
+comm = MPI.COMM_WORLD
+rank = comm.Get_rank()
+import params
+import os
+
+import time
+
+num_processors = parameter('num_processors')
+#read parameters in from params file
+calc = parameter('calc')
+fmax = parameter('fmax', 1e-3)
+cb = parameter('cb', 'None')
+r_I_vals = parameter('r_I_vals')
+r_III_vals = parameter('r_III_vals')
+cutoff = parameter('cutoff')
+k1_curve_file = parameter('k1_curve_file')
+num_data_points = parameter('num_data_points')
+follow_G_contour = parameter('follow_G_contour',False)
+folder_name = parameter('folder_name','data')
+a0 = parameter('a0') # lattice constant
+k0 = parameter('k0', 1.0)
+alpha0 = parameter('alpha0', 0.0) # initial guess for crack position
+traj_file = parameter('traj_file', 'x_traj.h5')
+extended_far_field = parameter('extended_far_field', False)
+vacuum = parameter('vacuum', 10.0)
+flexible = parameter('flexible', True)
+cont_k = parameter('cont_k','k1')
+clusters = parameter('clusters')#get full set of clusters being used
+
+if cb == 'None':
+    cb = None
+    
+if cb is not None:
+    if rank == 0:
+        #cb.initial_regression_fit()
+        #cb.save_regression_model()
+        cb.load_regression_model()
+        cb_mod = cb.get_model()
+        #communicate cb model to all other processors using mpi broadcast
+    else:
+        cb_mod = 0
+    cb_mod = comm.bcast(cb_mod, root=0)
+    cb.set_model(cb_mod)
+
+crk = CubicCrystalCrack(parameter('crack_surface'),
+                parameter('crack_front'),
+                C=parameter('C')/GPa,cauchy_born=cb)
+
+k1g = crk.k1g(parameter('surface_energy'))
+print('griffthk1,',k1g)
+
+if rank == 0:
+    #now read in the initial data file and start to send over data
+    hf = h5py.File(k1_curve_file, 'r')
+    x_traj = hf['x']
+    x = x_traj[:,:]
+    hf.close()
+
+    total_data_points = np.shape(x)[0]
+    sample_frac = int(total_data_points/num_data_points)
+    x = x[::sample_frac,:]
+
+    data_points_per_proc = int(np.floor(np.shape(x)[0]/num_processors))
+    #k2 distance to walk size
+
+    num_data_point_array = [data_points_per_proc for i in range(num_processors)]
+    sum_diff = np.shape(x)[0] - np.sum(num_data_point_array)
+    for i in range(sum_diff):
+        num_data_point_array[i] += 1
+    print(num_data_point_array)
+    time.sleep(3)
+    #now communicate over all data
+    for proc_num in range(1,num_processors):
+        print(proc_num)
+        comm.send(num_data_point_array[proc_num], dest=proc_num, tag=12)
+        comm.send(x[sum(num_data_point_array[:proc_num]):sum(num_data_point_array[:proc_num+1]),:], dest=proc_num, tag=10)
+        comm.send(os.getpid(),dest=proc_num,tag=13)
+    #the first set of data (proc_num 0) is assigned to rank 0 (this processor)
+    num_data_points = num_data_point_array[0]
+    x = x[0:num_data_points,:]
+    pid = os.getpid()
+else:
+    num_data_points = comm.recv(source=0, tag=12)
+    x = comm.recv(source=0, tag=10)
+
[…]
