### 2023-11-03 980ccf2
TST: Weakened cauchy-born corrector test tol a little
