### 2023-11-03 5e630f7
DOC: added ovito to pyproject.toml
