+++ pyproject.toml
@@ -34,7 +34,8 @@ docs = [
     "myst_nb",
     "nglview",
     "numpydoc",
-    "atomman"
+    "atomman",
+    "ovito"
 ]
 
 [project.urls]
