+++ docs/conf.py
@@ -44,7 +44,8 @@ extensions = [
     'sphinx.ext.autosummary',
     'sphinx.ext.mathjax',
     'myst_nb',
-    'numpydoc'
+    'numpydoc',
+    'sphinx_copybutton',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
