### 2023-11-03 46fe640
DOC: updated copyright
