### 2023-11-03 #5668 @pgrigorev
> @pgrigorev could you finish the requested changes?

I have just added the requested changes to [a pull request](https://github.com/libAtoms/matscipy/pull/188). Beginning of next week we will make sure everything works and it should be merged to update the documentation.
