### 2023-11-05 4287a0d
DOC: Add "interactive_view" from cylinder_configurations.ipynb to matscipy.utils
