### 2023-11-05 bc531fe
DOC: Shift interactive_view to visualisation.py
