### 2023-11-06 36a4689
TST: Added 3D build cell unit tests
