+++ matscipy/cauchy_born.py
@@ -10,6 +10,8 @@ from itertools import permutations
 from scipy.stats.qmc import LatinHypercube, scale
 
 import ase
+
+
 class RegressionModel:
     """Simple regression model object wrapper for np.linalg.lstsq for
     predicting shifts."""
@@ -27,7 +29,7 @@ class RegressionModel:
         values : array_like
             value of function to be fitted evaluated at each data point
         """
-        self.model = np.linalg.lstsq(phi, values,rcond=None)[0]
+        self.model = np.linalg.lstsq(phi, values, rcond=None)[0]
 
     def predict(self, phi):
         """Evaluate the simple least-squares regression model.
@@ -46,8 +48,8 @@ class RegressionModel:
         """
 
         return phi @ self.model
-    
-    def predict_gradient(self,grad_phi):
+
+    def predict_gradient(self, grad_phi):
         """Evaluate the simple least-squares regression model using
         the derivative of the basis functions to get the gradient.
 
@@ -64,11 +66,11 @@ class RegressionModel:
             1D vector of least squares model gradient predictions for simple model.
         """
         return grad_phi @ self.model
-    
+
     def save(self):
         """Saves the regression model to file: CB_model.npy"""
-        np.save('CB_model.npy',self.model)
-        
+        np.save('CB_model.npy', self.model)
+
     def load(self):
         """Loads the regression model from file: CB_model.npy"""
         self.model = np.load('CB_model.npy')
@@ -104,7 +106,7 @@ class CubicCauchyBorn:
         self.lattice1mask = None  # mask for the lattice 1 atoms
         self.lattice2mask = None  # mask for the lattice 2 atoms
 
-    def set_sublattices(self, atoms, A,read_from_atoms=False):
+    def set_sublattices(self, atoms, A, read_from_atoms=False):
         """Apply a small strain to all atoms in the supplied atoms structure
         and determine which atoms belong to which sublattice using forces. NOTE
         as this method is based on forces, it does not work in cases where atom
@@ -181,11 +183,13 @@ class CubicCauchyBorn:
                 if all(element == lattice1mask[0] for element in lattice1mask):
                     lattice1mask = force_diff_lattice[:, 2] > 0
                     lattice2mask = force_diff_lattice[:, 2] < 0
-            
-            #set new array in atoms
-            try: #make new arrays if they didn't already exist
-                atoms.new_array('lattice1mask',np.zeros(len(atoms),dtype=bool))
-                atoms.new_array('lattice2mask',np.zeros(len(atoms),dtype=bool))
+
+            # set new array in atoms
+            try:  # make new arrays if they didn't already exist
+                atoms.new_array('lattice1mask', np.zeros(
+                    len(atoms), dtype=bool))
+                atoms.new_array('lattice2mask', np.zeros(
+                    len(atoms), dtype=bool))
             except RuntimeError:
                 pass
             lattice1maskatoms = atoms.arrays['lattice1mask']
@@ -197,7 +201,7 @@ class CubicCauchyBorn:
         self.lattice1mask = lattice1mask
         self.lattice2mask = lattice2mask
 
-    def switch_sublattices(self,atoms):
+    def switch_sublattices(self, atoms):
         """Switch the sublattice masks on a set of atoms around
         such that lattice1mask = lattice2mask and vice versa
 
@@ -276,12 +280,14 @@ class CubicCauchyBorn:
             cell = unitcell_copy.get_cell()
             cell_rescale = np.transpose(x[:, :] @ cell)
[…]

[…]
