### 2023-11-06 f01d3ec
DOC: dislocation.py formatting for numpydoc
