### 2023-11-06 74649ba
tidy up AnisotropicDislocation class
