+++ matscipy/dislocation.py
@@ -2149,55 +2149,40 @@ def check_duplicates(atoms, distance=0.1):
     # print(f"found {duplicates.sum()} duplicates")
     return duplicates.astype(np.bool)
 
-class CubicCrystalDislocationGrad:
+class AnisotropicDislocation:
     """
-    Displacement and displacement gradient field of straight 
-    dislocation in anisotropic elastic media.
-    Ref: pp. 467 in J.P. Hirth and J. Lothe, Theory of Dislocations, 2nd ed.
-    The function is similar as class `CubicCrystalDislocation`.
+    Displacement and displacement gradient field of straight dislocation 
+    in anisotropic elastic media. Ref: pp. 467 in J.P. Hirth and J. Lothe, 
+    Theory of Dislocations, 2nd ed. Similar to class `CubicCrystalDislocation`.
     """
 
-    def __init__(self, a0, crack_surface, crack_front, n, xi, Burgers, C11=None, C12=None,
-                 C44=None):
+    def __init__(self, C11, C12, C44, axes,
+                 slip_plane, disloc_line, burgers):
         """
-        Insert a dislocation in a cubic crystal containing a crack. 
-        Elastic constants are C11, C12 and C44 (or optionally a full 6x6 elastic constant matrix C).
-        The dislocation is defined by m (dislocation glide direction), n (the slip plane normal), 
-        and xi (dislocation line direction).
+        Setup a dislocation in a cubic crystal. 
+        C11, C12 and C44 are the elastic constants in the Catersian geometry.
+        The arg 'axes' is a 3x3 array containing axes of frame of reference (eg.crack system).
+        The dislocation parameters required are the slip plane normal ('slip_plane'), the 
+        dislocation line direction ('disloc_line') and the Burgers vector ('burgers').
         """
 
-        # x (third_dir) - direction in which the crack is running
-        # y (crack_surface) - free surface that forms due to the crack
-        # z (crack_front) - direction of the crack front
-        # normalize axis of crack coordinates 
-        third_dir = np.cross(crack_surface, crack_front)
-        third_dir = np.array(third_dir) / np.sqrt(np.dot(third_dir,
-                                                         third_dir))
-        crack_surface = np.array(crack_surface) / \
-            np.sqrt(np.dot(crack_surface, crack_surface))
-        crack_front = np.array(crack_front) / \
-            np.sqrt(np.dot(crack_front, crack_front))
-        
-        # define rotation matrix from dislocation system to crack system
-        A = np.array([third_dir, crack_surface, crack_front])
-        if np.linalg.det(A) < 0:
-            third_dir = -third_dir
-        A = np.array([third_dir, crack_surface, crack_front])
-        m = np.cross(n, xi)
+        # normalize axes of the cell (for NCFlex: crack system coordinates) 
+        A = np.array([ np.array(v)/np.sqrt(np.dot(v,v)) for v in axes ])
 
-        # normalize m, n and xi
+        # normalize axis of dislocation coordinates
+        n = slip_plane ; n = np.array(n) / np.sqrt(np.dot(n, n))
+        xi = disloc_line ; xi = np.array(xi) / np.sqrt(np.dot(xi, xi))
+        m = np.cross(n, xi) # dislocation glide direction
         m = np.array(m) / np.sqrt(np.dot(m, m))
-        n = np.array(n) / np.sqrt(np.dot(n, n))
-        xi = np.array(xi) / np.sqrt(np.dot(xi, xi))
-        disloc_sys = np.array([m, n, xi])
-        if np.linalg.det(disloc_sys) < 0:
-            m = -m        
+        #disloc_sys = np.array([m, n, xi])
+        #if np.linalg.det(disloc_sys) < 0:
[…]
