### 2023-11-07 #5668 @pgrigorev
> > @pgrigorev could you finish the requested changes?
>
> I have just added the requested changes to [a pull request](https://github.com/libAtoms/matscipy/pull/188). Beginning of next week we will make sure everything works and it should be merged to update the documentation.

Hello @diehlpk and @mastricker. We have successfully updated the documentation by adding more details about tests and installation of `matscipy.dislocation` module as well as tutorials on how to create dislocations in BCC, FCC and Diamond structures. You can have a look [here](https://libatoms.github.io/matscipy/applications/plasticity.html).
