### 2023-11-07 bd41566
DOC: Merge pgrigorev changes to visualisation.py
