### 2023-11-08 481bdf8
MAINT: Remove dict from slots in CubicCrystalDislocation
