### 2023-11-08 dd5b5b8
ENH: GammaSurface called using dislocation class
