### 2023-11-08 22f3ae8
BUG: Pass crystalstructure and symbol through to bulk generator
