### 2023-11-08 834f1e3
DOC: First parts of stacking fault docs
