### 2023-11-09 4a16774
ENH: Set glide distance automatically based on dislocation glide distance
