### 2023-11-09 357fb71
MAINT: Add verbosity arg to get_elastic_constants
