### 2023-11-09 937ad04
AnisDisloc: fixed self.m bug, and included 3D
