### 2023-11-10 dc2b678
DOC: Adjustments to show_stacking_fault function
