+++ docs/applications/visualisation.py
@@ -1,10 +1,11 @@
 import numpy as np
 # interactive visualisation  inside the notebook with nglview
-from nglview import show_ase, ASEStructure
+from nglview import show_ase, show_asetraj, ASEStructure
 from ovito.io.ase import ase_to_ovito
 from ovito.modifiers import CommonNeighborAnalysisModifier, IdentifyDiamondModifier
 from ovito.pipeline import StaticSource, Pipeline
-from ase.visualize import view
+from ase.visualize.ngl import NGLDisplay
+
 
 # Get the results of Ovito Common Neighbor Analysis 
 # https://www.ovito.org/docs/current/reference/pipelines/modifiers/common_neighbor_analysis.html
@@ -164,12 +165,37 @@ def interactive_view(system, scale=0.5, name=""):
     view._js(tooltip_js)
     return view
 
-def show_stacking_fault(images):
-    # TODO: Nice NGL representation of these!
-    images = [image.copy() for image in images]
-    cell = images[0].cell[:, :]
-    for image in images:
-        image.set_cell(cell)
-        image.wrap()
-    v = view(images, viewer="ngl")
-    return v
+def show_stacking_fault(images, scale=0.5, CNA_color=True, diamond_structure=False):
+    center = images[0].cell[:, :]
+    # atom_labels, structure_names, colors = get_structure_types(system, 
+    #                                                            diamond_structure=diamond_structure)
+    images = [image.copy() * (2, 2, 2) for image in images]
+
+    display = NGLDisplay(images, 700, 300)
+    view = display.view
+    view.add_unitcell()
+    view.add_spacefill(radiusType='covalent', radiusScale=scale)
+    
+    #view.hide([0])
+    
+    # for structure_type in np.unique(atom_labels):
+    #     # every structure type is a different component
+    #     mask = atom_labels == structure_type
+    #     component = view.add_component(ASEStructure(system[mask]), 
+    #                                    default_representation=False, name=str(structure_names[structure_type]))
+    #     if CNA_color:
+    #         component.add_spacefill(color=colors[structure_type], radiusType='covalent', radiusScale=scale)
+    #     else:
+    #         component.add_spacefill(radiusType='covalent', radiusScale=scale)
+    
+    # Quaternion rotation
+    angle = 4 * np.pi/6
+    axis = np.array([1, 1, -1])
+    q1 = [np.cos(angle/2), *(np.sin(angle/2)*axis / np.linalg.norm(axis))]
+    view.control.rotate(q1)
+
+    view.camera = 'orthographic'
+    view.parameters = {"clipDist": 0}
+    view.center(center)
+
+    return view
