### 2023-11-10 7438931
AnisDisloc - resolved gradient sign issue
