### 2023-11-10 b92afaa
Update paper.md
Added Lei Zhang to author list (contributed the AnisotropicDislocation class in dislocation.py).
