### 2023-11-11 bc8fac3
DOC: switched to ase plots for SF visualisation
