+++ paper/paper.md
@@ -58,7 +58,7 @@ authors:
     orcid: 0000-0002-8008-1482
     affiliation: 7
   - name : Lei Zhang
-    orcid:
+    orcid: 0000-0003-4414-7111
     affiliation: 9
   - name: James R. Kermode
     orcid: 0000-0001-6755-6271
