### 2023-11-14 2c4ba8c
ENH: Stacking fault attached calculator, more docs
