### 2023-11-14 2959161
AD - integrated with CCD, shared self_consistency
