### 2023-11-15 9b4b00e
MAINT: Fixed bug with incorrect Voigt notation and cleaned code
