+++ matscipy/cauchy_born.py
@@ -8,7 +8,7 @@ from scipy.linalg import polar, sqrtm
 from itertools import permutations
 
 from scipy.stats.qmc import LatinHypercube, scale
-
+from matscipy.elasticity import full_3x3_to_Voigt_6_strain, Voigt_6_to_full_3x3_strain
 import ase
 
 
@@ -68,12 +68,12 @@ class RegressionModel:
         return grad_phi @ self.model
 
     def save(self):
-        """Saves the regression model to file: CB_model.npy"""
-        np.save('CB_model.npy', self.model)
+        """Saves the regression model to file: CB_model.txt"""
+        np.savetxt('CB_model.txt', self.model)
 
     def load(self):
-        """Loads the regression model from file: CB_model.npy"""
-        self.model = np.load('CB_model.npy')
+        """Loads the regression model from file: CB_model.txt"""
+        self.model = np.loadtxt('CB_model.txt')
 
 
 class CubicCauchyBorn:
@@ -140,6 +140,8 @@ class CubicCauchyBorn:
 
         # generate U (right hand stretch tensor) to apply
         if read_from_atoms:
+            if ('lattice1mask' not in atoms.arrays) or ('lattice2mask' not in atoms.arrays):
+                raise KeyError('Lattice masks not found in atoms object')
             lattice1mask = atoms.arrays['lattice1mask']
             lattice2mask = atoms.arrays['lattice2mask']
         else:
@@ -183,6 +185,8 @@ class CubicCauchyBorn:
                 if all(element == lattice1mask[0] for element in lattice1mask):
                     lattice1mask = force_diff_lattice[:, 2] > 0
                     lattice2mask = force_diff_lattice[:, 2] < 0
+                    if all(element == lattice1mask[0] for element in lattice1mask):
+                        raise RuntimeError('No forces detected in any direction when shift applied - cannot assign sublattices')             
 
             # set new array in atoms
             try:  # make new arrays if they didn't already exist
@@ -209,10 +213,11 @@ class CubicCauchyBorn:
         ----------
         atoms : ASE atoms object with defined sublattices
         """
-        tmpl1 = self.lattice1mask.copy()
-        tmpl2 = self.lattice2mask.copy()
-        self.lattice1mask = tmpl2.copy()
-        self.lattice2mask = tmpl1.copy()
+        self.lattice1mask, self.lattice2mask = self.lattice2mask.copy(), self.lattice1mask.copy()
+        #tmpl1 = self.lattice1mask.copy()
+        #tmpl2 = self.lattice2mask.copy()
+        #self.lattice1mask = tmpl2.copy()
+        #self.lattice2mask = tmpl1.copy()
         lattice1maskatoms = atoms.arrays['lattice1mask']
         lattice2maskatoms = atoms.arrays['lattice2mask']
         lattice1maskatoms[:] = self.lattice1mask
@@ -256,13 +261,7 @@ class CubicCauchyBorn:
             """
             # green lagrange version of function
             # turn E into matrix
-            E = np.zeros([3, 3])
-            E[0, 0] = E_vec[0]
-            E[1, 1] = E_vec[1]
-            E[2, 2] = E_vec[2]
-            E[2, 1], E[1, 2] = E_vec[3], E_vec[3]
-            E[2, 0], E[0, 2] = E_vec[4], E_vec[4]
-            E[0, 1], E[1, 0] = E_vec[5], E_vec[5]
[…]
