+++ tests/test_cauchy_born_corrector.py
@@ -10,6 +10,7 @@ from matscipy.calculators.manybody import Manybody
 from scipy.linalg import sqrtm
 import ase.io
 import matscipytest
+from matscipy.elasticity import Voigt_6_to_full_3x3_strain
 
 
 class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
@@ -44,64 +45,49 @@ class TestPredictCauchyBornShifts(matscipytest.MatSciPyTestCase):
 
     def test_fit_taylor_model(self):
         self.cb.fit_taylor()
-        # print(self.cb.grad_f)
-        # print(self.cb.hess_f)
-        grad_f = np.array([0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -
-                          1.09995082e+00, -2.22044605e-12, -3.33066907e-12])
-        hess_f = np.array([[0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.03958794e-01, 0.00000000e+00, 1.11022302e-08],
+        print(self.cb.grad_f)
+        print(self.cb.hess_f)
+        grad_f = np.array([0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
+                          -5.49975395e-01, 0.00000000e+00, -2.22044605e-12])
+        hess_f = np.array([[0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 
+                               1.01979464e-01, 1.66533454e-08, 1.66533454e-08],
                            [0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
-                               9.12311482e-01, -5.55111512e-09, 0.00000000e+00],
+                               4.56155597e-01, -1.66533454e-08, 0.00000000e+00],
                            [0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
-                               9.12311465e-01, 0.00000000e+00, 0.00000000e+00],
-                           [2.03958794e-01, 9.12311493e-01, 9.12311465e-01,
-                            0.00000000e+00, 0.00000000e+00, 0.00000000e+00],
-                           [0.00000000e+00, -5.55111512e-09, 0.00000000e+00,
-                            0.00000000e+00, 0.00000000e+00, 0.00000000e+00],
-                           [1.11022302e-08, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00]])
+                               4.56155608e-01, 0.00000000e+00, 0.00000000e+00],
+                           [1.01979464e-01, 4.56155597e-01, 4.56155608e-01,
+                               0.00000000e+00, 0.00000000e+00, 0.00000000e+00],
+                           [1.66533454e-08, -1.66533454e-08, 0.00000000e+00,
+                               0.00000000e+00, 0.00000000e+00, 0.00000000e+00],
+                           [1.66533454e-08, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00]])
         assert np.allclose(self.cb.grad_f, grad_f, atol=1e-6)
         assert np.allclose(self.cb.hess_f, hess_f, atol=1e-6)
 
     def E_cart3D(self, x, y, z, eps=None):
[…]
