### 2023-11-15 9213fc3
MAINT: Code restructuring and error catching
