+++ matscipy/surface_reconstruction.py
@@ -20,7 +20,8 @@ class SurfaceReconstruction:
     can be mapped to the internal crack surfaces of an atoms object.
     """
 
-    def __init__(self, el, a0, calc, directions, surf_dir, lattice='diamond', multilattice=False):
+    def __init__(self, el, a0, calc, directions,
+                 surf_dir, lattice, multilattice=False):
         """Parameters
         ----------
         el : string
@@ -46,7 +47,7 @@ class SurfaceReconstruction:
         self.calc = calc
         self.directions = directions
         self.surf_dir = surf_dir
-        self.inter_planar_dist = a0/np.linalg.norm(directions[surf_dir])
+        self.inter_planar_dist = a0 / np.linalg.norm(directions[surf_dir])
         # build the rotation matrix from the directions
         A = np.zeros([3, 3])
         for i, direction in enumerate(directions):
@@ -62,7 +63,8 @@ class SurfaceReconstruction:
                 self.el, self.a0, self.calc, lattice=self.lattice)
             self.multilattice = True
 
-    def map_surface(self, fmax=0.0001, layers=6, cutoff=10, shift=0, switch=False, invert_dirs=[False, False, False]):
+    def map_surface(self, fmax=0.0001, layers=6, cutoff=10,
+                    shift=0, switch=False, invert_dirs=[False, False, False]):
         """Map the relaxation of the crystal surface a certain number of layers deep.
         Parameters
         ----------
@@ -72,7 +74,7 @@ class SurfaceReconstruction:
             Number of layers deep to map the free surface
         cutoff : float
             Cutoff of the potential being used - used to decide the number of
-            atoms beneath the mapped layers that need to be modelled 
+            atoms beneath the mapped layers that need to be modelled
         shift : float
             Amount to shift the bulk structure before the surface is mapped
             such that the top layer of surface is physically meaningful. One
@@ -96,8 +98,8 @@ class SurfaceReconstruction:
             symbol=self.el,
             latticeconstant=self.a0,
             pbc=(1, 1, 1))
-        ase.io.write(
-            f'{self.directions[self.surf_dir]}_unitcell.xyz', single_cell)
+        # ase.io.write(
+        #    f'{self.directions[self.surf_dir]}_unitcell.xyz', single_cell)
 
         if self.multilattice:
             # for a multilattice, get number of layers by just looking
@@ -109,7 +111,7 @@ class SurfaceReconstruction:
             diff = single_cell.get_positions()[self.cb.lattice1mask][0, self.surf_dir] - \
                 single_cell.get_positions(
             )[self.cb.lattice2mask][0, self.surf_dir]
-            print(diff)
+            # print(diff)
             # if np.abs(diff)<0.001:
             #    self.split_cb_pair = False
             # else:
@@ -125,8 +127,8 @@ class SurfaceReconstruction:
         # set the number of atomic unit cells to map
         # as twice the number of layers to map + a number of layers
         # equal to the potential cutoff for fixing
-        height = 2*int(layers/n_layer_per_cell) + \
-            int(np.round(cutoff/self.inter_planar_dist))
+        height = 2 * int(layers / n_layer_per_cell) + \
+            int(np.round(cutoff / self.inter_planar_dist))
         size = [1, 1, 1]
         size[self.surf_dir] *= height
         self.layers = layers
@@ -160,17 +162,17 @@ class SurfaceReconstruction:
[…]
