### 2023-11-15 67fbfbb
ENH: Package generate_images to new private function
