### 2023-11-15 8fccf86
ENH: Togglable compression mode for gamma surface images
