### 2023-11-15 55dd566
MAINT: Initialise 'Es'
