### 2023-11-15 1fe44f1
ENH: Allow Es to be directly passed to plotting
