### 2023-11-15 0d61636
DOC: Add Es docs to stacking fault plot numpydoc
