+++ tests/test_gamma_surface.py
@@ -31,14 +31,16 @@ class GammaSurfaceTest(matscipytest.MatSciPyTestCase):
         Es = surface.get_surface_energies(self.model)
 
         assert np.min(Es) >= 0.0
-        assert np.allclose(np.max(Es), 0.1946074084327025)
+        print("Gamma Surface: ", np.max(Es))
+        assert np.allclose(np.max(Es), 0.20510077433464793)
 
     def test_stacking_fault(self):
         surface = StackingFault(self.at0, np.array([1, 1, 0]), np.array([-1, 1, 0]))
         surface.generate_images(9, z_replications=1, path_ylims=[0, 0.5])
         surface.relax_images(self.model, self.fmax)
         Es = surface.get_surface_energies(self.model)[0, :]
-        assert np.allclose(np.max(Es), 0.07233706030475648)
+        print("Stacking Fault: ", np.max(Es))
+        assert np.allclose(np.max(Es), 0.07150004280093887)
 
     def test_disloc_stacking_fault(self):
         from matscipy.dislocation import DiamondGlideScrew
@@ -46,6 +48,5 @@ class GammaSurfaceTest(matscipytest.MatSciPyTestCase):
         surface.generate_images(9, z_replications=1, path_ylims=[0, 0.5])
         surface.relax_images(self.model, self.fmax)
         Es = surface.get_surface_energies(self.model)[0, :]
-        print(np.max(Es))
 if __name__ == '__main__':
     unittest.main()
