### 2023-11-16 7fa8fcf
ENH: atom offset for dislocs relative to core position
