### 2023-11-17 33efaba
ENH: complete_basis function in utils.py
