### 2023-11-17 1daac47
MAINT: get_surface_energies -> get_energy_densities
