### 2023-11-17 3ebc245
TST: Remove quippy import
