### 2023-11-17 b59324e
MAINT: Switch FIRE optimiser to PreconLBFGS
