### 2023-11-17 2859eec
ENH: Port stacking fault plotting to class method
