### 2023-11-17 6cf424e
BUG: Fixes to uncompressed structure generation
