### 2023-11-17 432c37c
ENH: Add manual rotation to stacking fault show
