### 2023-11-17 0eb87d0
ENH: Gamma Surface show() function
