### 2023-11-17 b30d31a
ENH: _vec_to_miller function
