### 2023-11-17 68b29e3
ENH: Add miller directions to .show() methods
