### 2023-11-18 a163a32
MAINT: Rename plot_gamma_surface -> plot_energy_densities
