### 2023-11-18 63bcc40
MAINT: Fix basis sometimes having wrong handedness
