### 2023-11-18 455a34b
ENH: Expose interpolation scheme for gamma surface plot
