### 2023-11-18 f6b6040
MAINT: Remove show_stacking_fault from visualisation.py, as not needed
