+++ docs/applications/plasticity.rst
@@ -23,3 +23,4 @@ Tutorials:
     cylinder_configurations.ipynb
     multispecies_dislocations.ipynb
     gamma_surfaces.ipynb
+    gamma_surface_advanced.ipynb
