### 2023-11-21 645f27f
ENH: DIslocation Quadrupole class basics
