### 2023-11-21 6ea2bac
ENH: ray casting algorithm for points inside polygon
