+++ matscipy/utils.py
@@ -105,3 +105,98 @@ def validate_cubic_cell(a, symbol="w", axes=None, crystalstructure=None, pbc=Tru
         raise TypeError("type(a) is not a float, or an Atoms object")
 
     return alat, unit_cell
+
+
+def line_intersect_2D(p1, p2, x1, x2):
+    '''
+    Test if 2D finite line defined by points p1 & p2 intersects with the finite line defined by x1 & x2.
+    Essentially a Python conversion of the ray casting algorithm suggested in: 
+    https://stackoverflow.com/questions/217578/how-can-i-determine-whether-a-2d-point-is-within-a-polygon
+
+    Returns True if lines intersect (including if they are collinear)
+
+    p1, p2, x1, x2: np.array
+        2D points defining lines
+    '''
+
+    # Convert from pointwise p1, p2 form to ax + by + c = 0 form
+    a_p = p2[0] - p1[0]
+    b_p = p1[1] - p2[1]
+    c_p = (p2[0] * p1[1]) - (p1[0] * p2[1])
+
+    # Sub in x1, x2
+    d1 = a_p * x1[0] + b_p * x1[1] + c_p
+    d2 = a_p * x2[0] + b_p * x2[1] + c_p
+
+    # Both points above the line
+    if (d1 > 0 and d2 > 0) return False
+    # Both points below the line
+    if (d1 < 0 and d2 < 0) return False
+
+    # One point above, one point below the *infinite* line defined by p1, p2
+    # Instead of testing for finite line, switch p <-> x and repeat
+    # (Both finite lines intersect if both sets of points intersect the infinite
+    # lines defined by the other set of points)
+
+    # Convert from pointwise x1, x2 form to ax + by + c = 0 form
+    a_x = x2[0] - x1[0]
+    b_x = x1[1] - x2[1]
+    c_x = (x2[0] * x1[1]) - (x1[0] * x2[1])
+
+    # Sub in p1, p2
+    d1 = a_x * p1[0] + b_x * p1[1] + c_x
+    d2 = a_x * p2[0] + b_x * p2[1] + c_x
+
+    # Both points above the line
+    if (d1 > 0 and d2 > 0) return False
+    # Both points below the line
+    if (d1 < 0 and d2 < 0) return False
+
+    # 2 possibilities left: the lines intersect at a single point, or the lines are collinear,
+    # both result in an intersection, so return True
+    return True
+
+def points_in_polygon2D(p, poly_points):
+    '''
+    Test if points lies within the closed 2D polygon defined by poly_points
+    Uses ray casting algorithm, as suggested by:
+    https://stackoverflow.com/questions/217578/how-can-i-determine-whether-a-2d-point-is-within-a-polygon
+
+    p: np.array
+        (n, 2) array of points to test
+    poly_points: np.array
+        Points defining the 2D polygon
+
+    Returns
+    -------
+    mask: np.array
+        Boolean mask. True for points inside the poylgon
+    '''
+    if len(p.shape) == 1:
+        # Single point provided
+        points = p[np.newaxis, :]
+    else:
+        points = p
+
+    # Ensure polygon is closed
+    if poly_points[0, :] != poly_points[-1, :]:
+        poly_points = np.append(poly_points, poly_points[0, :], axis=0)
+
+
+
+    npoints = points.shape[0]
+
+    mask = np.zeros(npoints, dtype=bool)
+
+    # Get point that is definitely outside the polygon
+    test_point = np.array([2 * np.max(poly_points[:, 0]), 2 * np.max(poly_points[:, 1])])
+
+    for i in range(npoints):
+        intersections = 0
+        for j in range(poly_points.shape[0] - 1):
[…]
