### 2023-11-22 9d007cc
ENH: Dislocation Quadrupole Generation
