### 2023-11-22 9e2541f
MAINT: Variable renaming, formatting
