### 2023-11-22 da5cade
ENH: Add failure if relaxation does not converge, add steps limit
