### 2023-11-22 c0c47e5
BUG: Carry renamed variables through to unittests
