### 2023-11-22 72101e1
MAINT: Replace if statement with one-liner
