### 2023-11-23 03844bd
ENH: added `save_count` to suppress warning
