+++ matscipy/gamma_surface.py
@@ -6,29 +6,35 @@ import warnings
 from ase.constraints import UnitCellFilter
 from matscipy.utils import validate_cubic_cell, complete_basis
 import inspect
-from matscipy.dislocation import CubicCrystalDislocation, CubicCrystalDissociatedDislocation
+from matscipy.dislocation import CubicCrystalDislocation, \
+                                 CubicCrystalDissociatedDislocation
 from ase.units import _e
+
 
 class GammaSurface():
     '''
     A class for generating gamma surface/generalised stacking fault images & plots
     '''
-    def __init__(self, a, surface_direction, glide_direction=None, crystalstructure=None, symbol="C"):
+    def __init__(self, a, surface_direction, glide_direction=None,
+                 crystalstructure=None, symbol="C"):
         '''
         Initialise by cutting and rotating the input structure.
 
         Parameters
         ----------
         a: float or ase.Atoms
-            Lattice Constant or Starting structure to generate gamma surface from
-            (Operates similarly to CubicCrystalDislocation)
+            Lattice Constant or Starting structure to generate gamma surface
+            from (Operates similarly to CubicCrystalDislocation)
             If lattice constant is provided, crystalstructure must also be set
-        surface_direction: np.array of int or subclass of matscipy.dislocation.CubicCrystalDissociatedDislocation
+        surface_direction: np.array of int or subclass of
+            matscipy.dislocation.CubicCrystalDissociatedDislocation
             Vector direction of gamma surface, in miller index notation
             EG: np.array([0, 0, 1]), np.array([-1, 1, 0]), np.array([1, 1, 1])
-            A subclass of matscipy.dislocation.CubicCrystalDissociatedDislocation (EG: DiamondGlideScrew or FCCEdge110Dislocation)
+            A subclass of matscipy.dislocation.CubicCrystalDissociatedDislocation 
+            (EG: DiamondGlideScrew or FCCEdge110Dislocation)
         glide_direction: np.array of int or None
-            Basis vector (in miller indices) to form the glide direction of the stacking fault, which is oriented along the y axis of generated images
+            Basis vector (in miller indices) to form the glide direction of 
+            the stacking fault, which is oriented along the y axis of generated images
             Should be orthogonal to surface_direction
             If None, a suitable glide_direction will be found automatically
         crystalstructure: str
@@ -117,8 +123,10 @@ class GammaSurface():
                 "z": np.array(z)
                 }
             ax = np.array([x, y, z])
-        
-        alat, self.cut_at = validate_cubic_cell(a, axes=ax, crystalstructure=crystalstructure, symbol=symbol)
+
+        alat, self.cut_at = validate_cubic_cell(a, axes=ax, 
+                                                crystalstructure=crystalstructure,
+                                                symbol=symbol)
         self.offset *= alat
 
     def _vec_to_miller(self, vec, latex=True):
@@ -129,7 +137,7 @@ class GammaSurface():
             array to convert to string miller notation
             e.g. (100) or (11-2)
         latex: bool
-            Use LaTeX expressions to contstruct representation
+            Use LaTeX expressions to construct representation
         '''
         int_vec = vec.astype(int)
         l = []
@@ -140,25 +148,26 @@ class GammaSurface():
                 l.extend(str(item))
         return "(" + "".join(l) + ")"
 
-    def _gen_cellmove_images(self, base_struct, nx, ny, x_points, y_points, vacuum=0.0):
+    def _gen_cellmove_images(self, base_struct, 
+                             nx, ny, x_points, y_points, vacuum=0.0):
         # Add vacuum
         half_dist = base_struct.cell[2, 2] / 2
 
[…]
