### 2023-11-23 d3120da
ENH: More compressed quadrupole representation
