### 2023-11-24 cd8acb1
BUG: Re-add x and ylims to image gen path
