### 2023-11-24 6984748
MAINT: Docs gen now fails if notebook crashes
