### 2023-11-24 f3bffd9
BUG: Fix path lims being incorrectly calculated
