### 2023-11-27 e880924
ENH: Correct quadrupole geometry for 90deg diamond
