### 2023-11-27 610da4a
WIP: Allowed crack tip finding to work when crack not centred in cell
