+++ matscipy/fracture_mechanics/crack.py
@@ -2600,7 +2600,7 @@ def fit_crack_stress_field(atoms, r_range=(0., 50.), initial_params=None, fix_pa
     return params, err
 
 
-def find_tip_coordination(a, bondlength=2.6, bulk_nn=4):
+def find_tip_coordination(a, bondlength=2.6, bulk_nn=4, calculate_midpoint=False):
     """
     Find position of tip in crack cluster from coordination
     """
@@ -2609,14 +2609,19 @@ def find_tip_coordination(a, bondlength=2.6, bulk_nn=4):
 
     a.set_array('n_neighb', nn)
     g = a.get_array('groups')
-
+    #get midpoint of atoms
+    if calculate_midpoint:
+        atom_midpoint = (np.max(a.get_positions()[:,1])+np.min(a.get_positions()[:,1]))/2
+    else:
+        #assume crack is central in cell
+        atom_midpoint = a.cell[1, 1] / 2.0
     y = a.positions[:, 1]
-    above = (nn < bulk_nn) & (g != 0) & (y > a.cell[1, 1] / 2.0)
-    below = (nn < bulk_nn) & (g != 0) & (y < a.cell[1, 1] / 2.0)
+    above = (nn < bulk_nn) & (g != 0) & (y > atom_midpoint)
+    below = (nn < bulk_nn) & (g != 0) & (y < atom_midpoint)
 
     a.set_array('above', above)
     a.set_array('below', below)
-
+    ase.io.write('test.xyz', a)
     bond1 = above.nonzero()[0][a.positions[above, 0].argmax()]
     bond2 = below.nonzero()[0][a.positions[below, 0].argmax()]
 
