### 2023-11-27 6ed1c19
WIP: function for exact CB shift for homogeneous strain field
