### 2023-11-29 bd1276c
ENH: Rework of quadrupole structure constructor
