+++ matscipy/dislocation.py
@@ -3025,15 +3025,26 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
         core_separation = glide_separation * self.glide_distance
         core_vec = np.array([core_separation, 0, 0])
         
-
         # Replicate the unit cell enough times to fill the target cell
         cell = self.unit_cell.cell[:, :].copy()
 
-        fac = np.ceil(core_separation / np.sum(cell[0, :])).astype(int)
+        # Integer number of glide distances which fit in a single unit cell
+        glides_per_unit_cell_x = np.floor(np.linalg.norm(cell[0, :]) / (self.glide_distance - 1e-2)).astype(int)
+        glides_per_unit_cell_y = np.floor(np.linalg.norm(cell[1, :]) / (self.glide_distance - 1e-2)).astype(int)
 
-        xreps = 2*fac
-        # Try to keep vertical separation similar to horizontal separation
-        yreps = np.ceil(core_separation / np.sum(cell[1, :])).astype(int)
+        # Number of unit cells in x (glide) direction
+        xreps = int((2*glide_separation) // glides_per_unit_cell_x)
+
+        if (2*glide_separation) % glides_per_unit_cell_x:
+            # Add extra rep when we still have glides remaining
+            xreps += 1
+
+        # Number of unit cells in x (glide) direction
+        yreps = int(glide_separation // glides_per_unit_cell_y)
+
+        if glide_separation % glides_per_unit_cell_y:
+            # Add extra rep when we still have glides remaining
+            yreps += 1
 
         quad_bulk = self.unit_cell.copy() * (xreps, yreps, 1)
         cell = quad_bulk.cell[:, :].copy()
@@ -3042,29 +3053,10 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
         # New cell based on old cell vectors
         # Rhomboid shape enclosing both cores, + any stacking fault
         new_cell = np.array([
-            core_vec + cell[1, :],
-            (fac-1) * core_vec,
+            0.5 * cell[0, :] - cell[1, :],
+            0.5 * cell[0, :] + cell[1, :],
             cell[2, :]
         ])
-
-        # mask out atoms outside the cell
-        cell_points_2d = np.array([
-            [0, 0],
-            new_cell[0, :2],
-            new_cell[0, :2] + new_cell[1, :2],
-            new_cell[1, :2]
-        ])
-
-        pos = quad_bulk.get_positions()
-
-        p = pos.copy()
-        # Add small offset to prevent atoms lying on cell lines
-        # More offset in x, as one of the cell lines follows a y=x slope
-        p[:, 0] += 1e-2
-        p[:, 1] += 5e-3
-        mask = points_in_polygon2D(p[:, :2], cell_points_2d)
-        quad_bulk = quad_bulk[mask]
-
         quad_bulk.set_cell(new_cell)
         quad_bulk.wrap()
         pos = quad_bulk.get_positions()
@@ -3075,7 +3067,7 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
         lens = np.sum(new_cell, axis=0)
         core_pos_1 = lens/2 - 0.5 * core_vec
         core_pos_2 = core_pos_1 + core_vec
-        pos += core_pos_1 - self.left_dislocation.unit_cell_core_position
+        pos += core_pos_1 + self.left_dislocation.unit_cell_core_position
 
         quad_bulk.set_positions(pos)
         quad_bulk.wrap()
@@ -3087,25 +3079,6 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
         pos += disps
         quad_disloc.set_positions(pos)
[…]
