### 2023-11-30 df1758e
ENH: Add error handling for short glide_distances
