### 2023-12-05 dd425df
WIP: Thin strip fracture simulation script and utilities
