+++ matscipy/fracture_mechanics/thin_strip_utils.py
@@ -1,0 +1,496 @@
+import numpy as np
+from ase.lattice.cubic import Diamond
+from ase.constraints import ExpCellFilter
+from ase.optimize import LBFGS
+import ase.io
+import ase.io.lammpsdata
+import ase.units as units
+from matscipy.elasticity import youngs_modulus, poisson_ratio
+from matscipy.fracture_mechanics.crack import G_to_strain, thin_strip_displacement_y
+from matscipy.fracture_mechanics.clusters import set_groups
+from matscipy.fracture_mechanics.crack import find_tip_coordination
+import matplotlib.pyplot as plt
+from matscipy.cauchy_born import CubicCauchyBorn
+
+class ThinStripBuilder:
+
+    def __init__(self,el,a0,C,calc,lattice,directions,cb=None,multilattice=False,switch_sublattices=False):
+        self.el = el
+        self.a0 = a0
+        self.C = C
+        self.directions = directions
+        self.E = youngs_modulus(C, directions[1])
+        self.nu = poisson_ratio(C, directions[1], directions[0])
+        self.multilattice = multilattice
+        self.calc = calc
+        self.lattice = lattice
+        self.y_strain = 0
+        self.total_added_dist = 0
+        if self.multilattice:
+            assert cb is not None, 'Must provide a Cauchy Born object if multilattice is True'
+            self.cb = cb
+            #build rotation matrix
+            self.A = np.zeros([3, 3])
+            for i, direc in enumerate(directions):
+                self.A[:, i] = direc / np.linalg.norm(direc)
+            if switch_sublattices:
+                self.switch_sublattices = True
+
+
+    def K_to_strain(self,K,strip_height):
+        """
+        convert a stress intensity factor (in MPa sqrt(m)) to a strain for a given unstrained strip height
+        """
+        K_ase = (K/1000) * (units.GPa*np.sqrt(units.m))
+        initial_G = ((K_ase)**2)/self.E
+        strain = G_to_strain(initial_G, self.E, self.nu, strip_height)
+        return strain
+
+    def E_tensor_from_strain(self,strain_y,strain_x):
+        E = np.zeros([3,3])
+        E[0,0] = strain_y
+        E[1,1] = strain_x
+        return E
+
+    def get_equilibrium_x_strain(self,y_strain):
+        if y_strain == self.y_strain:
+            return self.x_strain
+        else:
+            #build an ASE unit cell with the correct strain
+            unitcell = self.lattice(
+                self.el, latticeconstant=self.a0, size=[1, 1, 1],directions=self.directions)
+            ase.io.write('unitcell_0.xyz',unitcell)
+            #set the strain
+            cell = unitcell.get_cell()
+            a_x_before = cell[0,0]
+            print('a_x_before',a_x_before)
+            #multiply the y component of the cell by the strain
+            cell[1,1] *= (1+y_strain)
+            #set the cell, shifting the atoms
+            unitcell.set_cell(cell,scale_atoms=True)
+            unitcell.set_pbc((True,True,True))
+            #set the calculator
+            unitcell.calc = self.calc
+            #set the expcellfilter constraint, only allowing strain to move in x direction
+            expcellfilter = ExpCellFilter(unitcell, mask=[1, 0, 0, 0, 0, 0])
+            #run the optimisation
+            opt = LBFGS(expcellfilter)
+            opt.run(fmax=1e-3)
+            ase.io.write('unitcell_1.xyz',unitcell)
+            #ge the lattice param in x direction
+            a_x = unitcell.get_cell()[0,0]
+            print('a_x_after',a_x)
+            self.y_strain = y_strain
[…]
