+++ scripts/fracture_mechanics/thin_strip_lammps.py
@@ -1,0 +1,256 @@
+from matscipy.fracture_mechanics.thin_strip_utils import ThinStripBuilder
+from matscipy import parameter
+from matscipy.fracture_mechanics.crack import G_to_strain
+import ase.io.lammpsdata
+import ase.io
+import numpy as np
+from mpi4py import MPI
+from lammps import lammps
+from matscipy.fracture_mechanics.crack import find_tip_coordination
+from matscipy.fracture_mechanics.clusters import set_groups
+import sys
+import matplotlib.pyplot as plt
+sys.path.insert(0, '.')
+import os 
+#print local directory
+print(os.getcwd())
+#print files in current working directory
+print(os.listdir())
+import params
+import ase.units as units
+
+
+me = MPI.COMM_WORLD.Get_rank()
+nprocs = MPI.COMM_WORLD.Get_size()
+strip_height = parameter('strip_height')
+strip_width = parameter('strip_width')
+strip_thickness = parameter('strip_thickness')
+vacuum = parameter('vacuum')
+crack_seed_length = parameter('crack_seed_length')
+strain_ramp_length = parameter('strain_ramp_length')
+C = parameter('C')
+a0 = parameter('a0')
+directions = [parameter('crack_direction'),
+              parameter('cleavage_plane'),
+              parameter('crack_front')]
+el = parameter('el')
+bondlength = parameter('bondlength')
+bulk_nn = parameter('bulk_nn')
+paste_threshold = parameter('paste_threshold')
+track_spacing = parameter('track_spacing',150) #distance between tracked atoms
+track_tstep = parameter('track_tstep',5) #timestep between tracked atom dumping
+lmp = lammps()
+restart = parameter('restart', False)
+checkpointing = parameter('checkpointing',False)
+checkpoint_filename = parameter('checkpoint_filename','thin_strip')
+initial_checkpoint_num = parameter('initial_checkpoint_num',1)
+lammps_set_up_path = parameter('lammps_set_up_path')
+cpnum = initial_checkpoint_num
+results_path = parameter('results_path','./results')
+checkpoint_interval = parameter('checkpoint_interval',1)
+kvals = parameter('kvals') #initial stress intensity factor (in MPa sqrt(m))
+cb = parameter('cb','None')
+multilattice = parameter('multilattice',False)
+right_hand_edge_dist = parameter('right_hand_edge_dist',0.0)
+crack_reset_pos = parameter('crack_reset_pos',strip_width/2)
+lattice = parameter('lattice')
+calc = parameter('calc')
+
+initial_K = kvals[0]
+if cb == 'None':
+    cb = None
+
+sim_time = 0
+
+plot_num = 0
+if me == 0:
+    os.makedirs(results_path, exist_ok=False)
+    tsb = ThinStripBuilder(el,a0,C,calc,lattice,directions,multilattice=multilattice,cb=cb,switch_sublattices=True)
+
+    crack_slab = tsb.build_thin_strip_with_crack(initial_K,strip_width,strip_height,strip_thickness\
+                                               ,vacuum,crack_seed_length,strain_ramp_length,track_spacing=track_spacing)
+    if not restart:
+        tracked_array = crack_slab.arrays['tracked']
+        print('Writing crack slab to file "crack.xyz"')
+        crack_slab.set_velocities(np.zeros([len(crack_slab),3]))
+        ase.io.write('crack.xyz', crack_slab, format='extxyz')
+        ase.io.lammpsdata.write_lammps_data('crack.lj',crack_slab,velocities=True)
+    else:
+        #if restarting
+        tracked_array = np.loadtxt('tracked_array.txt',dtype=int)
[…]
