### 2023-12-06 bd494f0
WIP: Removed LAMMPS input script and allowed greater parameter control thin strip
