+++ scripts/fracture_mechanics/thin_strip_lammps.py
@@ -1,4 +1,4 @@
-from matscipy.fracture_mechanics.thin_strip_utils import ThinStripBuilder
+from matscipy.fracture_mechanics.thin_strip_utils import ThinStripBuilder, set_up_simulation_lammps
 from matscipy import parameter
 from matscipy.fracture_mechanics.crack import G_to_strain
 import ase.io.lammpsdata
@@ -12,12 +12,9 @@ import sys
 import matplotlib.pyplot as plt
 sys.path.insert(0, '.')
 import os 
-#print local directory
-print(os.getcwd())
-#print files in current working directory
-print(os.listdir())
 import params
 import ase.units as units
+import warnings
 
 
 me = MPI.COMM_WORLD.Get_rank()
@@ -39,12 +36,10 @@ bulk_nn = parameter('bulk_nn')
 paste_threshold = parameter('paste_threshold')
 track_spacing = parameter('track_spacing',150) #distance between tracked atoms
 track_tstep = parameter('track_tstep',5) #timestep between tracked atom dumping
-lmp = lammps()
 restart = parameter('restart', False)
 checkpointing = parameter('checkpointing',False)
 checkpoint_filename = parameter('checkpoint_filename','thin_strip')
 initial_checkpoint_num = parameter('initial_checkpoint_num',1)
-lammps_set_up_path = parameter('lammps_set_up_path')
 cpnum = initial_checkpoint_num
 results_path = parameter('results_path','./results')
 checkpoint_interval = parameter('checkpoint_interval',1)
@@ -55,65 +50,105 @@ right_hand_edge_dist = parameter('right_hand_edge_dist',0.0)
 crack_reset_pos = parameter('crack_reset_pos',strip_width/2)
 lattice = parameter('lattice')
 calc = parameter('calc')
+mass = parameter('mass') #atomic mass
+cmds = parameter('cmds')
+sim_tstep = parameter('sim_tstep',0.001) #ps
+temp_path = parameter('temp_path','./temp')
+damping_strength = parameter('damping_strength', 0.1)
+dump_freq = parameter('dump_freq',100)
+dump_name = parameter('dump_name','dump.lammpstrj')
+thermo_freq = parameter('thermo_freq',100)
+initial_damp = parameter('initial_damp', True)
+initial_damping_time = parameter('initial_damping_time',10)
+initial_damping_strength = parameter('initial_damping_strength',0.1)
+n_steps_per_loop = parameter('n_steps_per_loop',1000)
+steady_state_check_dist = parameter('steady_state_check_dist',strip_width/5)
+max_loop_number = parameter('max_loop_number',500)
+left_damp_thickness = parameter('left_damp_thickness',60.0)
+right_damp_thickness = parameter('right_damp_thickness',60.0)
+n_v_compare = parameter('n_v_compare',10) # number of velocity values to compare to check for steady state
+ss_tol = parameter('ss_tol',0.01) # user defined tolerance for steady state around velocity mean
+
+if restart and initial_damp:
+    initial_damp = False
+
+
+lmp = lammps()
 
 initial_K = kvals[0]
 if cb == 'None':
     cb = None
 
 sim_time = 0
-
 plot_num = 0
+
+######################SET UP SIMULATION OR RESTART FROM CHECKPOINT FILES##############################
 if me == 0:
     os.makedirs(results_path, exist_ok=False)
+    os.makedirs(temp_path,exist_ok=True)
     tsb = ThinStripBuilder(el,a0,C,calc,lattice,directions,multilattice=multilattice,cb=cb,switch_sublattices=True)
 
     crack_slab = tsb.build_thin_strip_with_crack(initial_K,strip_width,strip_height,strip_thickness\
[…]
