+++ matscipy/fracture_mechanics/thin_strip_utils.py
@@ -59,7 +59,7 @@ class ThinStripBuilder:
             #build an ASE unit cell with the correct strain
             unitcell = self.lattice(
                 self.el, latticeconstant=self.a0, size=[1, 1, 1],directions=self.directions)
-            ase.io.write('unitcell_0.xyz',unitcell)
+            #ase.io.write('unitcell_0.xyz',unitcell)
             #set the strain
             cell = unitcell.get_cell()
             a_x_before = cell[0,0]
@@ -76,7 +76,7 @@ class ThinStripBuilder:
             #run the optimisation
             opt = LBFGS(expcellfilter)
             opt.run(fmax=1e-3)
-            ase.io.write('unitcell_1.xyz',unitcell)
+            #ase.io.write('unitcell_1.xyz',unitcell)
             #ge the lattice param in x direction
             a_x = unitcell.get_cell()[0,0]
             print('a_x_after',a_x)
@@ -112,12 +112,12 @@ class ThinStripBuilder:
         nx = int(width / unit_slab.cell[0, 0])
         ny = int(height/ unit_slab.cell[1, 1])
         nz = int(thickness/ unit_slab.cell[2, 2])
-        ase.io.write('unitslab.xyz',unit_slab)
+        #ase.io.write('unitslab.xyz',unit_slab)
         self.single_cell_width = unit_slab.cell[0,0]
         self.min_x_pos_dist = np.min(unit_slab.get_positions()[:,0])
         self.max_x_pos_dist = self.single_cell_width - np.max(unit_slab.get_positions()[:,0])
-        print('min x pos', self.min_x_pos_dist)
-        print('max x pos', self.max_x_pos_dist)
+        #print('min x pos', self.min_x_pos_dist)
+        #print('max x pos', self.max_x_pos_dist)
         # make sure ny is even so slab is centered on a bond
         if ny % 2 == 1:
             ny += 1
@@ -198,7 +198,7 @@ class ThinStripBuilder:
 
         crack_slab.positions[:, 1] += strain*crack_slab.positions[:, 1]
         #now apply the Poisson strain in the x direction
-        print('NU',self.nu)
+        #print('NU',self.nu)
         if force_spacing is None:
             #x_strain = -self.nu*strain
             x_strain = self.get_equilibrium_x_strain(strain)
@@ -295,7 +295,7 @@ class ThinStripBuilder:
         crack_slab.positions[:, 0] += xmean
         crack_slab.positions[:, 1] += ymean
         #write to file
-        ase.io.write('3_slab.xyz',crack_slab)
+        #ase.io.write('3_slab.xyz',crack_slab)
         return crack_slab
 
 
@@ -326,7 +326,7 @@ class ThinStripBuilder:
         old_atoms = old_atoms.copy()
         #build a thin strip at some defined strain
         new_strip = self.build_thin_strip_with_strain(K,width,height,thickness,vacuum,force_spacing=avg_cell_length)
-        ase.io.write('new_strip_temp.xyz',new_strip)
+        #ase.io.write('new_strip_temp.xyz',new_strip)
         paste_num_cells = int(crop/self.single_cell_width)
         right_hand_edge_cells = int(right_hand_edge_dist/self.single_cell_width)
         print(f'Copying and pasting {paste_num_cells} unit cells')
@@ -348,18 +348,18 @@ class ThinStripBuilder:
         
         #align the old_atom positions such that the right most atom aligns with
         #the right most atom of the new strip
-        ase.io.write('new_strip_init.xyz',new_strip)
+        #ase.io.write('new_strip_init.xyz',new_strip)
[…]
