### 2023-12-06 88a6d8e
ENH: Dissociated Quadrupoles, extensions
