+++ matscipy/dislocation.py
@@ -2517,7 +2517,7 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
                                    f'did not converge in {max_iter} cycles')
         return disp2
 
-    def displacements(self, bulk_positions, center, use_atomman=True,
+    def displacements(self, bulk_positions, center, use_atomman=True, partial_distance=0,
                       self_consistent=True, tol=1e-6, max_iter=100, verbose=True):
         
         if use_atomman:
@@ -2881,6 +2881,9 @@ class CubicCrystalDissociatedDislocation(CubicCrystalDislocation, metaclass=ABCM
         self.left_dislocation.invert_burgers()
         self.right_dislocation.invert_burgers()
 
+        # Flip dislocations
+        self.left_dislocation, self.right_dislocation = self.right_dislocation, self.left_dislocation
+
 
     def build_cylinder(self, radius, partial_distance=0,
                        core_position=np.array([0., 0., 0.]),
@@ -2969,7 +2972,8 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
         self.left_dislocation.invert_burgers()
         super().__init__(*args, **kwargs)
 
-    def periodic_displacements(self, positions, v1, v2, left_pos, right_pos, disp_tol=1e-3, max_neighs=60, **kwargs):
+    def periodic_displacements(self, positions, v1, v2, left_pos, right_pos, disp_tol=1e-3, max_neighs=60, 
+                               partial_distance=0, **kwargs):
         '''
         Calculate the stroh displacements for the periodic structure defined by 2D lattice vectors v1 & v2
         given the core positions.
@@ -2987,12 +2991,22 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
         max_neighs: int
             Maximum number of nth-nearest neighbour cells to consider
             max_neighs=0 is just the current cell, max_neighs=1 also includes nearest-neighbour cells, ...
+        partial_distance: int or list of int
+            Distance between dissociated partial dislocations
+            If a list, specify a partial distance for each dislocation
+            If an int, specify a partial distance for both dislocations
 
         returns
         -------
         disps: np.array
             displacements
         '''
+
+        if np.issubdtype(type(partial_distance), np.integer):
+            partial_dist = [partial_distance, partial_distance]
+        else:
+            # Assume list or array
+            partial_dist = partial_distance
 
         displacements = np.zeros_like(positions)
         disp_update = np.zeros_like(positions)
@@ -3007,9 +3021,11 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
 
                     # In a new cell
                     disp_update += self.left_dislocation.displacements(
-                        positions, left_pos+ ix * v1 + iy * v2, **kwargs)
+                        positions, left_pos+ ix * v1 + iy * v2, partial_distance=partial_dist[0], 
+                        **kwargs)
                     disp_update += self.right_dislocation.displacements(
-                        positions, right_pos+ ix * v1 + iy * v2, **kwargs)
+                        positions, right_pos+ ix * v1 + iy * v2, partial_distance=partial_dist[1],
+                        **kwargs)
 
             displacements += disp_update
             max_disp_change = np.max(np.linalg.norm(disp_update, axis=-1))
@@ -3019,17 +3035,35 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
         return displacements
 
 
-    def build_quadrupole(self, glide_separation=4, **kwargs):
+    def build_quadrupole(self, glide_separation=4, partial_distance=0, extension=0, **kwargs):
+        '''
[…]
