### 2023-12-06 954bb23
ENH: Verbosity control for periodic dislocation convergence
