### 2023-12-06 4d3540b
WIP: steady state check uses shorter trajectory
