### 2023-12-07 3cc63b0
WIP: Fixed error bar plot
