+++ scripts/fracture_mechanics/thin_strip_lammps.py
@@ -308,9 +308,9 @@ for knum,K in enumerate(kvals):
         res = np.reshape(np.loadtxt(f'{results_path}/K_vs_velocity.txt'),[-1,4])
         print('res',res)
         plt.figure()
-        mask = np.array(res[3,:].flatten(),dtype=bool)
-        plt.errorbar(res[0,:][mask],res[1,:][mask],yerr=res[2,:][mask],fmt='o',color='blue',ecolor='blue',capsize=5,label='Steady state reached')
-        plt.errorbar(res[0,:][~mask],res[1,:][~mask],yerr=res[2,:][~mask],fmt='o',color='red',ecolor='red',capsize=5,label='Steady state not reached')
+        mask = np.array(res[:,3].flatten(),dtype=bool)
+        plt.errorbar(res[:,0][mask],res[:,1][mask],yerr=res[:,2][mask],fmt='o',color='blue',ecolor='blue',capsize=5,label='Steady state reached')
+        plt.errorbar(res[:,0][~mask],res[:,1][~mask],yerr=res[:,2][~mask],fmt='o',color='red',ecolor='red',capsize=5,label='Steady state not reached')
         plt.xlabel('K (MPa sqrt(m))')
         plt.ylabel('Final crack velocity (km/s)')
         plt.legend()
