### 2023-12-07 e4261b0
WIP: Fixed num points determination and added group limiter
