### 2023-12-07 326fff9
WIP: Bug fix with copy-paste window
