### 2023-12-08 e66356e
WIP: Optional dumping and better guess for pasting lattice param
