### 2023-12-08 b964001
ENH: Dislocation NGL viewer method
