+++ matscipy/dislocation.py
@@ -2340,7 +2340,7 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
     # (see https://stackoverflow.com/questions/472000/usage-of-slots for more details on __slots__)
     __slots__ = ("burgers_dimensionless", "unit_cell_core_position_dimensionless",
                  "glide_distance_dimensionless", "crystalstructure", "axes",
-                 "C11", "C12", "C44", "alat", "unit_cell")
+                 "C11", "C12", "C44", "alat", "unit_cell", "name")
 
     # Attributes with defaults
     # These may be overridden by child classes
@@ -2633,6 +2633,11 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
         disloc.set_array('fix_mask', final_fix_mask)
         disloc.set_constraint(FixAtoms(mask=final_fix_mask))
 
+        disloc.info["core_positions"] = [list(shifted_center)]
+        disloc.info["burgers_vectors"] = [list(self.burgers)]
+        disloc.info["dislocation_types"] = [self.name]
+        disloc.info["dislocation_classes"] = [str(self.__class__)]
+
         # adding vacuum and centering breaks consistency
         # of displacement =  dislo.positions - bulk.positions
         # which is essential for plot_vitek and other tools
@@ -2779,6 +2784,161 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
         impurities_disloc.cell = disloc.cell
 
         return impurities_disloc
+    
+    @staticmethod
+    def view_cyl(system, scale=0.5, CNA_color=True, add_bonds=False,
+                     line_color=[0, 1, 0], disloc_names=None):
+        '''
+        NGLview-based visualisation tool for structures generated from the dislocation class
+
+        Parameters
+        ----------
+        system: ase.Atoms
+            Dislocation system to view
+        scale: float
+            Adjust radiusScale of add_ball_and_stick (add_bonds=True) or add_spacefill (add_bonds=False)
+        CNA_color: bool
+            Turn on atomic/bond colours based on Common Neighbour Analysis structure identification
+        add_bonds: bool
+            Add atomic bonds
+        line_color: list or list of lists
+            [R, G, B] values for the colour of dislocation lines
+            if a list of lists, line_color[i] defines the colour of the ith dislocation (see structure.info["dislocation_types"])
+        disloc_names: list of str
+            Manual override for the automatic naming of the dislocation lines 
+            (see structure.info["dislocation_types"] for defaults)
+        '''
+
+        def _plot_disloc_line(view, disloc_pos, disloc_name, z_length, color=[0, 1, 0]):
+            '''Add dislocation line to the view as a cylinder and two cones.
+            The cylinder is hollow by default so the second cylinder is needed to close it.
+            In case partial distance is provided, two dislocation lines are added and are shifter accordingly.
+            
+            Parameters
+            ----------
+            view: NGLview
+                NGLview plot of the atomic structure
+            disloc_pos: list or array
+                Position of the dislocation core
+            color: list
+                [R, G, B] colour of the dislocation line
+            '''
+
+            view.shape.add_cylinder((disloc_pos[0], disloc_pos[1], -2.0), 
+                                    (disloc_pos[0], disloc_pos[1], z_length - 0.5),
+                                    color,
+                                    [0.3],
+                                    disloc_name)
+            
+            view.shape.add_cone((disloc_pos[0], disloc_pos[1], -2.0), 
[…]
