### 2023-12-09 95d5ef2
WIP: Fixed all edges of strip
