+++ matscipy/fracture_mechanics/thin_strip_utils.py
@@ -185,7 +185,7 @@ class ThinStripBuilder:
         test_slab.set_pbc((False,True,True))
         return test_slab
 
-    def build_thin_strip_with_strain(self,K,width,height,thickness,vacuum,force_spacing=None):
+    def build_thin_strip_with_strain(self,K,width,height,thickness,vacuum,apply_x_strain=False,force_spacing=None):
         #force spacing is the width of a single unit cell to force upon the slab
         crack_slab = self.build_thin_strip(width,height,thickness,vacuum)
         strain = self.K_to_strain(K,height)
@@ -199,12 +199,15 @@ class ThinStripBuilder:
         crack_slab.positions[:, 1] += strain*crack_slab.positions[:, 1]
         #now apply the Poisson strain in the x direction
         #print('NU',self.nu)
-        if force_spacing is None:
-            #x_strain = -self.nu*strain
-            x_strain = self.get_equilibrium_x_strain(strain)
+        if apply_x_strain:
+            if force_spacing is None:
+                #x_strain = -self.nu*strain
+                x_strain = self.get_equilibrium_x_strain(strain)
+            else:
+                x_strain = (force_spacing-self.single_cell_width)/self.single_cell_width
+                print('effective x strain', x_strain)
         else:
-            x_strain = (force_spacing-self.single_cell_width)/self.single_cell_width
-            print('effective x strain', x_strain)
+            x_strain = 0
         
         crack_slab.positions[:, 0] += x_strain*crack_slab.positions[:, 0]
         
@@ -229,7 +232,7 @@ class ThinStripBuilder:
         return crack_slab
 
 
-    def build_thin_strip_with_crack(self,K,width,height,thickness,vacuum,crack_seed_length,strain_ramp_length,track_spacing=0):
+    def build_thin_strip_with_crack(self,K,width,height,thickness,vacuum,crack_seed_length,strain_ramp_length,track_spacing=0,apply_x_strain=False):
         """
         build a thin strip at some defined strain with a crack of some defined length
         When building the crack, add tracked atoms every track_spacing along x 
@@ -274,21 +277,22 @@ class ThinStripBuilder:
                                         left + crack_seed_length +
                                                 strain_ramp_length)
         
-        #first part has no associated Poisson strain
-        #last part has easy to calculate Poisson strain
-        #middle part has a poisson strain gradient
-        a = left + crack_seed_length
-        b = left + crack_seed_length + strain_ramp_length
-        y = crack_slab.positions[:,1]
-        x = crack_slab.positions[:,0]
-        u_x = np.zeros_like(x)
-        #get x_strain
-        x_strain = self.get_equilibrium_x_strain(strain)
-        u_x[x>b] = x_strain*x[x>b]
-        middle = (x >= a) & (x <= b)
-        f = (x[middle] - a) / (b - a)
-        u_x[middle] = (x_strain * f * x[middle])
-        crack_slab.positions[:,0] += u_x
+        if apply_x_strain:
+            #first part has no associated Poisson strain
+            #last part has easy to calculate Poisson strain
+            #middle part has a poisson strain gradient
+            a = left + crack_seed_length
+            b = left + crack_seed_length + strain_ramp_length
+            y = crack_slab.positions[:,1]
+            x = crack_slab.positions[:,0]
+            u_x = np.zeros_like(x)
+            #get x_strain
+            x_strain = self.get_equilibrium_x_strain(strain)
[…]
