+++ scripts/fracture_mechanics/thin_strip_lammps.py
@@ -93,7 +93,7 @@ if me == 0:
     tsb = ThinStripBuilder(el,a0,C,calc,lattice,directions,multilattice=multilattice,cb=cb,switch_sublattices=True)
 
     crack_slab = tsb.build_thin_strip_with_crack(initial_K,strip_width,strip_height,strip_thickness\
-                                               ,vacuum,crack_seed_length,strain_ramp_length,track_spacing=track_spacing)
+                                               ,vacuum,crack_seed_length,strain_ramp_length,track_spacing=track_spacing,apply_x_strain=False)
     
     if not restart:
         tracked_array = crack_slab.arrays['tracked']
