### 2023-12-09 0743b28
WIP: Fixed simulation timestep bug
