+++ scripts/fracture_mechanics/thin_strip_lammps.py
@@ -265,13 +265,13 @@ for knum,K in enumerate(kvals):
                     
                     steady_state_data = np.reshape(np.loadtxt(f'{results_path}/steady_state.txt'),[-1,3])
                     plt.figure()
-                    plt.plot((steady_state_data[:,0]/(10**6)),steady_state_data[:,1],'o')
+                    plt.plot((steady_state_data[:,0]/(10**3)),steady_state_data[:,1],'o')
                     plt.xlabel('Time (ns)')
                     plt.ylabel('Measured crack velocity (km/s)')
                     plt.savefig(f'{results_path}/velocity_steady_state.png')
                     plt.close()
                     plt.figure()
-                    plt.plot((steady_state_data[:,0]/(10**6)),steady_state_data[:,2],'o')
+                    plt.plot((steady_state_data[:,0]/(10**3)),steady_state_data[:,2],'o')
                     plt.xlabel('Time (ns)')
                     plt.ylabel('Steady state measure')
                     plt.savefig(f'{results_path}/steady_state_measure.png')
