### 2023-12-11 5c5ec0c
WIP: Allowed tracked atoms to be beneath cracks and fixed elastic constants
