### 2023-12-11 60eac1c
WIP: Strip strain now computed with true strip height, not user supplied
