### 2023-12-11 13276c0
ENH: view_quad for NGLview on quadrupoles (WIP)
