### 2023-12-11 ecf4708
ENH: Quadrupole Glide image gen
