### 2023-12-12 de0e73f
WIP: Added crack arrest detection
