### 2023-12-13 889c905
BUG: Fix for BCC bulk cells not being correctly constructed
