### 2023-12-13 9f5da16
ENH: Allow core shifts for glide & kink initial conditions
