### 2023-12-13 b9932a5
WIP: Added initial kick
