+++ scripts/fracture_mechanics/thin_strip_lammps.py
@@ -60,7 +60,10 @@ dump_name = parameter('dump_name','dump.lammpstrj')
 thermo_freq = parameter('thermo_freq',100)
 initial_damp = parameter('initial_damp', True)
 initial_damping_time = parameter('initial_damping_time',10)
-initial_damping_strength = parameter('initial_damping_strength',0.1)
+initial_damping_strengths = parameter('initial_damping_strengths',[0.1*(i+1) for i in range(initial_damping_time)])
+
+assert len(initial_damping_strengths) == initial_damping_time, 'initial_damping_strengths must be a list of length initial_damping_time'
+
 n_steps_per_loop = parameter('n_steps_per_loop',1000)
 steady_state_check_dist = parameter('steady_state_check_dist',strip_width/5)
 max_loop_number = parameter('max_loop_number',500)
@@ -69,6 +72,8 @@ right_damp_thickness = parameter('right_damp_thickness',60.0)
 n_v_compare = parameter('n_v_compare',10) # number of velocity values to compare to check for steady state
 ss_tol = parameter('ss_tol',0.01) # user defined tolerance for steady state around velocity mean
 dump_files = parameter('dump_files', True)
+initial_kick = parameter('initial_kick', False) #give the crack an initial kick at the end of the damping time
+kick_timestep = parameter('kick_timestep', 9) # timestep to give the crack an initial kick
 
 cpnum = initial_checkpoint_num
 
@@ -79,7 +84,8 @@ if int(strip_width/track_spacing) > 25:
     raise ValueError('LAMMPS only allows 32 groups total, reduce track spacing')
 
 if restart and initial_damp:
-    initial_damp = False
+    #initial_damp = False
+    initial_damp = True
 
 
 lmp = lammps()
@@ -146,21 +152,22 @@ for knum,K in enumerate(kvals):
         unique_v_vals = 0
         unscaled_crack = ase.io.lammpsdata.read_lammps_data(f'{temp_path}/crack.lj',atom_style='atomic')
         unscaled_crack.set_pbc([False,False,True])
-        # ase.io.write(f'{knum}_rescale0.xyz',unscaled_crack)
-        #print out y position of last atom of unscaled crack
-        # print('in pos',unscaled_crack.get_positions()[-1,1])
+
         rescale_crack = tsb.rescale_K(unscaled_crack,K_curr,K,strip_height,strip_width,strip_thickness,vacuum,tip_pos)
-        #print out y position of last atom of rescaled crack
-        # print('out pos',rescale_crack.get_positions()[-1,1])
         #re-write lammps data file
         ase.io.lammpsdata.write_lammps_data(f'{temp_path}/crack.lj',rescale_crack,velocities=True,masses=True)
-        # ase.io.write(f'{knum}_rescale1.xyz',rescale_crack)
     
     K_curr = K
     if (knum == 0) and initial_damp:
-        intial_damp = True
+        initial_damp = True
     else:
-        intial_damp = False
+        initial_damp = False
+    
+    if (knum == 0) and initial_kick:
+        initial_kick = True
+    else:
+        initial_kick = False
+
     at_steady_state = False
     i = -1
     final_v = -1 #set to -1 so that if steady state is not reached, it is obvious
@@ -177,9 +184,9 @@ for knum,K in enumerate(kvals):
         set_up_simulation_lammps(lmp,temp_path,mass,cmds,sim_tstep=sim_tstep,damping_strength_right=damping_strength_right,damping_strength_left=damping_strength_left
                                  , dump_freq=dump_freq, dump_name=dump_name, thermo_freq=thermo_freq, dump_files=dump_files,
                                  left_damp_thickness=left_damp_thickness, right_damp_thickness=right_damp_thickness)
-        if (i < initial_damping_time) and (intial_damp):
[…]
