### 2023-12-13 67a829a
WIP: Fixed restart bug and added warning
