### 2023-12-15 10850b7
ENH: Change bracketing of miller index axes labels in gamma surface plots
