### 2023-12-21 0c9e80e
DOC: Add example for topology building of amorphous carbon for non-reactive simulations with LAMMPS
