+++ docs/index.rst
@@ -8,6 +8,7 @@
    ./applications/index
    ./tools/index
    ./calculators/index
+   ./topology/index
    ./api
 
 
