+++ docs/topology/parameters_Reichenbach2023.in
@@ -1,0 +1,1667 @@
+# An All-Atom Force Field for Dry and Water-Lubricated Carbon Tribological Interfaces
+# Thomas Reichenbach, Severin Sylla, Leonhard Mayrhofer, Pedro Antonio Romero, Paul Schwarz, Michael Moseler, Gianpietro Moras
+# DATE: 2023-12-14
+#
+# Parameters adopted from the original OPLS FF:
+# H1 charge: Jorgensen JACS 1996 - RH alkanes
+# H2 charge: Jorgensen JACS 1996 - H benzene
+# Dihedrals:
+# C-C-C(sp2)-H2, H2-C(sp2)-CZ-H2, O-C-C(sp2)-H2: Weiner JCP 1986 benzene
+# C-CY-O2-H5: Jorgensen JACS 1996 - phenol
+# C-C(sp3)-O1-H4, H1-C(sp3)-O1-H4: Jorgensen JACS 1996 - alcohols
+#
+# Parameters adopted from the TIP3P water model:
+# HW and OW charges and LJ parameters:  Price J Chem Phys 2004
+# HW-OW bond and HW-OW-HW angle: MacKerell J Phys Chem B 1998, Jorgensen JACS 1981
+#
+#
+# Cutoffs-LJ-Coulomb
+# Name Cutoff(A)
+LJ 14.0  #
+C  7.4   #
+
+# Lennard-Jones and Charges
+# Element Epsilon(eV) Sigma(A) Charge(e)
+C1 0.007150 3.801 -0.060 #
+C2 0.007150 3.801 0.180 #
+C3 0.007150 3.801 -0.120 #
+C4 0.007150 3.801 0.120 #
+C5 0.007150 3.801 -0.180 #
+C6 0.007150 3.801 0.060 #
+CA 0.002906 3.340 0.000 #
+CB 0.000000 0.000 0.000 #
+CD 0.000000 0.000 0.000 #
+CG 0.002906 3.340 0.000 #
+CX 0.004309 3.729 -0.230 #
+CY 0.004309 3.729 0.020 #
+CZ 0.004309 3.729 -0.115 #
+H1 0.000921 2.176 0.060 #
+H2 0.001626 2.120 0.115 #
+H4 0.000000 0.000 0.410 #
+H5 0.000000 0.000 0.420 #
+HW 0.000000 0.000 0.415 #
+O1 0.005573 3.061 -0.590 #
+O2 0.005262 2.993 -0.440 #
+OW 0.004423 3.188 -0.830 #
+
+# LJ-pairs
+# Name Epsilon(eV) Sigma(A) Cutoff(A)
+CG-CG 0.002906 3.34 3.7490232413533056 # purely repulsive
+CA-CG 0.002906 3.34 3.7490232413533056 # purely repulsive
+C1-CG 0.004558278183700508 3.5630520624880013 14.0 #
+C2-CG 0.004558278183700508 3.5630520624880013 14.0 #
+C3-CG 0.004558278183700508 3.5630520624880013 14.0 #
+C4-CG 0.004558278183700508 3.5630520624880013 14.0 #
[…]
