### 2023-12-21 eb4f525
BUGL Quadrupole -> CCDQuadrupole, not just CCD
