### 2023-12-22 c274857
ENH: Radial masking for arbitrary polygons
