### 2023-12-22 2f3feb6
ENH: Generic _build_bulk_cyl method
