### 2023-12-22 2f33284
ENH: CCD.build_cylinder using new generic method
