+++ matscipy/dislocation.py
@@ -2477,10 +2477,10 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
 
         base_cell = self.unit_cell
 
-        xreps = np.ceil(base_cell.cell[0, 0] / (targetx - 1e-2))
-        yreps = np.ceil(base_cell.cell[0, 0] / (targety - 1e-2))
-
-        return base_cell * (xreps, yreps, 1)
+        xreps = np.ceil(targetx / (base_cell.cell[0, 0] - 1e-2)).astype(int)
+        yreps = np.ceil(targety / (base_cell.cell[1, 1] - 1e-2)).astype(int)
+        sup = base_cell.copy() * (xreps, yreps, 1)
+        return sup
 
     def _build_bulk_cyl(self, radius, core_positions, fix_rad):
         '''
@@ -2504,28 +2504,37 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
             Bulk cyl, cut down based on radius, complete
             with FixAtoms constraints
         '''
-        from utils import radial_mask_from_polygon2D
+        from matscipy.utils import radial_mask_from_polygon2D
 
         if len(core_positions.shape) == 1:
             core_positions = core_positions[np.newaxis, :]
 
-        xlen = np.max(core_positions[:, 0]) - np.min(core_positions[:, 0])
-        ylen = np.max(core_positions[:, 1]) - np.min(core_positions[:, 1])
+        xlen = np.max(core_positions[:, 0]) - np.min(core_positions[:, 0]) + 2*radius
+        ylen = np.max(core_positions[:, 1]) - np.min(core_positions[:, 1]) + 2*radius
 
         sup = self._build_supercell(xlen, ylen)
 
-        mask = radialmask_from_polygon2D(sup.get_positions(), core_positions, radius, inner=True)
+        center = np.diag(sup.cell) / 2
+
+        pos = sup.get_positions()
+        pos += center - core_positions[0, :]
+        sup.set_positions(pos)
+        sup.wrap()
+
+        new_core_positions = core_positions.copy()
+        new_core_positions += center - core_positions[0, :]
+        
+        mask = radial_mask_from_polygon2D(sup.get_positions(), new_core_positions, radius, inner=True)
 
         cyl = sup[mask]
 
         if fix_rad:
-            fix_mask = not radialmask_from_polygon2D(sup.get_positions(), core_positions, radius - fix_rad, inner=True)
+            fix_mask = ~radial_mask_from_polygon2D(cyl.get_positions(), new_core_positions, radius - fix_rad, inner=True)
 
             fix_atoms = FixAtoms(mask=fix_mask)
             cyl.set_constraint(fix_atoms)
-        return cyl
-
-
+            cyl.arrays["fix_mask"] = fix_mask
+        return cyl, new_core_positions
 
     def plot_unit_cell(self, ms=250, ax=None):
         import matplotlib.pyplot as plt
@@ -2606,92 +2615,22 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
 
     def build_cylinder(self, radius,
                        core_position=np.array([0., 0., 0.]),
-                       extension=np.array([0., 0., 0.]),
+                       extra_glide_widths=0,
                        fix_width=10.0, self_consistent=None):
 
         if self_consistent is None:
             self_consistent = self.self_consistent
+        
 
-        extent = np.array([2 * (radius + fix_width),
-                           2 * (radius + fix_width), 1.])
-        repeat = np.ceil(extent / np.diag(self.unit_cell.cell)).astype(int)
-
-        # if the extension and core position is
-        # within the unit cell, do not add extra unit cells
-        repeat_extension = np.floor(2.0 * extension /
[…]
