### 2023-12-22 9d3679d
ENH: CCD.displacements using core positions array
