### 2023-12-22 5ff9d22
ENH: Generalise more of build_cylinder functionality
