### 2023-12-22 c7f4b44
BUG: Fix bug in polys sometime not being closed
