+++ matscipy/dislocation.py
@@ -2584,9 +2584,38 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
         ax.set_xlabel(r"$\AA$")
         ax.set_ylabel(r"$\AA$")
 
-    def self_consistent_displacements(self, solver, bulk_positions, core_positions, 
+    def self_consistent_displacements(self, solvers, bulk_positions, core_positions, 
                                       tol=1e-6, max_iter=100, verbose=True):
-        
+        '''
+        Compute dislocation displacements self-consistently, with max_iter capping the number of iterations
+        Each dislocation core uses a separate solver, which computes the displacements associated with positions 
+        relative to that core (i.e. that the core is treated as lying at [0, 0, 0])
+
+        Parameters
+        ----------
+        solvers: list
+            List of functions, solvers[i] computes the displacement field for the dislocation
+            defined by core_positions[i, :]
+        bulk_positions: np.array
+            (N, 3) array of positions of atoms in the bulk configuration
+        core_positions: np.array
+            Positions of each dislocation core (shape = (ncores, 3))
+        tol: float
+            Displacement tolerance for the convergence of the self-consistent loop
+        max_iter: int
+            Maximum number of iterations of the self-consistent cycle to run before throwing 
+            an exception
+        verbose: bool
+            Enable/Disable printing progress of the self-consistent cycle each iteration
+
+        Returns
+        -------
+        displacements: np.array
+        shape (N, 3) array of displacements to apply to bulk_positions to form
+        the dislocation structure
+
+        Raises a RuntimeError if max_iter is reached without convergence
+        '''
         if len(core_positions.shape) == 1:
             core_positions = core_positions[np.newaxis, :]
 
@@ -2594,7 +2623,7 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
 
         disp1 = np.zeros_like(bulk_positions)
         for i in range(ncores):
-            disp1 += solver(bulk_positions - core_positions[i, :]).real
+            disp1 += solvers[i](bulk_positions - core_positions[i, :]).real
         
         if max_iter == 0:
             return disp1
@@ -2605,8 +2634,8 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
             disloc_positions = bulk_positions + disp1
             
             disp2 = np.zeros_like(bulk_positions)
-            for i in range(ncores):
-                disp2 += solver(disloc_positions - core_positions[i, :]).real
+            for j in range(ncores):
+                disp2 += solvers[j](disloc_positions - core_positions[j, :]).real
 
             res = np.abs(disp1 - disp2).max()
             disp1 = disp2
@@ -2617,10 +2646,25 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
             
         raise RuntimeError('Self-consistency' +
                             f'did not converge in {max_iter} cycles')
+    
+    def get_solvers(self, use_atomman=True):
+        '''
+        Get list of dislocation displacement solvers
 
-    def displacements(self, bulk_positions, core_positions, use_atomman=True, partial_distance=0,
-                      self_consistent=True, tol=1e-6, max_iter=100, verbose=True):
-        
+        As CubicCrystalDislocation models a single core, there is only one solver,
+        therefore solvers is a len(1) list
+
+        Parameters
+        ----------
+        use_atomman: bool
+            Use the Stroh solver included in atomman (requires atomman package) to
+            solve for displacements, or use the AnisotropicDislocation class
+
+        Returns
+        -------
+        solvers: list
+            List of functions f(positions) = displacements
+        '''
         if use_atomman:
             # Use stroh solver from atomman package
             if self.stroh is None:
[…]
