### 2023-12-22 bd3a8a9
ENH: Apply changes to displacements to Quadrupole builder
