+++ matscipy/dislocation.py
@@ -3274,8 +3274,8 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
         self.glides_per_unit_cell = np.floor((self.unit_cell.cell[0, 0] + self.unit_cell.cell[1, 0]) 
                                               / (self.glide_distance - 1e-2)).astype(int)
 
-    def periodic_displacements(self, positions, v1, v2, left_pos, right_pos, disp_tol=1e-3, max_neighs=60, 
-                               partial_distance=0, verbose="periodic", **kwargs):
+    def periodic_displacements(self, positions, v1, v2, core_positions, disp_tol=1e-3, max_neighs=60, 
+                               verbose="periodic", **kwargs):
         '''
         Calculate the stroh displacements for the periodic structure defined by 2D lattice vectors v1 & v2
         given the core positions.
@@ -3284,19 +3284,13 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
             Atomic positions to calculate displacements at
         v1, v2: np.array
             Lattice vectors defining periodicity in XY plane (dislocation line is in Z)
-        left_pos: np.array
-            position of the left dislocation core
-        right_pos: np.array
-            position of the right dislocation core
+        core_positions: np.array
+            positions of the all dislocation cores (including all partial cores)
         disp_tol: float
             Displacement tolerance controlling the termination of displacement convergence
         max_neighs: int
             Maximum number of nth-nearest neighbour cells to consider
             max_neighs=0 is just the current cell, max_neighs=1 also includes nearest-neighbour cells, ...
-        partial_distance: int or list of int
-            Distance between dissociated partial dislocations
-            If a list, specify a partial distance for each dislocation
-            If an int, specify a partial distance for both dislocations
         verbose: bool, str or None
             Controls verbosity of the displacement solving
             False, None, or "off" : No printing
@@ -3309,6 +3303,8 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
         disps: np.array
             displacements
         '''
+
+        ncores = core_positions.shape[0]
 
         if type(verbose) == str:
             if verbose.lower() == "off":
@@ -3341,12 +3337,6 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
             # CubicCrystalDislocation.displacements doesn't print unless SCF is turned on 
             disloc_verbose = disloc_verbose * kwargs["self_consistent"]
 
-        if np.issubdtype(type(partial_distance), np.integer):
-            partial_dist = [partial_distance, partial_distance]
-        else:
-            # Assume list or array
-            partial_dist = partial_distance
-
         displacements = np.zeros_like(positions)
         disp_update = np.zeros_like(positions)
         
@@ -3360,14 +3350,13 @@ class CubicCrystalDislocationQuadrupole(CubicCrystalDissociatedDislocation):
 
                     # In a new cell
                     if disloc_verbose:
-                        print(f"Left dislocation displacements for cell {ix} {iy}")
-                    disp_update += self.left_dislocation.displacements(
-                        positions, left_pos+ ix * v1 + iy * v2, partial_distance=partial_dist[0], 
-                        verbose=disloc_verbose, **kwargs)
-                    if disloc_verbose:
-                        print(f"Right dislocation displacements for cell {ix} {iy}")
-                    disp_update += self.right_dislocation.displacements(
-                        positions, right_pos+ ix * v1 + iy * v2, partial_distance=partial_dist[1],
[…]
