### 2023-12-22 7268e15
ENH: Allow for extensions from multiple dislocation cores
