+++ matscipy/dislocation.py
@@ -2432,7 +2432,6 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
                                            axes, slip_plane, disloc_line, 
                                            self.burgers)
 
-
     # @property & @var.setter decorators used to ensure var and var_dimensionless don't get out of sync
     @property
     def burgers(self):
@@ -2483,7 +2482,8 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
         sup = base_cell.copy() * (xreps, yreps, 1)
         return sup
 
-    def _build_bulk_cyl(self, radius, core_positions, fix_rad, self_consistent):
+    def _build_bulk_cyl(self, radius, core_positions, fix_rad, extension,
+                        self_consistent, use_atomman, ext_idx):
         '''
         Build bulk cylinder config from args supplied by self.build_cylinder
 
@@ -2498,6 +2498,16 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
             e.g. for glide configs, etc
         fix_rad: float
             Radius for fixed atoms constraints
+        extension: int
+            Ensure the cell extends out from the center to center + extension
+        ext_idx: int or np.array of ints
+            idxs of the cores each part of extension should apply to
+            extension = np.array([
+                            [-5.0, 0, 0],
+                            [5.0, 0, 0.0]
+                            ])
+            and ext_idxs = np.array([0, -1]) will extend the cell by 5Ang left of the first core, 
+            and 5Ang right of the last core
 
         Returns
         -------
@@ -2514,18 +2524,28 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
         if len(core_positions.shape) == 1:
             core_positions = core_positions[np.newaxis, :]
 
-        ncores = core_positions.shape[0]
+        if len(extension.shape) == 1:
+            extension = extension[np.newaxis, :]
+        ext_idx = np.array(ext_idx)
 
-        xlen = np.max(core_positions[:, 0]) - np.min(core_positions[:, 0]) + 2*radius
-        ylen = np.max(core_positions[:, 1]) - np.min(core_positions[:, 1]) + 2*radius
+        exts = extension + core_positions[ext_idx, :]
+
+        xmax = np.max([np.max(core_positions[:, 0]), np.max(exts[:, 0])])
+        xmin = np.min([np.min(core_positions[:, 0]), np.min(exts[:, 0])])
+
+        ymax = np.max([np.max(core_positions[:, 1]), np.max(exts[:, 1])])
+        ymin = np.min([np.min(core_positions[:, 1]), np.min(exts[:, 1])])
+
+        xlen = xmax - xmin + 2*radius
+        ylen = ymax - ymin + 2*radius
 
         sup = self._build_supercell(xlen, ylen)
 
         center = np.diag(sup.cell) / 2
 
+        # Shift everything so that the dislocations are reasonably central
 
-        # Shift everything so that the dislocations ae centralised
-        shift = center - np.average(core_positions, axis=0)  
+        shift = center - 0.5 * (np.array([xmax, ymax, 0]) + np.array([xmin, ymin, 0]))
 
         pos = sup.get_positions()
         pos += shift
@@ -2533,9 +2553,17 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
         sup.wrap()
 
         new_core_positions = core_positions.copy()
+
         new_core_positions += shift
+
+        mask_positions = np.vstack([new_core_positions, exts + shift])
+
+        idxs = np.argsort(mask_positions, axis=0)
+        mask_positions = np.take_along_axis(mask_positions, idxs, axis=0)
+
+        print(mask_positions)
         
[…]
