### 2024-01-02 248a19c
TST: Fix failing CCDD tests for new displacements
