### 2024-01-02 0f3ad06
MAINT: Remove commented out code
