### 2024-01-02 dc8b0a6
WIP: Added support for non-linear stress strain relationship
