+++ matscipy/fracture_mechanics/thin_strip_utils.py
@@ -21,8 +21,7 @@ class ThinStripBuilder:
         self.directions = directions
         self.E = youngs_modulus(C, directions[1])
         print('youngs modulus', self.E)
-        pr = poisson_ratio(C, directions[1], directions[0])
-        print('elastic constant poisson ratio',pr)
+
         self.multilattice = multilattice
         self.calc = calc
         self.lattice = lattice
@@ -38,19 +37,81 @@ class ThinStripBuilder:
             if switch_sublattices:
                 self.switch_sublattices = True
         
-        self.nu = self.measure_poisson_ratio()
-        print('measured poisson ratio',self.nu)
+        #self.nu = self.measure_poisson_ratio()
+        #print('measured poisson ratio',self.nu)
+        #get effective modulus from poisson ratios
+        #
+        nu_yx = poisson_ratio(C, directions[1], directions[0])
+        nu_zx = poisson_ratio(C, directions[2], directions[0])
+        nu_zy = poisson_ratio(C, directions[2], directions[1])
+        nu_xy = poisson_ratio(C, directions[0], directions[1])
+        nu_xz = poisson_ratio(C, directions[0], directions[2])
+        nu_yz = poisson_ratio(C, directions[1], directions[2])
 
+        nu_prime = (nu_yx + (nu_yz*nu_zx))/(nu_yz + (nu_yx*nu_xz))
+        print('nu_prime',nu_prime)
+        self.E_effective = self.E/(1-(nu_zy + nu_xy*nu_prime)*((nu_yz)/(1-nu_prime*nu_xz)))
+        print('E',self.E)
+        print('E effective',self.E_effective)
+        self.nu = nu_yx
+        self.nu_xy = nu_xy
+        self.nu_yx = nu_yx
 
+    def measure_energy_strain_relation(self,resolution=100):
+        supercell = self.lattice(size=[1,1,1],symbol=self.el,latticeconstant=self.a0,pbc=(1,1,1),directions=self.directions)
+        supercell.calc = self.calc
+        strains = np.linspace(0,0.3,resolution)
+        stresses = []
+        running_strains = []
+        energies_per_unit_vol = []
+        for strain in strains:
+            new_cell = supercell.copy()
+            new_cell.set_cell(new_cell.get_cell()*[1,1+strain,1],scale_atoms=True)
+            new_cell.calc = self.calc
+            #now relax
+            opt = LBFGS(new_cell)
+            opt.run(fmax=1e-3)
+            stresses.append(new_cell.get_stress()[1])
+            running_strains.append(strain)
+            #get energy per unit volume
+            energies_per_unit_vol.append(np.trapz(stresses,running_strains))
+        
+        #plot stresses and strains
+        plt.plot(running_strains,stresses)
+        plt.xlabel('strain')
+        plt.ylabel('stress')
+        plt.savefig('stress_strain_relation.png')
+        #now integrate to get energy per unit volume
+        print('energies per unit vol',energies_per_unit_vol)
+        energy_strain_relation = np.zeros([len(strains),2])
+        energy_strain_relation[:,0] = strains
+        energy_strain_relation[:,1] = energies_per_unit_vol
+        self.energy_strain_relation = energy_strain_relation
 
-    def K_to_strain(self,K,strip_height):
+    def interpolate_energy_strain_relation(self,energy):
+        #interpolate the energy strain relation to find the strain corresponding to a given energy per unit vol
+        return np.interp(energy,self.energy_strain_relation[:,1],self.energy_strain_relation[:,0])
+    
+
+    def K_to_strain(self,K,strip_height,approximate=False):
         """
[…]
