### 2024-01-04 4fe2a8c
DOC: Added information on restrictions to editable installs
