### 2024-01-04 e98e300
DOC: Added CI badge
