### 2024-01-04 c4cbd0a
DOC: Fixed CI badge for RST
