### 2024-01-04 ea8ab17
MAINT: Replace deprecated distutils.version
