+++ pyproject.toml
@@ -21,6 +21,7 @@ dependencies = [
     "numpy>=1.16.0",
     "scipy>=1.2.3",
     "ase>=3.16.0",
+    "looseversion"
 ]
 
 [project.optional-dependencies]
