### 2024-01-04 cf9041c
CI: Wheels for Python 3.12
