+++ .github/workflows/wheels.yml
@@ -28,7 +28,7 @@ jobs:
         - [macos-12, macosx, x86_64]
         - [windows-2019, win, AMD64]
 
-        python: ["cp38", "cp39", "cp310", "cp311"]
+        python: ["cp38", "cp39", "cp310", "cp311", "cp312"]
       fail-fast: false
 
     env:
