### 2024-01-04 24c9a23
TST: Use absolute path names for example data files
