### 2024-01-04 0db4a03
CI: Updated cibuildwheel version; added install of atomman
