+++ .github/workflows/tests.yml
@@ -23,6 +23,8 @@ jobs:
           sudo apt-get install -y python3-dev python3-pip python3-venv libxml2-dev libxslt-dev zlib1g-dev
           # Upgrade to latest meson and ninja
           sudo pip install --upgrade meson ninja
+          # Install atomman so tests that depend on atomman run
+          sudo pip install atomman
 
       - name: build_c
         run: |
