### 2024-01-04 c883fbc
TST: More absolute paths
