### 2024-01-04 12cae8c
DOC: Address JOSS reviewer comments, #183
