+++ paper/paper.bib
@@ -377,8 +377,7 @@ abstract = {Calculations of dislocation-defect interactions are essential to mod
 }
 
 @article{Seidl2021,
-  title = {Molecular {{Simulations}} of {{Electrotunable Lubrication}}: {{Viscosity}} and {{Wall Slip}} in {{Aqueous Electrolytes}}},
-  shorttitle = {Molecular {{Simulations}} of {{Electrotunable Lubrication}}},
+  title = {Molecular Simulations of Electrotunable Lubrication: Viscosity and Wall Slip in Aqueous Electrolytes},
   author = {Seidl, Christian and H{\"o}rmann, Johannes L. and Pastewka, Lars},
   year = {2021},
   month = jan,
@@ -946,33 +945,6 @@ url={https://openreview.net/forum?id=YPpSngE-ZU}
   doi     = "10.1038/nmat2902"
 }
 
-@ARTICLE{Seidl2021-vn,
-  title    = "Molecular Simulations of Electrotunable Lubrication: Viscosity
-              and Wall Slip in Aqueous Electrolytes",
-  author   = "Seidl, Christian and H{\"o}rmann, Johannes L and Pastewka, Lars",
-  abstract = "We study the frictional response of water-lubricated gold
-              electrodes subject to an electrostatic potential difference using
-              molecular dynamics simulations. Contrary to previous studies on
-              electrotunable lubrication that were carried out by fixing the
-              charges, our simulations keep electrodes at fixed electrostatic
-              potential using a variable charge method. For pure water and NaCl
-              solutions, viscosity is independent of the polarization of the
-              electrodes, but wall slip depends on the potential difference.
-              Our findings are in agreement with previous analytical theories
-              of how wall slip is affected by interatomic interactions. The
-              simulations shed light on the role of electrode polarization for
-              wall slip and illustrate a mechanism for controlling friction and
-              nanoscale flow in simple aqueous lubricants.",
-  journal  = "Tribol. Lett.",
-  volume   =  69,
-  number   =  1,
-  pages    = "22",
-  month    =  jan,
-  year     =  2021,
-  language = "en",
-  doi      = "10.1007/s11249-020-01395-6"
-}
-
 @ARTICLE{Hormann2023-ml,
   title    = "Molecular simulations of sliding on {SDS} surfactant films",
   author   = "H{\"o}rmann, Johannes L and Liu, Chenxu and Meng, Yonggang and
@@ -1180,4 +1152,32 @@ url={https://openreview.net/forum?id=YPpSngE-ZU}
   publisher = {American Physical Society},
   doi = {10.1103/PhysRevB.29.6443},
   url = {https://link.aps.org/doi/10.1103/PhysRevB.29.6443}
+}
+
+@article{preCICEv2,
+  author = {Chourdakis, G and Davis, K and Rodenberg, B and Schulte, M and Simonis, F and Uekermann, B and Abrams, G and Bungartz, HJ and Cheung Yau, L and Desai, I and Eder, K and Hertrich, R and Lindner, F and Rusch, A and Sashko, D and Schneider, D and Totounferoush, A and Volland, D and Vollmer, P and Koseomur, OZ},
+  title = {{preCICE} v2: A sustainable and user-friendly coupling library [version 2; peer review: 2 approved]
+  },
+  journal = {Open Research Europe},
+  volume = {2},
+  year = {2022},
+  number = {51},
+  doi = {10.12688/openreseurope.14445.2},
+  url = {https://doi.org/10.12688/openreseurope.14445.2}
+}
+
+@InProceedings{Dehning2014,
+author="Dehning, Carsten
+and Bierwisch, Claas
+and Kraft, Torsten",
+editor="Griebel, Michael
+and Schweitzer, Marc Alexander",
[…]
