+++ paper/paper.md
@@ -103,12 +103,13 @@ The Python package `matscipy` contains a set of tools for researchers using atom
 
 We never see individual atoms at macroscopic scales. To understand the behaviour of everyday objects, atomic-scale information needs to be transferred to the continuum scale. This is the primary objective of multi-scale modelling. `matscipy` focuses on atomic representations of materials, but implements tools for connecting to continuum descriptions in mechanics and transport theory. Each of the application domains described in the following therefore relies on the computation of continuum fields, that is realized through analytic or numerical solutions.
 
-There is no other package that we are aware of, which fills the particular niche of the application domains in the next section. The package addresses the boundary between atomic-scale and continuum modelling in materials with particular emphasis on plasticity, fracture and tribology. The `atomman` atomistic manipulation toolkit [@AtomMan] and the `atomsk` package [@Hirel2015-ts] have some overlapping objectives but are restricted to a narrower class of materials systems, principally point defects, stacking faults and dislocations. We target interoperability with the widely used Atomic Simulation Environment (ASE) [@Larsen2017], which offers great flexibility by providing a Python interface to tens of simulation codes implementing different physical models. While ASE's policy is to remain pure Python, we augment some of ASE's functionality with more efficient implementations in C, such as the computation of the neighbour list. 
-Large scale molecular dynamics (MD) simulations are most efficiently performed with optimized codes such as LAMMPS [@Thompson2022], with `matscipy`'s main contribution being to setup input structures and to post-process output trajectories.
+There is no other package that we are aware of, which fills the particular niche of the application domains in the next section. The package addresses the boundary between atomic-scale and continuum modelling in materials with particular emphasis on plasticity, fracture and tribology. The `atomman` atomistic manipulation toolkit [@AtomMan] and the `atomsk` package [@Hirel2015-ts] have some overlapping objectives but are restricted to a narrower class of materials systems, principally point defects, stacking faults and dislocations. We target interoperability with the widely used Atomic Simulation Environment (ASE) [@Larsen2017], which offers great flexibility by providing a Python interface to tens of simulation codes implementing different physical models. While ASE's policy is to remain pure Python, we augment some of ASE's functionality with more efficient implementations in C, such as the computation of the neighbour list. Large scale molecular dynamics (MD) simulations are most efficiently performed with optimized codes such as LAMMPS [@Thompson2022], with `matscipy`'s main contribution being to setup input structures and to post-process output trajectories.
 
 The central class in ASE is the `Atoms` class, which is a container that stores atomic positions, velocities and other properties. `Calculator`s describe relationships between atoms, and are used for example to compute energies and forces, and can be attached to `Atoms` objects. All other `matscipy` functionality is implemented as functions acting on `Atoms` objects. This is comparable to the design of `numpy` [@Harris2020-it] or `scipy` [@Virtanen2020-tq], that are collections of mathematical functions operating on core `array` container objects. In our experience, separating code into functions and containers lowers the barrier to entry for new users and eases testability of the underyling code base.
 
-We point out that other generic multi-scale coupling packages exist. Examples of open-source codes are `libmultiscale` [@Anciaux2007;@Anciaux2018], `MultiBench` [@Miller2009] and Green's function molecular dynamics (`GFMD`) [@Campana2006;@Pastewka2012]. These packages target two-way coupling between atomic-scale and continuum formulations, while `matscipy` focuses on constructing atomic domains from continuum information and extracting continuum field from atomic structures. `matscipy` has only limited capabilities for two-way coupling, which makes it significantly easier to use and reduces the barrier to engaging in multi-scale modelling of materials.
[…]
