### 2024-01-04 libAtoms/matscipy#183 @pastewka
Dear @mbarzegary - we have addressed your comments, see below:

> * As mentioned in the paper, there exists other packages for multi-scale coupling such as LibMultiScale and MultiBench, but the necessity of the proposed package is not well elaborated. Since it’s the most important paragraph of the statement of need, I appreciate it if the authors elaborate on this.

I have updated that paragraph. Essentially, most coupling tools deal with mesh interpolation and not with construction of atomic domains. I hope that this is now clearer from the description. I've also added preCICE and MpCCI as examples for generic multi-scale coupling tools in addition to the atomistic-continuum tools mentioned already in the prior iteration of the paper.

> * I see duplicate references. For example, there are two identical Seidel et al. (2021a and 2021b). Can you please check the paper for this?

This was a good catch. We fixed this.

> * The authors have referred to 2 papers using the coupling of the numerical models of the Poisson–Nernst–Planck equation in electrochemistry (Hörmann et al., 2023; Seidl et al. 2021b), but it doesn't seem that those papers use this coupling. Can you please elaborate on the mentioned usage?

This is correct, they don't use the coupling. The paper were just quoted as example of discrete (atomic-scale) effects that need to be captured. We agree that this is confusing and have removed those citations.
