+++ tests/test_bop.py
@@ -144,9 +144,9 @@ def test_crystal_elastic_constants(a0, par):
     TersoffBrenner(tersoff_brenner.Tersoff_PRB_39_5566_Si_C),
     StillingerWeber(stillinger_weber.Stillinger_Weber_PRB_31_5262_Si)
 ])
-def test_amorphous(par):
+def test_amorphous(par, datafile_directory):
     # Tests for amorphous Si
-    aSi = ase.io.read(f'{os.path.dirname(__file__)}/aSi_N8.xyz')
+    aSi = ase.io.read(f'{datafile_directory}/aSi_N8.xyz')
     aSi.calc = Manybody(**par)
     # Non-zero forces and Hessian
     compute_forces_and_hessian(aSi, par)
