### 2024-01-04 libAtoms/matscipy#183 @pastewka
PS: The changes are in the still open PR #208 and not yet in the PDF on the JOSS review page.
