### 2024-01-04 162d3cb
TST: Added pytest fixture for data file directory
