### 2024-01-04 0ae0535
TST: Added atomman and ovito as a test dependency (to pyproject.toml)
