+++ pyproject.toml
@@ -27,17 +27,19 @@ dependencies = [
 [project.optional-dependencies]
 test = [
     "pytest",
-    "sympy"
+    "sympy",
+    "atomman",
+    "ovito"
 ]
 docs = [
     "sphinx",
-    "sphinx-rtd-theme",
     "myst_nb",
     "nglview",
     "numpydoc",
     "atomman",
     "ovito",
-    "sphinx_copybutton"
+    "sphinx_copybutton",
+    "pydata-sphinx-theme"
 ]
 
 [project.urls]
