### 2024-01-04 158a14c
MAINT: Removed distributed computation as this is essentially untested and unused
