### 2024-01-04 libAtoms/matscipy#185 @pastewka
Dear @mbarzegary - see comments below:

> * I cannot find the FEniCS functionality being tested via the pytest tests. Is it because of the complicated installation procedure? I care about this functionality not only due to its usefulness but also because it's mentioned in the paper.

FEniCS is not testes via pytest because of the complicated installation procedure. However, we now added a test to github actions that essentially runs the FEniCS examples that are provided for testing (@jotelha).

> * All tests pass on my local machine, but some of them produce unclear characters. This is the output of the pytest run:

Some tests were skipped ('s'), likely because `atomman` and `ovito` was not installed. We now added those packages to the test dependencies.

Note that some changes are in PR #208 and not yet in master.
