### 2024-01-04 libAtoms/matscipy#186 @pastewka
@jotelha Can you please add documentation on the electrochemistry module? Maybe we can turn part of the example into a documentation of the module. Please also document the FEniCS coupling there.
