### 2024-01-04 0dc8b3e
DOC: Added angle distribution documentation; added formulas for distribution function; remove pair distribution example (which appears redundant to the documentation notebook)
