### 2024-01-04 639a36f
DOC: Removed explicit mention of opls module name
