### 2024-01-08 61af193
ENH: Convert solver switching to string param from bool
