### 2024-01-09 d23af14
DOC: docstring for electrochemistry module
