### 2024-01-09 319e089
MAINT: installable command line interface
