### 2024-01-09 e83b042
DOC: created CLI documentation stub
