+++ docs/index.rst
@@ -9,6 +9,7 @@
    ./tools/index
    ./calculators/index
    ./topology/index
+   ./cli
    ./api
 
 
