+++ matscipy/elasticity.py
@@ -209,6 +209,57 @@ def cubic_to_Voigt_6x6(C11, C12, C44):
                      [  0,  0,  0,  0,C44,  0],
                      [  0,  0,  0,  0,  0,C44]])
 
+def coalesce_elastic_constants(C11=None, C12=None, C44=None, C=None, standard="Cij"):
+    '''
+    Convert either all of C11, C12, C44, or 6x6 C, or 3x3x3x3 C to the form defined by standard.
+
+    standard="cubic": returns C11, C12, C44
+    standard="Cij": returns 6x6 C matrix
+    standard="Cijkl": returns 3x3x3x3 C tensor
+    
+    '''
+    # Aim to convert to 3x3x3x3 first, as this contains the maximal information
+    Cijkl = None
+
+    if C is not None:
+        # Prefer to use full matrix, as this contains more info
+        C = np.array(C)
+        if C.shape == (6, 6):
+            Cijkl = Voigt_6x6_to_full_3x3x3x3(C)
+
+
+        elif C.shape == (3, 3, 3, 3):
+            Cijkl = C.copy()
+        else:
+            # C not of the correct shape
+            warnings.warn("C was not the correct shape. Expected shape (6, 6) or (3, 3, 3, 3)")
+        
+        if Cijkl is not None:
+            # Check if C11, C12, or C44 have also been passed
+            # Ignore these inputs, but warn user
+            has_other_elast = [var is not None for var in [C11, C12, C44]]
+
+            if any(has_other_elast):
+                for i, var in enumerate(["C11", "C12", "C44"]):
+                    if has_other_elast[i]:
+                        warnings.warn(f"Ignoring argument {var} as C was passed")
+
+    if Cijkl is None:
+        # Didn't get correct info from C, look at C11, C12, C44
+        if C11 is not None and C12 is not None and C44 is not None:
+            Cijkl = Voigt_6x6_to_full_3x3x3x3(cubic_to_Voigt_6x6(C11, C12, C44))
+        else:
+            raise RuntimeError("Elastic Constants not correctly defined. "+
+                                "Pass either C, or all of C11, C12, C44.")
+
+    if standard.lower() == "cijkl":
+        # 3x3x3x3 standard
+        return Cijkl
+    elif standard.lower() == "cij":
+        return full_3x3x3x3_to_Voigt_6x6(Cijkl)
+    elif standard.lower() == "cubic":
+        return Voigt_6x6_to_cubic(full_3x3x3x3_to_Voigt_6x6(Cijkl))
+
 def _invariants(s, syy=None, szz=None, syz=None, sxz=None, sxy=None,
                full_3x3_to_Voigt_6=full_3x3_to_Voigt_6_stress):
     """
