### 2024-01-10 da981a4
MAINT: Rename method inits to match method name
