### 2024-01-10 ef5c8ab
DOC: Better docs of _build_supercell
