### 2024-01-10 4fd5237
MAINT: Remove confusing ext_idx arg
