+++ matscipy/dislocation.py
@@ -2588,7 +2588,7 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
         return sup
 
     def _build_bulk_cyl(self, radius, core_positions, fix_rad, extension,
-                        self_consistent, method, ext_idx):
+                        self_consistent, method):
         '''
         Build bulk cylinder config from args supplied by self.build_cylinder
 
@@ -2603,16 +2603,11 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
             e.g. for glide configs, etc
         fix_rad: float
             Radius for fixed atoms constraints
-        extension: int
-            Ensure the cell extends out from the center to center + extension
-        ext_idx: int or np.array of ints
-            idxs of the cores each part of extension should apply to
-            extension = np.array([
-                            [-5.0, 0, 0],
-                            [5.0, 0, 0.0]
-                            ])
-            and ext_idxs = np.array([0, -1]) will extend the cell by 5Ang left of the first core, 
-            and 5Ang right of the last core
+        extension: np.array
+            (N, 3) array of vector extensions from each core.
+            Will add a fictitious core at position 
+            core_positions[i, :] + extension[i, :], for the 
+            purposes of adding extra atoms
 
         Returns
         -------
@@ -2631,9 +2626,20 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
 
         if len(extension.shape) == 1:
             extension = extension[np.newaxis, :]
-        ext_idx = np.array(ext_idx)
 
-        exts = extension + core_positions[ext_idx, :]
+        if np.all(extension.shape != core_positions.shape):            
+            if extension.shape[0] < core_positions.shape[0] and \
+                extension.shape[1] == core_positions.shape[1]:
+                # Too few extensions passed, assume that the extensions are for the first N cores
+                # Pad extension with zeros
+                extension = np.vstack([extension, np.zeros((
+                    core_positions.shape[0] - extension.shape[0], core_positions.shape[1]
+                ))])
+            else:
+                raise ValueError(f"{extension.shape=} does not match {core_positions.shape=}")
+
+        # Only care about non-zero extensions
+        exts = extension + core_positions
 
         xmax = np.max([np.max(core_positions[:, 0]), np.max(exts[:, 0])])
         xmin = np.min([np.min(core_positions[:, 0]), np.min(exts[:, 0])])
@@ -2660,6 +2666,11 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
         new_core_positions = core_positions.copy()
 
         new_core_positions += shift
+
+        # Mask out extensions that are all zeros
+        nonzero_exts = np.sum(np.abs(extension), axis=-1) > 0.0
+
+        exts = exts[nonzero_exts, :]
 
         mask_positions = np.vstack([new_core_positions, exts + shift])
 
@@ -2829,8 +2840,7 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
                        core_position=np.array([0., 0., 0.]),
                        extension=np.array([0, 0, 0]),
                        fix_width=10.0, self_consistent=None,
-                       method="atomman",
-                       ext_idxs = -1):
+                       method="atomman"):
         '''
         Build dislocation cylinder for single dislocation system
 
@@ -2842,7 +2852,7 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
             Vector offset of the core position from default site
         extension: np.array
             Add extra bulk to the system, to also surround a fictitious core
-            that is extension away from the center of the cell
+            that is at position core_position + extension
         fix_width: float
             Defines a region to apply the FixAtoms ase constraint to
             Fixed region is given by (radius - fix_width) <= r <= radius,
[…]
