### 2024-01-10 81967ff
MAINT: Rename 'standard' -> 'convention'
