### 2024-01-10 6b604bb
BUG: Remove old init_stroh call
