### 2024-01-10 ccce7e3
DOC: add electrochemistry jupyter notebook in form of ipynb and jupytext-generated percent-.py file
