+++ docs/applications/electrochemistry_1.py
@@ -1,0 +1,572 @@
+# %% [markdown]
+# # From continuous double layer theory to discrete coordinate sets
+#
+# ## Poisson-Nernst-Planck systems
+
+# %% [markdown]
+# The `matscipy.electrochemistry.PoissonNernstPlanckSystem` solver provides an interface for solving classical, one-dimensional electrochemical systems with an arbitrary number of $N$ ionic species by means of a controlled-volumes method.
+
+# %% [markdown]
+# ### The inert electrode
+
+# %% [markdown]
+# An archetypical system is the inert electrode at the open half-space governed by the Poisson equation and $N$ Nernst-Planck equations with Dirichlet and Neumann boundary conditions as shown in Figure 1 below.
+
+# %% [markdown]
+# <figure>
+# <img src="electrochemistry/inertElectrode.svg" style="width:65%">
+# <figcaption align="center">Figure 1: Inert electrode at the open half-space</figcaption>
+# </figure>
+#
+
+# %% [markdown]
+# Here, $D_i$ is the diffusivity of the *i*th species, $c_i$ its local concentration, $c_i^\infty$ its bulk concentration, $u_i$ its electric mobility, and $z_i$ its charge number. $\phi$ is the local electrostatic potential, and $\phi_0$ the electrode's potential compared to the bulk solution. $\epsilon_0$ is the vacuum permittivity, $\epsilon_r$ is the relative permittivity of the solvent, and $F$ is the Faraday constant.
+
+# %% [markdown]
+# In the following, we use `matscipy.electrochemistry.PoissonNernstPlanckSystem` to solve this system for $0.1\,\mathrm{mM}~\text{NaCl}$ at $0.05\,\mathrm{V}$ bias. Parameters are provided in SI units. We will compare the results against the analytical solution, which exists for this special case of a binary electrolyte.
+
+# %% [markdown]
+# #### Preparations
+
+# %% [markdown]
+# First, we import a few basic tools.
+
+# %%
+import logging
+import numpy as np
+import scipy.constants as sc
+import matplotlib.pyplot as plt
+
+# %% [markdown]
+# Then, we import the Poisson-Nernst-Planck solver.
+
+# %%
+from matscipy.electrochemistry import PoissonNernstPlanckSystem
+
+# %% [markdown]
+# `PoissonNernstPlanckSystem` makes extensive use of Python's logging module. We reconfigure to format and show logs as desired.
+#
+
+# %%
+# configure logging: verbosity level and format as desired
+standard_loglevel   = logging.INFO
+standard_logformat  = "[ %(filename)s:%(lineno)s - %(funcName)s() ]: %(message)s"
+
+# reset logger if previously loaded
+logging.shutdown()
+logging.basicConfig(level=standard_loglevel,
+                    format=standard_logformat,
+                    datefmt='%m-%d %H:%M')
+
+# in Jupyter notebooks, explicitly modifying the root logger necessary
+logger = logging.getLogger()
+logger.setLevel(standard_loglevel)
+
+# remove all handlers
+for h in logger.handlers: logger.removeHandler(h)
+
+# create and append custom handles
+ch = logging.StreamHandler()
+formatter = logging.Formatter(standard_logformat)
+ch.setFormatter(formatter)
+ch.setLevel(standard_loglevel)
[…]
