+++ docs/applications/electrochemistry.rst
@@ -1,0 +1,9 @@
+Electrochemistry
+================
+
+The electrochemistry module provides three components for preparing atomistic
+configurations usually encountered in the context of soluted electrolytes.
+
+.. toctree::
+
+    electrochemistry_1
