### 2024-01-10 dc0bfd3
DOC: electrochemistry module documentation draft for pnp solver and sampling
