### 2024-01-11 263be09
DOC: enable use of jupytext in mystnb and store electrochemistry_1 notebook in pure text format only
