### 2024-01-11 7a89f27
DOC: show view of sampled atoms in electrochemistry nb
