+++ docs/applications/electrochemistry_1.py
@@ -502,7 +502,7 @@ ax.set_axis_off()
 # # Acknowledgements
 
 # %% [markdown]
-# We thank Lukas Elflein for supporting the development of `matscipy.electrochemistry` and the according example notebooks, Christian Seidl for early experimentation with the module's ffunctionality, and Andreas Greiner for discussions and inspirations on solving transport problems with controlled-volume and finte-elements methods.
+# We thank Lukas Elflein for supporting the development of `matscipy.electrochemistry` and the according example notebooks, Christian Seidl for early experimentation with the module's functionality, Andreas Greiner for discussions and inspirations on solving transport problems with controlled-volume and finte-elements methods, as well as Meng, Yonggang and Zhang, Jun for posing an interesting tribological problem that sparked the development of this module initially.
 
 # %% [markdown]
 # # References
