### 2024-01-11 7505775
DOC: documentation on steric correction
