+++ docs/applications/electrochemistry_3.py
@@ -1,0 +1,307 @@
+# %% [markdown]
+# # Steric correction
+#
+# Continuum models do not account for the finite size of atoms. Sampling discrete particle coordinates from continuous concentration distributions may thus result in arbitrarily close atom positions. To avoid overlap and yield *usable* atomistic configurations from densily packed samples, `matscipy.electrochemistry` offers the `steric_correction` submodule. Its function `apply_steric_correction` applies a steric correction to the provided configuration, assuring the desired steric radii for all species if possible. To achieve this, the module reimplements a pseudo-potential minimization algorithm [](https://doi.org/10.1002/jcc.21224) used by [PACKMOL](https://m3g.github.io/packmol/) in Python 
+
+# %% [markdown]
+# Below, we generate continuous concentration distributions from classical electrochemical double layer theory, sample coordinates from thesedistributions, and apply a steric correction to the ion coordinates.
+
+# %% [markdown]
+# Theory and mechanisms involved in step 1 and 2 have been discussed in detail in previous documentation sections.
+
+# %% [markdown]
+# ## Step 1: generating continuous concentration distributions
+#
+
+# %% [markdown]
+# Adjacent to an inert electrode at an open half-space, concentration distributions of a binary electrolyte are retrieved from the analytical solution of the Poisson-Boltzmann equation. `matscipy.electrochemistry.poisson_boltzmann_distribution` provides a simple interface to this analytical solution. Let's consider 1 mM saline solution.
+
+# %%
+from matscipy.electrochemistry.poisson_boltzmann_distribution import (
+    potential, concentration, charge_density)
+
+# %% [markdown]
+# Although we look at a continuous 1d system now, we want to sample coordinates in a 3d box from it later. Hence, we specify the 3d box size here already and retrieve potential, concentration and charge distributions arising from solving the Poisson-Bolzmann equation analytically.
+
+# %%
+import numpy as np 
+
+# measures of box
+xsize = ysize = 5e-9 # nm, SI units
+zsize = 20e-9        # nm, SI units
+
+x = np.linspace(0, zsize, 2000)
+c = [1, 1]   # bulk concentrations of Na and Cl, mM, SI units
+z = [1, -1]  # number charges of species
+u = 0.05     # electrostatic potential across the interface, V
+
+phi = potential(x, c, z, u)
+C   = concentration(x, c, z, u)
+rho = charge_density(x, c, z, u)
+
+# %% [markdown]
+# Next, we visualize these distributions.
+
+# %%
+import matplotlib.pyplot as plt
+import scipy.constants as sc    # fundamental constants
+
+from matscipy.electrochemistry import debye
+
+def make_patch_spines_invisible(ax):
+    ax.set_frame_on(True)
+    ax.patch.set_visible(False)
+    for sp in ax.spines.values():
+        sp.set_visible(False)
+        
+
+deb = debye(c, z) 
+
+fig, ax1 = plt.subplots(figsize=[18,5])
+ax1.set_xlabel('distance x (nm)')
+ax1.plot(x/sc.nano, phi, marker='', color='red', label='potential', linewidth=1, linestyle='--')
+ax1.set_ylabel('potential $\phi$ (V)')
+ax1.axvline(x=deb/sc.nano, label='Debye Length', color='orange')
+
+ax2 = ax1.twinx()
[…]
