### 2024-01-11 c72250d
DOC: CLI docs layout optimized
