### 2024-01-11 1c274cf
DOC: reshuffled structure of electrochem docs a little
