### 2024-01-11 fed3837
DOC: corrected cli docs link
