### 2024-01-11 d73a7bd
MAINT: add jupytext to docs requirements for rendering py-percent plain text notebooks
