+++ docs/requirements.txt
@@ -1,3 +1,4 @@
 numpy>=1.16.0
 ase>=3.16.0
 sphinx_rtd_theme
+jupytext
