### 2024-01-11 a9abbb5
DOC: added fenics integration docs
