+++ docs/applications/electrochemistry_2.py
@@ -1,0 +1,295 @@
+# %% [markdown]
+# # Poisson-Nernst-Planck systems by finite element method
+
+# %% [markdown]
+# `PoissonNernstPlanckSystem` of `matscipy.electrochemistry.poisson_nernst_planck_solver` uses its own controlled-volumes solver implementation, while `PoissonNernstPlanckSystemFEniCS` of `matscipy.electrochemistry.poisson_nernst_planck_solver_fenics` interfaces the finite elements solver [FEniCS](https://fenicsproject.org/) for solving Poisson-Nernst-Planck systems. In highly nonlinear problems, latter may offer improved convergence above the former. Following examples illustrate this.
+
+# %% [markdown]
+# Again, we look at the inert electrode shown here.
+
+# %% [markdown]
+# <a id="figure1"></a><figure>
+# ![Figure 1](electrochemistry/inertElectrode.svg)
+#
+# *Figure 1*: Inert electrode at the open half-space
+
+# %% [markdown]
+# As usual, we begin with preparing a few necessities.
+
+# %%
+# basics
+import numpy as np
+import scipy.constants as sc
+import matplotlib.pyplot as plt
+
+# electrochemistry basics
+from matscipy.electrochemistry import debye, ionic_strength
+
+# Poisson-Bolzmann distribution
+from matscipy.electrochemistry.poisson_boltzmann_distribution import gamma, potential, concentration, charge_density
+
+# Poisson-Nernst-Planck solver
+from matscipy.electrochemistry import PoissonNernstPlanckSystem
+from matscipy.electrochemistry.poisson_nernst_planck_solver_fenics import PoissonNernstPlanckSystemFEniCS
+
+# 3rd party file output
+import ase
+import ase.io
+
+def make_patch_spines_invisible(ax):
+    ax.set_frame_on(True)
+    ax.patch.set_visible(False)
+    for sp in ax.spines.values():
+        sp.set_visible(False)
+
+pnp = {}
+
+# %% [markdown]
+# ## Agreement of solvers for moderate boundary conditions
+
+# %% [markdown]
+# At moderate potentials across the interface, both `PoissonNernstPlanckSystem` and `PoissonNernstPlanckSystemFEniCS` both converge quickly.
+
+# %%
+# Test case parameters
+c = [1.0, 1.0]
+z = [1, -1]
+L = 100e-9 # 100 nm
+delta_u = 0.05 # V
+N = 200 # number of discretization grid points
+
+# define desired system
+pnp['std_interface'] = PoissonNernstPlanckSystem(
+    c, z, L, delta_u=delta_u, N=N,
+    solver="hybr", options={'xtol':1e-12})
+
+pnp['std_interface'].useStandardInterfaceBC()
+uij, nij, lamj = pnp['std_interface'].solve()
+
+# define desired system
+pnp['fenics_interface'] = PoissonNernstPlanckSystemFEniCS(c, z, L, delta_u=delta_u, N=N)
+pnp['fenics_interface'].useStandardInterfaceBC()
+uij, nij, _ = pnp['fenics_interface'].solve()
+
+# %% [markdown]
+# The results fall onto each other and onto the analytical Poisson-Boltzmann solution, validating each other.
+
+# %%
+x = np.linspace(0,L,100)
+phi = potential(x, c, z, delta_u) 
+C =   concentration(x, c, z, delta_u)
+rho = charge_density(x, c, z, delta_u) 
+deb = debye(c, z)
+
[…]
