+++ docs/applications/electrochemistry.rst
@@ -7,9 +7,16 @@ configurations usually encountered in the context of soluted electrolytes.
 .. toctree::
 
     electrochemistry_1
+    electrochemistry_2
     electrochemistry_3
 
-Acknowledgments
----------------
+Acknowledgements
+----------------
 
-We thank Lukas Elflein for supporting the development of `matscipy.electrochemistry` and the according example notebooks, Christian Seidl for early experimentation with the module's functionality, Andreas Greiner for discussions and inspirations on solving transport problems with controlled-volume and finte-elements methods, as well as Meng, Yonggang and Zhang, Jun for posing an interesting tribological problem that sparked the development of this module initially.
+We thank Lukas Elflein for supporting the development of
+`matscipy.electrochemistry` and the according example notebooks, Christian
+Seidl for early experimentation with the module's functionality, Andreas
+Greiner for discussions and inspirations on solving transport problems with
+controlled-volume and finte-elements methods, as well as Meng Yonggang and
+Zhang Jun for posing an interesting tribological problem that sparked the
+development of this module initially.
