### 2024-01-11 9456d8c
CI: provide fenics in CI for building fenics-related docs
