### 2024-01-11 libAtoms/matscipy#186 @jotelha
@pastewka I think I am done with the electrochemistry and CLI documentation pages on https://github.com/libAtoms/matscipy/pull/210 . Generating the docs from jupyter notebooks works very neatly!
