### 2024-01-12 d0dd7ea
DOC: allow generating spelling reports with 'make spelling'
