### 2024-01-12 b32deff
DOC: corrected typos in electrochemistry docs
