+++ tests/test_poisson_nernst_planck_solver.py
@@ -24,6 +24,7 @@ import unittest
 
 from matscipy.electrochemistry import PoissonNernstPlanckSystem
 
+
 class PoissonNernstPlanckSolverTest(matscipytest.MatSciPyTestCase):
 
     def setUp(self):
@@ -44,6 +45,7 @@ class PoissonNernstPlanckSolverTest(matscipytest.MatSciPyTestCase):
         self.assertArrayAlmostEqual(pnp.grid, self.ref_data ['x'])
         self.assertArrayAlmostEqual(pnp.potential, self.ref_data ['u'])
         self.assertArrayAlmostEqual(pnp.concentration, self.ref_data ['c'], 1e-6)
+
 
 if __name__ == '__main__':
     unittest.main()
