### 2024-01-12 2cf4fd7
TST: skip test_poisson_nernst_planck_solver_fenics if fenics not available
