### 2024-01-12 libAtoms/matscipy#185 @jotelha
Fenics installation is indeed not of negligible effort, in particular since the fenics integration only plays a role in one of several sub-modules. I have added a test on fenics integration with https://github.com/libAtoms/matscipy/pull/213. The test is skipped if fenics is not available.
