### 2024-01-12 88216d7
DOC: remove all obsolete jupyter notebooks from electrochemistry examples folder
