+++ examples/electrochemistry/samples_pnp_c2d.py
@@ -1,1381 +1,0 @@
-#
-# Copyright 2019-2020 Johannes Hoermann (U. Freiburg)
-#
-# matscipy - Materials science with Python at the atomic-scale
-# https://github.com/libAtoms/matscipy
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 2 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-
-# %% [markdown]
-# # Poisson-Nernst-Planck systems & continuous2discrete 
-#
-# *Johannes Hörmann, Lukas Elflein, 2019*
-#
-# from continuous electrochemical double layer theory to discrete coordinate sets
-
-# %%
-# for dynamic module reload during testing, code modifications take immediate effect
-# %load_ext autoreload
-# %autoreload 2
-
-# %%
-# stretching notebook width across whole window
-try:
-    from IPython.core.display import display, HTML
-    display(HTML("<style>.container { width:100% !important; }</style>"))
-except:
-    pass
-
-# %%
-# basics
-import logging
-import numpy as np
-import scipy.constants as sc
-import matplotlib.pyplot as plt
-
-# %%
-# sampling
-from scipy import interpolate
-from matscipy.electrochemistry import continuous2discrete
-from matscipy.electrochemistry import get_histogram
-
-# %%
-# electrochemistry basics
-from matscipy.electrochemistry import debye, ionic_strength
-
-# %%
-# Poisson-Bolzmann distribution
-from matscipy.electrochemistry.poisson_boltzmann_distribution import gamma, potential, concentration, charge_density
-
-# %%
-# Poisson-Nernst-Planck solver
-from matscipy.electrochemistry import PoissonNernstPlanckSystem
-
-# %%
-# 3rd party file output
-import ase
-import ase.io
-
-# %%
-# helper functions
-def get_centers(bins):
-    """Return the center of the provided bins.
-
-    Example:
-    >>> get_centers(bins=np.array([0.0, 1.0, 2.0]))
-    array([ 0.5,  1.5])
-    """
-    bins = bins.astype(float)
-    return (bins[:-1] + bins[1:]) / 2
-
-def plot_dist(histogram, name, reference_distribution=None):
-    """Plot histogram with an optional reference distribution."""
-    hist, bins = histogram
-    width = 1 * (bins[1] - bins[0])
[…]
