### 2024-01-12 ac07aa9
DOC: updated electrochemistry parametric study example bash scripts
