+++ examples/electrochemistry/pcs_pipeline.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# Sample on parameter sweep through the pnp > c2d > stericify pipeline.
+# Sample on parameter sweep through the poisson-nernst-planck > continuous2discrete > stericify pipeline.
 # see sub-directories for detailed understanding.
 
 system=cell_1d
