### 2024-01-12 libAtoms/matscipy#184 @jotelha
https://github.com/libAtoms/matscipy/pull/192 resolved bullet point 4.

Addressing bullet point 3 for the electrochemistry submodule examples: Within https://github.com/libAtoms/matscipy/pull/210, I have turned the electrochemistry example notebooks into proper documentation and consequently removed them from the not very neatly curated `examples/electrochemistry` folder. This folder now only contains sample bash scripts for a simple parametric studies with the electrochemistry CLI.
