### 2024-01-12 2b43690
TST: Add test of scf of sum vs sum of scf
