### 2024-01-12 f97d417
BUG: Bugfix & testing of complete_basis
