### 2024-01-12 a084f8d
TST: Unittests for get_distance_from_polygon2D
