### 2024-01-12 875be87
TST: Unittests for radial_mask_from_polygon2D
