### 2024-01-12 39bb2d2
BUG: Fixed skipping of FENICS tests
