### 2024-01-12 fed34e8
CI: rm should succeed even if file does not exist
