### 2024-01-13 4889b56
TST: fix skippable test forfenics integration
