### 2024-01-13 3fcaad6
MAINT: Camel case to lower case for electrochemistry methods, and other minor adaptions for improved PEP-8 adherence
