+++ matscipy/electrochemistry/poisson_nernst_planck_solver.py
@@ -225,7 +225,6 @@ class PoissonNernstPlanckSystem:
             self.convergenceStepRelative = []
             self.convergenceResidualAbsolute = []
 
-
             dxij = np.zeros(self.N)
 
         self.xij.append(xij)
@@ -257,7 +256,7 @@ class PoissonNernstPlanckSystem:
         """Sets up discretization scheme and initial value"""
         # indices
         self.Ni = self.N+1
-        I = np.arange(self.Ni)
+        index_i = np.arange(self.Ni)
 
         self.logger.info('{:<{lwidth}s} {:> 8.4g}'.format(
           'discretization segments N', self.N, lwidth=self.label_width))
@@ -271,7 +270,7 @@ class PoissonNernstPlanckSystem:
           'dx', self.dx, lwidth=self.label_width))
 
         # positions (scaled)
-        self.X = self.x0_scaled + I*self.dx
+        self.X = self.x0_scaled + index_i * self.dx
 
         # Boundary & initial values
 
@@ -288,9 +287,6 @@ class PoissonNernstPlanckSystem:
             self.ni0 = np.kron(self.c_scaled, np.ones((self.Ni, 1))).T
         if self.zi0 is None:
             self.zi0 = np.kron(self.z, np.ones((self.Ni, 1))).T  # does not change
-
-
-        # self.initial_values()
 
     def initial_values(self):
         """
@@ -379,7 +375,7 @@ class PoissonNernstPlanckSystem:
         return self.uij, self.nij, self.lamj
 
     # standard sets of boundary conditions:
-    def useStandardInterfaceBC(self):
+    def use_standard_interface_bc(self):
         """Interface at left hand side and open bulk at right hand side"""
         self.boundary_conditions = []
 
@@ -392,24 +388,24 @@ class PoissonNernstPlanckSystem:
             'Right hand side Dirichlet boundary condition:                              u1 = {:> 8.4g}'.format(self.u1))
 
         self.boundary_conditions.extend([
-            lambda x: self.leftPotentialDirichletBC(x, self.u0),
-            lambda x: self.rightPotentialDirichletBC(x, self.u1)])
+            lambda x: self.left_potential_dirichlet_bc(x, self.u0),
+            lambda x: self.right_potential_dirichlet_bc(x, self.u1)])
 
         for k in range(self.M):
             self.logger.info(
                 'Ion species {:02d} left hand side concentration Flux boundary condition:       j0 = {:> 8.4g}'.format(
-                    k,0))
+                    k, 0))
             self.logger.info(
                 'Ion species {:02d} right hand side concentration Dirichlet boundary condition: c1 = {:> 8.4g}'.format(
                     k, self.c_scaled[k]))
             self.boundary_conditions.extend([
-                lambda x, k=k: self.leftControlledVolumeSchemeFluxBC(x, k),
-                lambda x, k=k: self.rightDirichletBC(x, k, self.c_scaled[k])])
+                lambda x, k=k: self.left_controlled_volume_scheme_flux_bc(x, k),
+                lambda x, k=k: self.right_dirichlet_bc(x, k, self.c_scaled[k])])
         # counter-intuitive behavior of lambda in loop:
         # https://stackoverflow.com/questions/2295290/what-do-lambda-function-closures-capture
         # workaround: default parameter k=k
 
-    def useStandardCellBC(self):
+    def use_standard_cell_bc(self):
         """Interfaces at left hand side and right hand side"""
         self.boundary_conditions = []
 
[…]
