+++ matscipy/electrochemistry/poisson_boltzmann_distribution.py
@@ -86,8 +86,7 @@ def debye(c, z,
     lambda_D : float
         Debye length, sqrt( epsR*eps*R*T/(2*F^2*I) ) [m]
     """
-    I = ionic_strength(c, z)
-    return np.sqrt(relative_permittivity*vacuum_permittivity*R*T/(2.0*F**2*I))
+    return np.sqrt(relative_permittivity*vacuum_permittivity*R*T/(2.0*F**2*ionic_strength(c, z)))
 
 
 def gamma(u, T=298.15):
