### 2024-01-13 cbd9a65
CI: run tests on pull requests
