+++ .github/workflows/tests.yml
@@ -6,6 +6,8 @@ on:
       - '**'
     tags:
       - '**'
+  pull_request:
+    branches: [ master ]
 
 jobs:
   tests:
