### 2024-01-13 0326294
CI: provide fenics in tests work flow to avoid skipping fenics-dependent tests
