### 2024-01-14 2c64685
TST: print stderr of CLI commands to understand why and where pipeline test might fail in CI
