### 2024-01-14 2c39664
TST: inspect both stdout and stderr of CLI test thorougly
