### 2024-01-14 aca11d8
MAINT: a little bit of trial and error on what will make fenics logs silent in CI
