### 2024-01-14 2a5ecac
MAINT: more trial and error on what will make fenics logs silent
