### 2024-01-14 3036e23
MAINT: desperate trial and error on what will make fenics logs silent
