### 2024-01-14 5aec6e3
MAINT: endless trial and error on what will make fenics logs silent
