### 2024-01-14 c4dde28
WIP: Moved some scripts to CLI and added Meson build files
