### 2024-01-14 2d8102b
MAINT: Moved some script to proper CLI
