+++ matscipy/cli/glasses/quench.py
@@ -67,102 +67,100 @@ def random_solid(els, density, sx=None, sy=None):
 
 ###
 
-# For coordination counting
-cutoff = 1.85
+def main():
+    els = parameter('stoichiometry')
+    densities  = parameter('densities')
 
-els = parameter('stoichiometry')
-densities  = parameter('densities')
+    T1 = parameter('T1', 5000*kB)
+    T2 = parameter('T2', 300*kB)
+    dt1 = parameter('dt1', 0.1*fs)
+    dt2 = parameter('dt2', 0.1*fs)
+    tau1 = parameter('tau1', 5000*fs)
+    tau2 = parameter('tau2', 500*fs)
+    dtdump = parameter('dtdump', 100*fs)
+    teq = parameter('teq', 50e3*fs)
+    tqu = parameter('tqu', 20e3*fs)
+    nsteps_relax = parameter('nsteps_relax', 10000)
 
-T1 = parameter('T1', 5000*kB)
-T2 = parameter('T2', 300*kB)
-dt1 = parameter('dt1', 0.1*fs)
-dt2 = parameter('dt2', 0.1*fs)
-tau1 = parameter('tau1', 5000*fs)
-tau2 = parameter('tau2', 500*fs)
-dtdump = parameter('dtdump', 100*fs)
-teq = parameter('teq', 50e3*fs)
-tqu = parameter('tqu', 20e3*fs)
-nsteps_relax = parameter('nsteps_relax', 10000)
+    ###
 
-###
+    quick_calc = parameter('quick_calc')
+    calc = parameter('calc')
 
-quick_calc = parameter('quick_calc')
-calc = parameter('calc')
+    ###
 
-###
+    for _density in densities:
+        try:
+            density, sx, sy = _density
+        except:
+            density = _density
+            sx = sy = None
+        print('density =', density)
 
-for _density in densities:
-    try:
-        density, sx, sy = _density
-    except:
-        density = _density
-        sx = sy = None
-    print('density =', density)
+        initial_fn = 'density_%2.1f-initial.traj' % density
 
-    initial_fn = 'density_%2.1f-initial.traj' % density
+        liquid_fn = 'density_%2.1f-liquid.traj' % density
+        liquid_final_fn = 'density_%2.1f-liquid.final.traj' % density
 
-    liquid_fn = 'density_%2.1f-liquid.traj' % density
-    liquid_final_fn = 'density_%2.1f-liquid.final.traj' % density
+        quench_fn = 'density_%2.1f-quench.traj' % density
+        quench_final_fn = 'density_%2.1f-quench.final.traj' % density
 
-    quench_fn = 'density_%2.1f-quench.traj' % density
-    quench_final_fn = 'density_%2.1f-quench.final.traj' % density
+        print('=== LIQUID ===')
 
-    print('=== LIQUID ===')
+        if not os.path.exists(liquid_final_fn):
+            if not os.path.exists(liquid_fn):
+                print('... creating new solid ...')
+                a = random_solid(els, density, sx=sx, sy=sy)
+                n = a.get_atomic_numbers().copy()
 
-    if not os.path.exists(liquid_final_fn):
-        if not os.path.exists(liquid_fn):
-            print('... creating new solid ...')
-            a = random_solid(els, density, sx=sx, sy=sy)
-            n = a.get_atomic_numbers().copy()
+                # Relax with the quick potential
+                a.set_atomic_numbers([6]*len(a))
+                a.set_calculator(quick_calc)
+                FIRE(a, downhill_check=True).run(fmax=1.0, steps=nsteps_relax)
+                a.set_atomic_numbers(n)
+                write(initial_fn, a)
+            else:
+                print('... reading %s ...' % liquid_fn)
+                a = read(liquid_fn)
 
[…]
