### 2024-01-14 a21f4d8
DOC: Added short description of commands
