+++ docs/cli/glasses.md
@@ -1,0 +1,5 @@
+# Glasses
+
+## Overview of commands
+
+* `matscipy_quench`: Liquid quench of a glassy structure.
