+++ docs/cli/index.rst
@@ -1,13 +1,12 @@
 Command line interface
 ======================
 
-`matscipy` provides command line interfaces (CLI) for selected functionality.
-These CLI are not installed per default. Install with the `[cli]` extra, e.g.
-
-  pip install matscipy[cli]
-
-to create all provided CLI scripts.
+`matscipy` provides command line interfaces (CLIs) for selected functionality.
+These CLIs are installed per default and are prefixed with `matscipy_`.
 
 .. toctree::
 
+   diffusion
    electrochemistry
+   glasses
+   interatomic
