+++ docs/cli/interatomic.md
@@ -1,0 +1,5 @@
+# Interatomic potentials
+
+## Overview of commands
+
+* `matscipy_average_eam_potential`: Generate averaged EAM potential for an alloy.
