+++ docs/cli/diffusion.md
@@ -1,0 +1,5 @@
+# Diffusion
+
+## Overview of commands
+
+* `matscipy_rms`: Compute the root-mean-square (rms) displacement on a trajectory file.
