### 2024-01-15 02ccb14
DOC: Hyphenated commands and added description on rms command
