+++ docs/cli/index.rst
@@ -2,7 +2,7 @@ Command line interface
 ======================
 
 `matscipy` provides command line interfaces (CLIs) for selected functionality.
-These CLIs are installed per default and are prefixed with `matscipy_`.
+These CLIs are installed per default and are prefixed with `matscipy-`.
 
 .. toctree::
 
