+++ docs/cli/glasses.md
@@ -2,4 +2,4 @@
 
 ## Overview of commands
 
-* `matscipy_quench`: Liquid quench of a glassy structure.
+* `matscipy-quench`: Liquid quench of a glassy structure.
