+++ docs/cli/interatomic.md
@@ -2,4 +2,4 @@
 
 ## Overview of commands
 
-* `matscipy_average_eam_potential`: Generate averaged EAM potential for an alloy.
+* `matscipy-average-eam-potential`: Generate averaged EAM potential for an alloy.
