+++ docs/cli/diffusion.md
@@ -1,12 +1,6 @@
 # Diffusion
 
-## Overview of commands
-
-* `matscipy-rms`: Compute the root-mean-square (rms) displacement on a trajectory file.
-
-## Usage
-
-### `matscipy-rms`
+## `matscipy-rms`
 
 The `matscipy-rms` utility script computes the rms displacements of all atoms in a NetCDF trajectory file. It
 gracefully handles jumps of atoms through periodic boundaries by applying a minimum image convention to each
