+++ docs/cli/glasses.md
@@ -1,5 +1,0 @@
-# Glasses
-
-## Overview of commands
-
-* `matscipy-quench`: Liquid quench of a glassy structure.
