### 2024-01-15 a549670
DOC: Added documentation for `matscipy-average-eam-potential`
