### 2024-01-15 10be8a5
MAINT: Refactored average_eam_potential.py to calculators submodule
