### 2024-01-15 feafa35
MAINT: Refactored fracture mechanics scripts to CLI
