+++ scripts/fracture_mechanics/quasistatic_crack.py
@@ -1,206 +1,0 @@
-#
-# Copyright 2014-2015, 2021 Lars Pastewka (U. Freiburg)
-#           2015 James Kermode (Warwick U.)
-#
-# matscipy - Materials science with Python at the atomic-scale
-# https://github.com/libAtoms/matscipy
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 2 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-
-# ======================================================================
-# matscipy - Python materials science tools
-# https://github.com/libAtoms/matscipy
-#
-# Copyright (2014) James Kermode, King's College London
-#                  Lars Pastewka, Karlsruhe Institute of Technology
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 2 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-# ======================================================================
-
-# USAGE:
-#
-# Code imports the file 'params.py' from current working directory. params.py
-# contains simulation parameters. Some parameters can be omitted, see below.
-#
-# Parameters
-# ----------
-# calc : ase.Calculator
-#     Calculator object for energy and force computation.
-# tip_x0 : float
-#     Initial x-position of crack tip.
-# tip_y0 : float
-#     Initial y-position of crack tip.
-# tip_dx : array-like
-#     Displacement of tip in x-direction during run. x- and y-positions will be
-#     optimized self-consistently if omitted.
-# tip_dy : array-like
-#     Displacement of tip in y-direction during run. Position will be optimized
-#     self-consistently if omitted.
-
-import os
-import sys
-
-import numpy as np
-
-import ase
-import ase.io
-import ase.constraints
-import ase.optimize
-from ase.units import GPa
-
-import matscipy.fracture_mechanics.crack as crack
-from matscipy import parameter
-from matscipy.logger import screen
-
-from setup_crack import setup_crack
-from ase.optimize.precon import PreconLBFGS
-
-###
-
-sys.path += ['.', '..']
-import params
-
-###
-
-# Atom types used for outputting the crack tip position.
-ACTUAL_CRACK_TIP = 'Au'
-FITTED_CRACK_TIP = 'Ag'
-
-###
-
-logger = screen
-
-###
-
-a, cryst, crk, k1g, tip_x0, tip_y0, bond1, bond2, boundary_mask, \
-    boundary_mask_bulk, tip_mask = setup_crack(logger=logger)
-ase.io.write('notch.xyz', a, format='extxyz')   
-
-# Global parameters
-basename = parameter('basename', 'quasistatic_crack')
-calc = parameter('calc')
-fmax = parameter('fmax', 0.01)
-
[…]
