+++ scripts/fracture_mechanics/optimize_crack_tip.py
@@ -45,23 +45,18 @@
 # Code imports the file 'params.py' from current working directory. params.py
 # contains simulation parameters. Some parameters can be omitted, see below.
 
-import os
-import sys
-
 import numpy as np
 
-import ase
 import ase.constraints
 import ase.io
 import ase.optimize
 from ase.data import atomic_numbers
-from ase.units import GPa
 
 import matscipy.fracture_mechanics.crack as crack
 from matscipy import parameter
 from matscipy.logger import screen
 
-from setup_crack import setup_crack
+from matscipy.cli.fracture_mechanics.setup_crack import setup_crack
 
 ###
 
