### 2024-01-15 23d8827
DOC: Added a very simple getting started page
