### 2024-01-15 256672d
MAINT: Renamed "scripts" to "staging" and added a README file
