### 2024-01-15 29d01cc
DOC: conditionally exclude fenics-dependent notebooks when building the documentation if fenics not available
