### 2024-01-15 d6d6548
add docs for fracture scripts
