### 2024-01-15 1a15973
BUG: Fixed quench example and added README.md
