+++ examples/glasses/README.md
@@ -1,0 +1,5 @@
+Glass examples
+==============
+
+Subdirectory `amorphous_SiC`: This contains a parameterization for the `matscipy-quench` command generating and
+amorphous Si-C configuration. Simply run `matscipy-quench` in this subdirectory.
