### 2024-01-15 55a0a0b
MAINT: Removed Hessian matrix example as this is now in the docs
