### 2024-01-16 bc88b77
remove socketcalculator example as not longer supported
