### 2024-01-16 995e8e6
docs for MCFM calculator example
