+++ examples/mcfm/README.md
@@ -1,0 +1,7 @@
+# Multi Cluster Force Mixing (MCFM) example
+
+This directory contains example input files and scripts for the MCFM calculator.
+
+Run the `example_simulation.py` script to start the simulation. This is a self-contained example which uses two Morse potentials with different parameters in place of 'real' QM and MM calculators.
+
+See [Golebiowski2018](http://dx.doi.org/10.1063/1.5035508) and [Golebiowski2020](http://dx.doi.org/10.1039/d0cp01841d) for applications of this approach.
