### 2024-01-16 ec57f7d
MAINT: pin nglview version to resolve https://github.com/libAtoms/matscipy/issues/222 for now
