### 2024-01-16 1ab5e1a
remove defunct examples
