+++ examples/fracture_mechanics/README.md
@@ -9,8 +9,6 @@ and renamed to `params.py` before invoking the relveant script.
 
 - `quasistatic_fracture` - example input for the `matscipy-quasistatic-fracture` CLI tool
 - `sinclair-crack` - example input for the `matscipy-sinclair-crack` CLI tool
-- `energy_barrier/` - example for `staging/fracture_mechanics/energy_barrier.py` which computes energy barriers to breaking the bond at the tip of a crack
-- `energy_barrier_multiple/` - example for `staging/fracture_mechanics/energy_barrier_multiple.py` where multiple bond breaking steps are considered
 - `ideal_brittle_solid/` - example for `staging/fracture_mechanics/run_ideal_brittle_solid.py` which models fracture in an ideal harmonic solid (following the approach of D. Holland and M. P. Marder, Cracks and Atoms, Advanced Materials.)
 - `make_crack_thin_strip` - example for `staging/fracture_mechanics/make_crack_thin_strip.py`, which constructs a fracture simulation cell with 'thin strip' boundary conditions. The same parameter files serves as input for the `staging/fracture_mechanics/run_crack_thin_strip` script which runs MD in this geometry.
 - `quartz_crack` - example input for `staging/fracture_mechanics/quartz_crack.py` which models fracture of quartz
