### 2024-01-16 2271f11
ENH: Porting CCD & CCDD tests to TestCases
