+++ tests/test_dislocation.py
@@ -55,7 +55,6 @@ try:
 except ImportError:
     print("ovito not found: skipping some tests")
 
-
 class TestDislocation(matscipytest.MatSciPyTestCase):
     """Class to store test for dislocation.py module."""
 
@@ -411,320 +410,6 @@ class TestDislocation(matscipytest.MatSciPyTestCase):
         # to length of kink_length * 3 - 1 (for left kink)
         self.assertEqual(len(sliced_left_kink), kink_length * 3 - 1)
 
-    def check_disloc(self, cls, ref_angle, structure="BCC", test_u=True,
-                     burgers=0.5 * np.array([1.0, 1.0, 1.0]), tol=10.0, gen_bulk=False):
-        alat = 3.14339177996466
-        C11 = 523.0266819809012
-        C12 = 202.1786296941397
-        C44 = 160.88179872237012
-
-        if gen_bulk:
-            a = ase_bulk("Al", structure.lower(), alat, cubic=True)
-        else:
-            a = alat
-
-        d = cls(a, C11, C12, C44, symbol="Al")
-        bulk, disloc = d.build_cylinder(20.0)
-
-        # test that assigning non default symbol worked
-        assert np.unique(bulk.get_chemical_symbols()) == "Al"
-        assert np.unique(disloc.get_chemical_symbols()) == "Al"
-
-        assert len(bulk) == len(disloc)
-
-        if test_u:
-            # test the consistency
-            # displacement = disloc.positions - bulk.positions
-            stroh_displacement = d.displacements(bulk.positions,
-                                                 np.array(disloc.info["core_positions"]),
-                                                 self_consistent=d.self_consistent)
-
-            displacement = disloc.positions - bulk.positions
-
-            np.testing.assert_array_almost_equal(displacement,
-                                                 stroh_displacement)
-
-        results = sd.ovito_dxa_straight_dislo_info(disloc, structure=structure)
-        assert len(results) == 1
-        position, b, line, angle = results[0]
-        self.assertArrayAlmostEqual(np.abs(b), burgers)  # signs can change
-
-        err = angle - ref_angle
-        print(f'angle = {angle} ref_angle = {ref_angle} err = {err}')
-        assert abs(err) < tol
-
-        if issubclass(cls, sd.CubicCrystalDislocationQuadrupole):
-            self.check_dissociated_disloc(cls, structure, gen_bulk)
-
-    def check_dissociated_disloc(self, cls, structure="BCC", gen_bulk=False):
-        alat = 3.14339177996466
-        C11 = 523.0266819809012
-        C12 = 202.1786296941397
-        C44 = 160.88179872237012
-
-        if gen_bulk:
-            a = ase_bulk("Al", structure.lower(), alat, cubic=True)
-        else:
-            a = alat
-
-        d = cls(a, C11, C12, C44, symbol="Al")
-        
-        partial_dist = 5.0        
-        bulk, disloc = d.build_cylinder(40.0, partial_distance=partial_dist)
-
-        # Check good agreement of displacements method with sum of 
-        # left and right displacements
-        # May not be equal, due to self_consistent
-
-        core_positions = np.array([
-            [0, 0, 0],
-            [partial_dist, 0, 0]
-        ])
-        
-        # Left + Right displacements
-        full_disps = d.displacements(bulk.positions, core_positions, 
-                                      self_consistent=d.self_consistent)
-        # Left only
-        left_disps = d.left_dislocation.displacements(bulk.positions, core_positions[0, :], 
-                                      self_consistent=d.self_consistent)
-        # Right only
-        right_disps = d.right_dislocation.displacements(bulk.positions, core_positions[1, :], 
[…]
