### 2024-01-16 c4740d2
TST: Revert skipped tests from commit bbf82ec
