### 2024-01-16 6071236
BUG: Fix issue with setting dissociated disloc "core_positions" info
