### 2024-01-16 f925515
MAINT:  Use default self_consistent if none provided
