### 2024-01-16 #5668 @pastewka
@diehlpk - We finished addressing the reviewer comments, response letter follows here. I am also tagging my coauthors @jameskermode and @pgrigorev

We thank all three reviewers for the time devoted to looking at manuscript, documentation and code. This was extremely helpful to streamlining matscipy and we think that in particular the documentation of the code has tremendously improved in the process.

Below we have collected the reviewers comments from different locations (JOSS issue track, matscipy issue tracker) and briefly outline our actions in response to the comments. We hopefully didn't miss any comments as they were in different places - if we did, please accept our apologies and let us know where the additional comments are.

**From @mastricker:**

**_I have the impression that not all features are covered in the online documentation that is automatically generated by SPHINX; What would be of particular interest for my is a documentation how to run the DXA; or even to generate specific dislocation configurations in the first place. Part of the functionality is buried in the API reference section of the online documentation._**

We cannot run DXA within matscipy but rely on the Ovito module for it. We are using DXA to test the dislocation generation. In the matscipy version during original submission those tests were actually deactivated in CI, we now activated them. Ovito and atomman are now installed as dependencies when installing matscipy with the [test] option.

Regarding dislocation-generation, we have substantially expanded the online documentation to include examples of it: [https://libatoms.github.io/matscipy/applications/plasticity.html](https://libatoms.github.io/matscipy/applications/plasticity.html)

**_Presumably test_dislocation.py provides unclean output of test_dislocation.py ssssssssssssss.Fssssssss.Fs.ssssssss.s_**

This was due to a lack of atomman and ovito packages. We have added both packages to the test dependencies in pyproject.toml, see also the above comment.

**_Presumably test_energy_release.py provides additional s as output python -m pytest . results in a segmentation fault in test_neighbours.py -> Fatal Python error: Segmentation fault_**

Unfortunately we cannot reproduce this failure. We will keep this in mind when continuing to develop matscipy. Note that our longer term agenda includes switching the C-bindings to pybind11, which should make the code more robust, in particular with regards to buffer overflows which may have been behind this segmentatio fault.

**From @rashatwi:**

**_Wondering if there is a way to show some of the equations used in the analysis codes like the radial and angular correlation functions? Multiple variations exist for these calculations (e.g., different normalization methods) and it would be nice for the user to know how these are exactly calculated without having to go over the code (either in the paper or on the package website)._**

We have now put the equation into the documentation and added a documentation page on the angle distribution. We think that those equations would be too much detail for the JOSS paper, but agree that they should be mentioned in the documentation.

**_In the All-purpose atomic analysis tools section of the paper, only the topology building for non-reactive MD simulations subsection includes the specific module where the functionality is implemented, the other 4 subsections do not refer to the module and I had to go over the package modules to know where these are implemented._**

We removed the mention of explicit module names from the paper because we want to keep the option to refactor the code in the future without invalidating the paper. The package documentation should be the central source of information for how to actually run the code.

**_I suggest adding the package sphinx documentation link in the “About” section at the top right of the GitHub repository. It makes it much more accessible than just referring to it in the README file._**

This is a good suggestion. We added this link to the about section.

**_I see there is support for command line interface, but there is no details about how to use it. I’m not sure if this is a preliminary implementation of the cli. If not, then maybe adding simple examples in the README file will be helpful._**

It is correct that the code has a CLI that packages some workflows. There was also a subdirectory “scripts” that is not described anywhere. We have now moved some of the “scripts” to an actual (installable) CLI and added a section to the online documentation of the code on the CLI, and renamed “scripts” to “staging” to emphasize the transitory nature of the scripts stored there. Once matured, those scripts will move to the actual CLI, which is installed with matscipy and documented.

**_In the matscipy.elasticity module, lines 1303 and 1304, spilu and LinearOperator are not properly imported from scipy._**

This was a good catch. We fixed this.

**_I see there is an examples folder in the root directory with descriptive names for the python scripts, but it is still not very clear to me what these examples do. I think it will be helpful to have a tutorials section on the website or Jupyter notebooks added to the main GitHub repository that guide the user step-by-step to how they can start using the functionalities described in the paper and also contain detailed explanations._**

We thank the reviewer for this criticism, which we agree with. We have transitioned many of the examples to notebook tutorials included in the documentation pages (which has the further benefit that they are run in CI), and for the others have at the  minimum added README files to all subdirectories. \

**_On a side note, a lot of the examples provided in that directory have not been updated in a long time, e.g., in the example under fracture_mechanics/energy_barrier/params.py, I’m getting this import error: \
 ImportError: cannot import name 'diamond_111_110' from 'matscipy.fracture_mechanics.clusters’_**

We have removed the defunct examples and replaced them with tutorials which are included in the CI and run automatically to ensure they remain up to date. The examples which remain have had minimal README files added and over time they will be transitioned into full tutorial notebooks which can be included in the documentation and CI.

**From @mbarzegary:**

**_Some of the mentioned functionalities of the package are not documented. For example, there is no documentation for the electrochemistry. Also, some of the atomic analysis tools and all the interatomic potential features are only documented via autogenerated API docs. I think this makes it hard for beginner users to get started with the package. The application domains (except the electrochemistry, as mentioned) are well documented and described. It can be beneficial to talk a little bit about the package features in the documentation._**

We are sorry for this, as there was indeed no documentation for the electrochemistry module. We now added this part of the documentation. We agree that not all code functionality is documented in the hand-written documentation, but we do think that the core is now documented. We will add documentation as we continue to develop the code.

**_It's mentioned that the package has an interface to the FEniCS code. This functionality is supported by one example that does not run. FEniCS dependency is not documented and elaborated._**

This is now documented in the electrochemistry documentation.

**_I cannot find the FEniCS functionality being tested via the pytest tests. Is it because of the complicated installation procedure? I care about this functionality not only due to its usefulness but also because it's mentioned in the paper._**

We now added FEniCS tests to CI. It will run automatically on a user installation if a FEniCS installation is detected. Note that (unlike Ovito and atomman) we have not added FEniCS to the [test] optional specialization of matscipy as FEniCS is not simply installable via pip.

**_All tests pass on my local machine, but some of them produce unclear characters. This is the output of the pytest run:_**

Some tests were probably skipped (resulting in an ‘s’ state) because of missing dependencies (Ovito, atomman, FEniCS). All tests that rely on any of these modules skip automatically if they are not present. Note that all tests run in CI.

**_It's not clear where the user should start using the package after installing it. The way the provided examples are placed is a bit difficult to follow. Except for the electrochemistry and the fracture_mechanics directories, the rest of the examples directories are not clear regarding which functionality they belong to. I suggest that you add a new example directory for getting started with the package, putting some basic examples grouped per feature._**

We changed the entry page of the documentation and replaced it with a very brief getting started page.

**_It's not clear how the distributed computation example is supposed to work._**

We have removed this module. It was a legacy module that was unmaintained.

**_Generally speaking, the example directory contains a large number of non-trivial files, most of which are not documented unfortunately. I think this makes it difficult to get started with the package._**

This comment was also raised by rashatwi. We have added CLI documentation, moved some examples to the documentation and added README files to the examples. The longer term goal is to transition all examples to the documentation.

**_Running the samples_pb_c2d.py and samples_pnp_c2d.py examples results in the following error: ModuleNotFoundError: No module named 'matscipy.electrochemistry.utility'. The error for samples_pnp_by_fenics_fem.py is related to lack of matscipy.electrochemistry.poisson_nernst_planck_solver_fenics module._**

We thank you for checking this, this problem has been fixed.

**_Some examples rely on IPython, which doesn't get installed via the installation procedure of the package._**

We are not sure which examples the reviewer refers to. Note that we transitionen some examples to docs, and none of the remaining examples require IPython.

**_As mentioned in the paper, there exists other packages for multi-scale coupling such as LibMultiScale and MultiBench, but the necessity of the proposed package is not well elaborated. Since it’s the most important paragraph of the statement of need, I appreciate it if the authors elaborate on this._**

We have rewritten this part of the paper to more explicitly mention other packages and their capabilities.

**_I see duplicate references. For example, there are two identical Seidel et al. (2021a and 2021b). Can you please check the paper for this?The authors have referred to 2 papers using the coupling of the numerical models of the Poisson–Nernst–Planck equation in electrochemistry (Hörmann et al., 2023; Seidl et al. 2021b), but it doesn't seem that those papers use this coupling. Can you please elaborate on the mentioned usage?_**

We have removed those citations. They were examples for specific atomic phenomena but the reviewer is right that matscipy is not used in those works and citing them may confuse the reader.
