### 2024-01-16 e6471f8
TST: Decrease printing, reduct number of skipped tests
