+++ tests/test_dislocation.py
@@ -106,7 +106,7 @@ class TestDislocation(matscipytest.MatSciPyTestCase):
         pot_path = os.path.join(test_dir, pot_name)
         calc_EAM = EAM(pot_path)
         obtained_values = sd.get_elastic_constants(calculator=calc_EAM,
-                                                   delta=1.0e-3)
+                                                   delta=1.0e-3, verbose=False)
 
         self.assertArrayAlmostEqual(obtained_values, target_values, tol=1e-2)
 
@@ -135,7 +135,7 @@ class TestDislocation(matscipytest.MatSciPyTestCase):
                            log_file="lammps.log")
 
         obtained_values = sd.get_elastic_constants(calculator=lammps,
-                                                   delta=1.0e-3)
+                                                   delta=1.0e-3, verbose=False)
 
         os.remove("lammps.log")
 
@@ -493,12 +493,12 @@ class TestDislocation(matscipytest.MatSciPyTestCase):
 
         # Setup CubicCrystalDislocation object, and get displacments
         ccd = cls(a0, C11, C12, C44, symbol="Fe")
-        bulk, sc_disloc1 = ccd.build_cylinder(20.0)
+        bulk, sc_disloc1 = ccd.build_cylinder(20.0, verbose=False)
         center = np.diag(bulk.cell) / 2
-        stroh_disp = ccd.displacements(bulk.positions, center, method="atomman", self_consistent=False)
+        stroh_disp = ccd.displacements(bulk.positions, center, method="atomman", self_consistent=False, verbose=False)
 
         # Get displacements using AnistoropicDislocation class
-        adsl_disp = ccd.displacements(bulk.positions, center, method="adsl", self_consistent=False)
+        adsl_disp = ccd.displacements(bulk.positions, center, method="adsl", self_consistent=False, verbose=False)
 
         # Setup the dislcation from the AnistoropicDislocation object
         disloc = bulk.copy()
@@ -590,6 +590,10 @@ class TestDislocation(matscipytest.MatSciPyTestCase):
 
 
 class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
+
+    has_atomman = "atomman" not in sys.modules
+    has_ovito = "ovito" not in sys.modules
+    
     def __init__(self, disloc_cls):
         '''
         Test a CubicCrystalDislocation class
@@ -603,6 +607,8 @@ class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
 
         self.set_test_name()
 
+        self.default_method = "atomman" if self.has_atomman else "adsl"
+
     def set_test_name(self):
         '''
         Hack for making the unittest test name include the name of the dislocation
@@ -613,16 +619,13 @@ class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
         self.__setattr__(test_method_name, self.perfect_disloc_tests)
         unittest.TestCase.__init__(self, methodName=test_method_name)
 
-    @unittest.skipIf("atomman" not in sys.modules or
-                     "ovito" not in sys.modules,
-                     "requires atomman and ovito")
     def perfect_disloc_tests(self):
         '''
         Test build_cylinder and build_glide_configurations
         '''
         with self.subTest("Checking construction of dislocation cylinders"):
-            self.check_disloc(gen_bulk=False)
-            self.check_disloc(gen_bulk=True)
+            self.check_dislocation_cylinder(gen_bulk=False)
+            self.check_dislocation_cylinder(gen_bulk=True)
 
         with self.subTest("Checking dislocation glide configurations"):
             self.check_glide_configs()
@@ -630,8 +633,11 @@ class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
[…]
