### 2024-01-16 4fd9df9
TST: Disabled printing of dd_test.png
