### 2024-01-17 86a4cef
MAINT: Allow testing of Lattice objects with assertAtomsAlmostEqual
