### 2024-01-17 5e415a2
TST: Basic unittests for quadrupoles
