### 2024-01-17 78dc9b8
MAINT: Add missing space to SCF failure error msg
