### 2024-01-18 2280b3c
ENH: Allow dissociated disloc properties as classmethods
