+++ matscipy/dislocation.py
@@ -2421,19 +2421,19 @@ class CubicCrystalDislocation(metaclass=ABCMeta):
         # Create copies of immutable class attributes
         # Prevents side effects when E.G. changing burgers vector of an instance
         # Also makes changing cls.var and instance.var work as expected for these variables
-        self.axes = self.axes.copy()
-        self.burgers_dimensionless = self.burgers_dimensionless.copy()
-        self.unit_cell_core_position_dimensionless = self.unit_cell_core_position_dimensionless.copy()
-        self.parity = self.parity.copy()
+        # Skipped for dissociated dislocations, as these operate via classmethod properties
+        if not issubclass(self.__class__, CubicCrystalDissociatedDislocation):
+            self.axes = self.axes.copy()
+            self.burgers_dimensionless = self.burgers_dimensionless.copy()
+            self.unit_cell_core_position_dimensionless = self.unit_cell_core_position_dimensionless.copy()
+            self.parity = self.parity.copy()
         
-        self.alat, self.unit_cell = validate_cubic_cell(a, symbol, self.axes, self.crystalstructure, self.pbc)
+            self.alat, self.unit_cell = validate_cubic_cell(a, symbol, self.axes, self.crystalstructure, self.pbc)
 
-        self.glide_distance = self.alat * self.glide_distance_dimensionless
-
-        # Empty dict for solver methods (e.g. AnisotropicDislocation.displacements)
-        self.solvers = {
-            m : None for m in self.avail_methods
-        }
+            # Empty dict for solver methods (e.g. AnisotropicDislocation.displacements)
+            self.solvers = {
+                m : None for m in self.avail_methods
+            }
 
         # Sort out elasticity matrix into 6x6 convention (as we know the system is cubic)
         self.C = coalesce_elastic_constants(C11, C12, C44, C, convention="Cij")
@@ -3208,9 +3208,11 @@ class CubicCrystalDissociatedDislocation(CubicCrystalDislocation, metaclass=ABCM
                         left and righ dislocations are not the same.
         """
 
-        if not isinstance(self.left_dislocation, CubicCrystalDislocation):
+        if not (isinstance(self.left_dislocation, CubicCrystalDislocation) or \
+                isinstance(self.left_dislocation, CubicCrystalDissociatedDislocation)):
             self.left_dislocation = self.left_dislocation(a, C11, C12, C44, C, symbol)
-        if not isinstance(self.right_dislocation, CubicCrystalDislocation):
+        if not (isinstance(self.right_dislocation, CubicCrystalDislocation) or \
+                isinstance(self.right_dislocation, CubicCrystalDissociatedDislocation)):
             self.right_dislocation = self.right_dislocation(a, C11, C12, C44, C, symbol)
 
         # Change disloc burgers vectors, if requested
@@ -3222,13 +3224,6 @@ class CubicCrystalDissociatedDislocation(CubicCrystalDislocation, metaclass=ABCM
         # Set self.attrs based on left disloc attrs
         left_dislocation = self.left_dislocation
         right_dislocation = self.right_dislocation
-        self.crystalstructure = left_dislocation.crystalstructure
-        self.axes = left_dislocation.axes.copy()
-        self.unit_cell_core_position_dimensionless = left_dislocation.unit_cell_core_position_dimensionless.copy()
-        self.parity = left_dislocation.parity
-        self.glide_distance_dimensionless = left_dislocation.glide_distance_dimensionless
-        self.n_planes = left_dislocation.n_planes
-        self.self_consistent = left_dislocation.self_consistent
 
         super().__init__(a, C11, C12, C44, C, symbol)
 
@@ -3239,7 +3234,7 @@ class CubicCrystalDissociatedDislocation(CubicCrystalDislocation, metaclass=ABCM
[…]
