### 2024-01-18 #5668 @mastricker
> @mastricker please take a look at the author's response.

I'm good - I did not know that I had to look again. In my earlier response I wrote that once these issues are addressed, you can proceed.

@pastewka @pgrigorev @jameskermode I applaud you for your thorough response and you software. I'm really looking forward to putting it to use very soon.

Cheers,
M
