### 2024-01-19 406ca66
MAINT: Remove unneeded extra displacement eval
