### 2024-01-19 73d9a5e
ENH: Simple mixing of scf iterations
