### 2024-01-19 cd2b397
ENH: Tests across different elastic constants
