+++ tests/test_dislocation.py
@@ -29,6 +29,10 @@ import numpy as np
 
 from ase.calculators.lammpslib import LAMMPSlib
 from matscipy.calculators.eam import EAM
+from matscipy.dislocation import DiamondGlide90degreePartial, get_elastic_constants
+from matscipy.calculators.manybody.explicit_forms.stillinger_weber import StillingerWeber,\
+                                                                Holland_Marder_PRL_80_746_Si
+from matscipy.calculators.manybody import Manybody
 
 from ase.build import bulk as ase_bulk
 
@@ -589,6 +593,24 @@ class TestDislocation(matscipytest.MatSciPyTestCase):
         self.check_anisotropic_disloc(sd.BCCMixed111Dislocation, 70.5)
 
 
+test_calculators = {
+    # Calculators for several bulk properties
+    "diamond" : {
+        "symbol" : "Si",
+        "calc" : Manybody(**StillingerWeber(Holland_Marder_PRL_80_746_Si))
+    },
+
+    "fcc" : {
+        "symbol" : "Ni",
+        "calc" : EAM(test_dir + os.sep + "FeCuNi.eam.alloy")
+    },
+
+    "bcc" : {
+        "symbol" : "W",
+        "calc" : EAM(test_dir + os.sep + "w_eam4.fs")
+    }
+}
+
 class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
 
     has_atomman = "atomman" in sys.modules
@@ -600,10 +622,12 @@ class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
         '''
         self.test_cls = disloc_cls
 
-        self.alat = 3.14339177996466
-        self.C11 = 523.0266819809012
-        self.C12 = 202.1786296941397
-        self.C44 = 160.88179872237012
+        structure = self.test_cls.crystalstructure
+
+        self.symbol = test_calculators[structure.lower()]["symbol"]
+        calc = test_calculators[structure.lower()]["calc"]
+
+        self.alat, self.C11, self.C12, self.C44 = sd.get_elastic_constants(calculator=calc, symbol=self.symbol, verbose=False)
 
         self.set_test_name()
 
@@ -648,21 +672,19 @@ class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
         Test construction of CubicCrystalDislocation.build_cylinder
         '''
 
-        d = self.test_cls(self.alat, self.C11, self.C12, self.C44, symbol="Al")
-        structure = d.crystalstructure
 
         if gen_bulk:
-            a = ase_bulk("Al", structure.lower(), self.alat, cubic=True)
+            a = ase_bulk(self.symbol, self.structure.lower(), self.alat, cubic=True)
         else:
             a = self.alat
 
-        d = self.test_cls(a, self.C11, self.C12, self.C44, symbol="Al")
+        d = self.test_cls(a, self.C11, self.C12, self.C44, symbol=self.symbol)
         bulk, disloc = d.build_cylinder(20.0, method=self.default_method, 
                                         self_consistent=d.self_consistent, verbose=False)
 
         # test that assigning non default symbol worked
-        assert np.unique(bulk.get_chemical_symbols()) == "Al"
-        assert np.unique(disloc.get_chemical_symbols()) == "Al"
+        assert np.unique(bulk.get_chemical_symbols()) == self.symbol
+        assert np.unique(disloc.get_chemical_symbols()) == self.symbol
 
         assert len(bulk) == len(disloc)
 
@@ -685,18 +707,17 @@ class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
             
         if not self.has_ovito:
             self.skipTest("ovito module not installed")
[…]
