### 2024-01-19 af5a75d
Revert "MAINT: Remove unneeded extra displacement eval"
This reverts commit 406ca66d7ffe059fcdd0eeb96757c0e0783013f1.
