### 2024-01-19 3be40e5
BUG: Fix issue with burgers dimensionless not being instance attribute
