### 2024-01-19 1020669
MAINT: Calc test props once, not per test
