### 2024-01-19 2e6a834
TST: Better docstring on test name changing
