### 2024-01-19 a229c65
WIP: Added functionality to cut out a box of atoms from a simulation
