### 2024-01-22 7a30783
DOC: Added DOI for co-simulation paper
