+++ paper/paper.bib
@@ -1179,5 +1179,6 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="61--79",
 abstract="This paper describes methods and algorithms implemented for the co-simulation between a mesh-free discrete element method (DEM) code and a finite element analysis (FEA) code under control of a co-simulation middleware software environment.",
-isbn="978-3-319-06898-5"
+isbn="978-3-319-06898-5",
+doi="10.1007/978-3-319-06898-5_4
 }
