### 2024-01-22 9e7928f
DOC: Removed superfluous dot
