### 2024-01-23 8cc66a2
DOC: Updated figure caption to include hardware information
