+++ paper/paper.bib
@@ -1167,18 +1167,15 @@ url={https://openreview.net/forum?id=YPpSngE-ZU}
 }
 
 @InProceedings{Dehning2014,
-author="Dehning, Carsten
-and Bierwisch, Claas
-and Kraft, Torsten",
-editor="Griebel, Michael
-and Schweitzer, Marc Alexander",
-title="Co-simulations of Discrete and Finite Element Codes",
-booktitle="Meshfree Methods for Partial Differential Equations VII",
-year="2015",
-publisher="Springer International Publishing",
-address="Cham",
-pages="61--79",
-abstract="This paper describes methods and algorithms implemented for the co-simulation between a mesh-free discrete element method (DEM) code and a finite element analysis (FEA) code under control of a co-simulation middleware software environment.",
-isbn="978-3-319-06898-5",
-doi="10.1007/978-3-319-06898-5_4
+  author    = "Dehning, Carsten and Bierwisch, Claas and Kraft, Torsten",
+  editor    = "Griebel, Michael and Schweitzer, Marc Alexander",
+  title     = "Co-simulations of Discrete and Finite Element Codes",
+  booktitle = "Meshfree Methods for Partial Differential Equations VII",
+  year      = "2015",
+  publisher = "Springer",
+  address   = "Cham",
+  pages     = "61--79",
+  abstract  = "This paper describes methods and algorithms implemented for the co-simulation between a mesh-free discrete element method (DEM) code and a finite element analysis (FEA) code under control of a co-simulation middleware software environment.",
+  isbn      = "978-3-319-06898-5",
+  doi       = "10.1007/978-3-319-06898-5_4"
 }
