### 2024-01-23 #5668 @pastewka
> @pastewka Can you please take a look at my comments?
>
> * For Figure 1, I could not find the hardware details, the code was tested and the version. Can this be added?
>
> ```
>   • Neighbour list. An efficient linear-scaling neighbour list implemented in C delivers orders-
> 154
>  of-magnitude faster performance for large systems than the pure Python implementation
> 155
>  in ASE (Larsen et al., 2017), see Figure 1.
> ```
>
> * Committee models..  - It seems here are two dots, please verify

@diehlpk I've updated the paper, can you check if your comments are addressed?
