+++ tests/test_dislocation.py
@@ -22,6 +22,7 @@
 import os
 import unittest
 import matscipytest
+import pytest
 import sys
 
 import matscipy.dislocation as sd
@@ -58,6 +59,33 @@ try:
     import ovito
 except ImportError:
     print("ovito not found: skipping some tests")
+
+# Generator functions for lists of cubic dislocations
+def cubic_dislocs():
+    import inspect
+
+    # Get all cubic dislocation classes
+    # So new disloc types are automatically added to testing framework
+    _cubic_dislocs = [item for name, item in sd.__dict__.items() if inspect.isclass(item) and issubclass(item, sd.CubicCrystalDislocation)]
+    _cubic_dislocs = [item for item in _cubic_dislocs if item not in [sd.CubicCrystalDislocation, 
+                                                                    sd.CubicCrystalDissociatedDislocation,
+                                                                    sd.CubicCrystalDislocationQuadrupole]]
+    return _cubic_dislocs
+
+def cubic_perfect_dislocs():
+    _cubic_dislocs = cubic_dislocs()
+
+    _cubic_perfect_dislocs = [item for item in _cubic_dislocs if not issubclass(item, sd.CubicCrystalDissociatedDislocation)]
+    
+    return _cubic_perfect_dislocs
+
+def cubic_dissociated_dislocs():
+    _cubic_dislocs = cubic_dislocs()
+
+    _cubic_dissociated_dislocs = [item for item in _cubic_dislocs if issubclass(item, sd.CubicCrystalDissociatedDislocation)]
+    
+    return _cubic_dissociated_dislocs
+
 
 class TestDislocation(matscipytest.MatSciPyTestCase):
     """Class to store test for dislocation.py module."""
@@ -614,15 +642,12 @@ test_props = {
     }
 }
 
-class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
+class BaseTestCubicCrystalDislocation(matscipytest.MatSciPyTestFixture):
 
     has_atomman = "atomman" in sys.modules
     has_ovito = "ovito" in sys.modules
     
-    def __init__(self, disloc_cls):
-        '''
-        Test a CubicCrystalDislocation class
-        '''
+    def set_up_cls(self, disloc_cls):
         self.test_cls = disloc_cls
 
         self.structure = self.test_cls.crystalstructure
@@ -630,8 +655,6 @@ class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
         self.symbol = test_props[self.structure.lower()]["symbol"]
 
         self.alat, self.C11, self.C12, self.C44 = test_props[self.structure.lower()]["props"]
-
-        self.set_test_name()
 
         self.default_method = "atomman" if self.has_atomman else "adsl"
 
@@ -641,39 +664,13 @@ class TestCubicCrystalDislocation(matscipytest.MatSciPyTestCase):
         else:
             self.ncores = 1
 
-    def set_test_name(self):
-        '''
-        Hack for making the unittest test name include the name of the dislocation
-        e.g. "TestDiamondGlide90degreePartial"
-        '''
-
-        test_method_name = "Test" + self.test_cls.__name__
-        self.__setattr__(test_method_name, self.perfect_disloc_tests)
-        unittest.TestCase.__init__(self, methodName=test_method_name)
-
-    def perfect_disloc_tests(self):
-        '''
-        Test build_cylinder, build_glide_configurations, and all displacement solver methods
-        '''
-        with self.subTest("Checking construction of dislocation cylinders"):
-            self.check_dislocation_cylinder(gen_bulk=False)
-            self.check_dislocation_cylinder(gen_bulk=True)
-
-        with self.subTest("Checking dislocation glide configurations"):
-            self.check_glide_configs()
-
[…]
