### 2024-01-24 #5668 @diehlpk
@pastewka thanks, please do the following

- [ ] Please generate a new release, including all changes made during the review. Post the version number here.
- [ ] Please upload the release to Zenodo and make sure the title and authors match with the paper. Post the DOI here.
