### 2024-01-24 fe5cb9e
TST: Custom parametrize printing for dislocs
