### 2024-01-24 d6bdfd8
MAINT: Hide disloc lines for quadrupoles, as not correctly placed
