### 2024-01-24 becae50
TST: Improved formatting for pytest parametrize tests
