### 2024-01-24 f8f3226
WIP: Added LAMMPS quench and atom by atom deposition
