+++ matscipy/cli/glasses/deposition.py
@@ -1,0 +1,285 @@
+#this is a function used for deposition of atoms using ASE alongside the python interface to LAMMPS
+import numpy as np
+from ase.units import fs
+from ase import Atom
+from ase.io.lammpsdata import write_lammps_data, read_lammps_data
+import ase.io
+
+def set_up_lammps(lmps,input_file,mass_cmds,potential_cmds):
+    #lmps : LAMMPS object
+    #input_file : str, path to input file
+    #mass_cmds : commands to set atomic masses in LAMMPS
+    #potential_cmds : commands to set potential in LAMMPS
+    
+    # ---------- Initialize Simulation --------------------- 
+    lmps.command('clear') 
+    lmps.command('dimension 3')
+    lmps.command('boundary p p f')
+    lmps.command('atom_style atomic')
+    lmps.command('units metal')
+    lmps.command('comm_style tiled')
+    print(input_file)
+    #----------Read atoms------------
+    lmps.command(f'read_data {input_file}')
+    
+    #----------Define masses and Interatomic Potential----------------
+    lmps.commands_list(mass_cmds)
+    lmps.commands_list(potential_cmds)
+
+    #----------Output thermo data----------------
+    lmps.command('thermo 20')
+
+    #----------Create dump file----------------
+    lmps.command('dump 1 all custom 10 dump.lammpstrj id type x y z vx vy vz')
+    #make sure to append
+    lmps.command('dump_modify 1 append yes')
+
+def deposit_single_atom(prev_config,atom_energy,atom_mass):
+    #convert atom_energy (in eV) to velocity in ASE units. Atom mass is in atomic mass units
+    v_mag = np.sqrt(2*(atom_energy)/atom_mass)
+
+    print(f'VELOCITY MAGNITUDE: {v_mag}')
+    #this is a velocity magnitude, we need a vector
+    theta = np.random.uniform(-np.pi/12,np.pi/12)
+    phi = np.random.uniform(-np.pi,np.pi)
+    v = np.array([v_mag*np.sin(theta)*np.cos(phi),v_mag*np.sin(theta)*np.sin(phi),-v_mag*np.cos(theta)])
+
+    print(f'VELOCITY VECTOR: {v}')
+    #now we need a random point near the top of the surface, a distance away such that the particle should strike within 5fs
+    strike_time = 5*fs
+    place_distance = np.abs(v[2])*strike_time
+
+    print(f'PLACE DISTANCE: {place_distance}')
+    #get the highest point on the surface
+    top_z = np.max(prev_config.positions[:,2])
+    place_height = top_z + place_distance
+
+    print(f'PLACE HEIGHT: {place_height}')
+
+    #get a random x and y position from the box dimensions
+    cell = prev_config.get_cell()
+    x = np.random.uniform(0,cell[0,0])
+    y = np.random.uniform(0,cell[1,1])
+
+    #now we have a random point near the top of the surface, we can place the atom there
+    new_config = prev_config.copy()
+    new_atom = Atom('C')
+    new_config.append(new_atom)
+    new_config.positions[-1,:] = np.array([x,y,place_height])
+    old_velocities = new_config.get_velocities()
+    old_velocities[-1,:] = v
+    new_config.set_velocities(old_velocities)
+    #write to xyz
+    #ase.io.write('config_with_new_atom.xyz',new_config,parallel=False)
+
+    return new_config
+
+# function 1: deposit atoms
+def deposit_atom_sim(rank,prev_config,atom_energy,atom_mass,lmps,mass_cmds,potential_cmds,separation_tol):
+    #this function takes a previous configuration, adds an atom at a random point near the top and gives it a velocity
+    #with a magnitude found from atom_energy toward the surface
+    #it then runs a short MD simulation in NVE to simulate the atom impact
[…]
