+++ matscipy/cli/glasses/quench.py
@@ -22,6 +22,9 @@ import os
 import signal
 import sys
 
+#add current folder to path
+sys.path.insert(0, '.')
+
 import numpy as np
 
 import ase
@@ -33,6 +36,7 @@ from ase.symbols import string2symbols
 from ase.units import mol, fs, kB
 
 from matscipy import parameter
+from matscipy.neighbours import neighbour_list
 
 ###
 
@@ -44,28 +48,249 @@ signal.signal(signal.SIGUSR2, handle_sigusr2)
 
 ###
 
-def random_solid(els, density, sx=None, sy=None):
-    if isinstance(els, str):
-        syms = string2symbols(els)
-    else:
-        syms = ""
-        for sym, n in els:
-            syms += n*sym
-    r = np.random.rand(len(syms), 3)
-    a = ase.Atoms(syms, positions=r, cell=[1,1,1], pbc=True)
+def random_solid(els, density, internal_cutoff=0.0, sx=None, sy=None):
+    stable_struct = False
+    while not stable_struct:
+        if isinstance(els, str):
+            syms = string2symbols(els)
+        else:
+            syms = ""
+            for sym, n in els:
+                syms += n*sym
+        r = np.random.rand(len(syms), 3)
+        a = ase.Atoms(syms, positions=r, cell=[1,1,1], pbc=True)
 
-    mass = np.sum(a.get_masses())
-    if sx is not None and sy is not None:
-        sz = ( 1e24*mass/(density*mol) )/(sx*sy)
-        a.set_cell([sx,sy,sz], scale_atoms=True)
-    else:
-        a0 = ( 1e24*mass/(density*mol) )**(1./3)
-        a.set_cell([a0,a0,a0], scale_atoms=True)
-    a.set_initial_charges([0]*len(a))
+        mass = np.sum(a.get_masses())
+        if sx is not None and sy is not None:
+            sz = ( 1e24*mass/(density*mol) )/(sx*sy)
+            a.set_cell([sx,sy,sz], scale_atoms=True)
+        else:
+            a0 = ( 1e24*mass/(density*mol) )**(1./3)
+            a.set_cell([a0,a0,a0], scale_atoms=True)
+        a.set_initial_charges([0]*len(a))
+
+        nl = neighbour_list('i',atoms=a, cutoff=internal_cutoff)
+        print(nl)
+        if len(nl)==0:
+            stable_struct = True
 
     return a
 
 ###
+
+def set_up_lammps(lmps, input_file, mass_cmds, potential_cmds):
+    #lmps : LAMMPS object
+    #input_file : str, path to input file
+    #mass_cmds : commands to set atomic masses in LAMMPS
+    #potential_cmds : commands to set potential in LAMMPS
+    
+    # ---------- Initialize Simulation --------------------- 
+    lmps.command('clear') 
+    lmps.command('dimension 3')
+    lmps.command('boundary p p p')
+    lmps.command('atom_style atomic')
+    lmps.command('units metal')
+
+    print(input_file)
+    #----------Read atoms------------
+    lmps.command(f'read_data {input_file}')
+    
+    #----------Define masses and Interatomic Potential----------------
+    lmps.commands_list(mass_cmds)
+    lmps.commands_list(potential_cmds)
+
+    #----------Output thermo data----------------
+    lmps.command('thermo 100')
+
+
+def main_lammps(lmps, rank):
+    # This is a version of the function below, which instead uses the LAMMPS python interface with MPI4PY
+    # to run the actual quenching simulation. This is faster than the ASE interface, but requires a 
+    # LAMMPS python installation and LAMMPS potential.
+
+    # it also (at the time of writing) requires a version of ASE, which has better LAMMPS file functionality
[…]
