### 2024-01-24 #5668 @pastewka
* We have release v1.0.0
* I've generate a Zenodo record but not yet released it. I would like to point the Zenodo record back to this paper once it has a DOI. The DOI for the Zenodo record is 10.5281/zenodo.10564956. If you can give me a DOI for the JOSS paper and a page number I can release the Zenodo record.
