### 2024-01-24 #5668 @pastewka
@diehlpk  PS: Here is a Zenodo preview, but I don't know if this is visible without my credentials: https://zenodo.org/records/10564956?preview=1
