### 2024-01-24 #5668 @diehlpk
@pastewka can you please check the isbn in your Bibtex file?
