### 2024-01-24 #5668 @pastewka
@diehlpk Fixed it. It was an ISSN, not an ISBN
