### 2024-01-24 036129a
DOC: Removed duplicate Jana2019 entry
