+++ paper/paper.bib
@@ -777,31 +777,6 @@ abstract = {Calculations of dislocation-defect interactions are essential to mod
   doi     = "10.1103/PhysRevB.78.161402"
 }
 
-@ARTICLE{Jana2019,
-  title    = "Structural and elastic properties of amorphous carbon from
-              simulated quenching at low rates",
-  author   = "Jana, Richard and Savio, Daniele and Deringer, Volker and
-              Pastewka, Lars",
-  abstract = "We generate representative structural models of amorphous carbon
-              (a-C) from constant-volume quenching from the liquid with
-              subsequent relaxation of internal stresses in molecular dynamics
-              simulations using empirical and machine-learning interatomic
-              potentials. By varying volume and quench rate we generate
-              structures with a range of density and amorphous morphologies. We
-              find that all a-C samples show a universal relationship between
-              hybridization, bulk modulus and density despite having distinctly
-              different cohesive energies. Differences in cohesive energy are
-              traced back to slight changes in the distribution of bond-angles
-              that will likely be linked to thermal stability of these
-              structures.",
-  journal  = "Modell. Simul. Mater. Sci. Eng.",
-  volume   =  27,
-  pages    = "085009",
-  year     =  2019,
-  language = "en",
-  doi      = "10.1088/1361-651X/ab45da"
-}
-
 @ARTICLE{Muser2023,
   title     = "Interatomic potentials: achievements and challenges",
   author    = "M{\"u}ser, Martin H and Sukhomlinov, Sergey V and Pastewka, Lars",
