### 2024-01-24 #5668 @pastewka
@diehlpk  I've fixed the duplicate reference (not yet merged to master) but I don't know what "Element inline-formula is not declared in institution list of possible children" refers to.
