### 2024-01-25 #5668 @pastewka
@diehlpk We have removed all abstract from the .bib file, which I suppose may have been the source of the above error.
