### 2024-01-25 7231f8f
WIP: Made it so assigned groups always works regardless of size
