### 2024-01-25 0c0a16e
WIP: Fixed velocity vector
