### 2024-01-26 #5668 @xuanxu
The error is generated because the JATS metadata file used to deposit the paper does not allow inline math formulas in the headers, so [this mu](https://github.com/libAtoms/matscipy/blob/master/paper/paper.md?plain=1#L82) is causing it to fail.
Is it possible to spell that name in any other way?
