### 2024-01-26 #5668 @pastewka
@xuanxu  Does the system support UTF-8?
