### 2024-01-26 b6c7293
Use UTF-8 instead of inline Math
