### 2024-01-26 9ca5e4c
BUG: Fix bug in self-consistent definition
