### 2024-01-28 #5668 @Kevin-Mattheus-Moerman
@pastewka As AEiC I will now help to process final steps. I have checked this review, the archive link, your repository, and the paper. I have only the below minor points that require your attention:

- [ ] Please edit the Zenodo listed version tag to be `v1.0.0`, i.e. include the `v` since that is part of your version tag (this should result in it being shown on Zenodo as `Version v1.0.0`).
- [ ] In your affiliations, please spell out country names (e.g. `UK` should be `United Kingdom`).
- [ ] `underyling` should read `underlying`
- [ ] With words like `neighbour`, `neighbouring`, `behaviour`, `analysing`, `generalised` and others, this paper mainly appears to feature British English. However, the paper also features the American spellings: `discretized`, `specialized`, `optimized`, and `realized`. Please choose one format and use it consistently.
- [ ] `charactize` is a type and should be -> `characterized` or in British English -> `characterised`

Suggestion/comment (no action required):
- [ ] JOSS does not require full postal addresses for the affiliations. Hence if you prefer to shorten them you can remove the street names and postal codes (mind you, I do know some institutes prefer the longer form, it may be some French universities for instance. So, you may need to check with co-authors).
