### 2024-01-28 fd4e246
DOC: add doi of Brenner reference
