+++ paper/paper.bib
@@ -196,7 +196,7 @@ author = {Petr Grigorev and Alexandra M. Goryaeva and Mihai-Cosmin Marinica and 
 	author = {Brenner, Donald W.},
 	year = {1990},
 	pages = {9458--9471},
-	optdoi = {10.1103/PhysRevB.42.9458},
+	doi = {10.1103/PhysRevB.42.9458},
 }
 
 @article{BKS1990,
