+++ paper/paper.md
@@ -99,7 +99,7 @@ Behaviour of materials is governed by physical phenomena that occur at an extrem
 
 # Statement of need
 
-The Python package `matscipy` contains a set of tools for researchers using atomic-scale models in materials science. In atomic-scale modelling, the primary numerical object is a discrete point in three-dimensional space that represents the position of an individual atom. Simulations are often dynamical, where configurations change over time and each atom carries a velocity. Complexity emerges from the interactions of many atoms, and numerical tools are required for generating initial atomic configurations and for analyzing output of such dynamical simulations, most commonly to connect local geometric arrangements of atoms to physical processes. An example, described in more detail below, is the detection of the tip of a crack that moves through a solid body.
+The Python package `matscipy` contains a set of tools for researchers using atomic-scale models in materials science. In atomic-scale modelling, the primary numerical object is a discrete point in three-dimensional space that represents the position of an individual atom. Simulations are often dynamical, where configurations change over time and each atom carries a velocity. Complexity emerges from the interactions of many atoms, and numerical tools are required for generating initial atomic configurations and for analysing output of such dynamical simulations, most commonly to connect local geometric arrangements of atoms to physical processes. An example, described in more detail below, is the detection of the tip of a crack that moves through a solid body.
 
 We never see individual atoms at macroscopic scales. To understand the behaviour of everyday objects, atomic-scale information needs to be transferred to the continuum scale. This is the primary objective of multi-scale modelling. `matscipy` focuses on atomic representations of materials, but implements tools for connecting to continuum descriptions in mechanics and transport theory. Each of the application domains described in the following therefore relies on the computation of continuum fields, that is realised through analytic or numerical solutions.
 
@@ -109,7 +109,7 @@ The central class in ASE is the `Atoms` class, which is a container that stores 
 
 `matscipy` is a toolbox that enables multi-scale coupling, but it is not a toolbox for actually carrying out two-way coupled calculations. Its target is the construction of atomic domains from continuum information and the extraction of continuum fields from atomic structures. Other packages exist that take care of the actual, two-way coupling. In contrast to `matscipy`, those have a primary focus on handling discretised continuum fields, typically in the form of finite-element meshes, and interpolating nodal or element values between atomic-scale and continuum descriptions. `matscipy` itself has no provisions for handling discrete continuum data, but does implement analytic expressions for continuum fields.
 
-Example implementations of two-way coupling codes are the open-source code `libmultiscale` [@Anciaux2007;@Anciaux2018] that explicitly targets atomistic-continuum coupling, or the generic coupling libraries `preCICE` [@preCICEv2] or `MpCCI` [@Dehning2014]. Another two-way coupling code is `MultiBench` [@Miller2009], that was specifically designed for benchmarking a wide range of two-way atomistic-continuum coupling schemes. Futhermore, there are specialised multiscale coupling code, such as Green's function molecular dynamics (`GFMD`) [@Campana2006;@Pastewka2012] which targets two-way coupling in contact mechanics and friction simulations. All of these packages have only limited capabilities for constructing atomistic domains. `matscipy` could be combined with these packages for two-way coupled simulation of plasticity, fracture or frictional processes.
[…]
