### 2024-01-28 0ec9fc9
DOC: add JOSS badge
