### 2024-01-28 17c1bb1
DOC: add citation guide in README
