### 2024-01-29 #5668 @Kevin-Mattheus-Moerman
Congratulations on this JOSS publication @pastewka!

Thanks for editing @diehlpk

And a special Thank you to the reviewers: @rashatwi, @mbarzegary, @mastricker
